+++ paper/paper.bib
@@ -56,7 +56,7 @@ year = {2025}
   volume={35},
   pages={6133--6140},
   year={2023},
-  doi={10.1021/acs.chemmater.3c00874}
+  doi={10.1021/acs.chemmater.3c01271}
 }
 
 @article{PoletayevEtAl_NatureMater2022,
