### 2025-05-27 39ae7ec
Added GEMDAT Zenodo doi
