+++ paper/paper.bib
@@ -33,7 +33,8 @@ license = {Apache-2.0},
 title = {{GEMDAT}},
 url = {https://github.com/GEMDAT-repos/GEMDAT},
 version = {1.6.2},
-year = {2025}
+year = {2025},
+doi = {10.5281/zenodo.8401669}
 }
 
 @misc{pymatgen-analysis-diffusion,
