### 2025-05-27 77243a4
Updated .bib
