+++ paper/paper.bib
@@ -1,22 +1,3 @@
-@article{Gowers2016,
-  title={MDAnalysis: A Python Package for the Rapid Analysis of Molecular Dynamics Simulations},
-  author={Gowers, Richard J. and Linke, Max and Barnoud, Jonathan and Reddy, Tyler J. E. and Melo, Manuel N. and Seyler, Sean L. and Domanski, Jan and Dotson, David L. and Buchoux, S{\'e}bastien and Kenney, Ian M. and Beckstein, Oliver},
-  journal={Proceedings of the 15th Python in Science Conference},
-  pages={98--105},
-  year={2016}
-}
-
-@article{McGibbon2015,
-  title={MDTraj: A Modern Open Library for the Analysis of Molecular Dynamics Trajectories},
-  author={McGibbon, Robert T. and Beauchamp, Kyle A. and Harrigan, Matthew P. and Klein, Christoph and Swails, Jason M. and Hern{\'a}ndez, Carlos X. and Schwantes, Christian R. and Wang, Lee-Ping and Lane, Thomas J. and Pande, Vijay S.},
-  journal={Biophysical Journal},
-  volume={109},
-  number={8},
-  pages={1528--1532},
-  year={2015},
-  doi={10.1016/j.bpj.2015.08.015}
-}
-
 @article{Ong2013,
   title={Python Materials Genomics (pymatgen): A robust, open-source python library for materials analysis},
   author={Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
