+++ .github/workflows/build.yml
@@ -35,7 +35,7 @@ jobs:
         uses: coverallsapp/github-action@v2
         with:
           path-to-lcov: ./coverage.lcov
-          github-token: ${{ secrets.COVERALLS_TOKEN }}
+          github-token: ${{ secrets.GITHUB_TOKEN }}
           parallel: true
       
       - name: Static type checking
