+++ .github/workflows/build.yml
@@ -1,5 +1,4 @@
 name: build
-
 # Controls when the action will run
 on: [push, pull_request]
 
@@ -37,7 +36,20 @@ jobs:
           path-to-lcov: ./coverage.lcov
           github-token: ${{ secrets.GITHUB_TOKEN }}
           parallel: true
+          flag-name: python-${{ matrix.python-version }}
       
       - name: Static type checking
         run: |
           mypy site_analysis
+
+  # This job finalizes the parallel coverage reports
+  finish:
+    needs: tests
+    if: ${{ always() }}
+    runs-on: ubuntu-latest
+    steps:
+      - name: Coveralls Finished
+        uses: coverallsapp/github-action@v2.3.6
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          parallel-finished: true
