### 2025-06-03 30f21fd
Bump version to 1.0.4
