+++ tests/test_dynamic_voronoi_site.py
@@ -1,4 +1,5 @@
 import unittest
+from unittest.mock import patch
 import numpy as np
 from collections import Counter
 from pymatgen.core import Structure, Lattice
@@ -8,237 +9,253 @@ from site_analysis.site import Site
 from site_analysis.atom import Atom
 
 class DynamicVoronoiSiteTestCase(unittest.TestCase):
-	
-	def setUp(self):
-		Site._newid = 1
-		
-	def test_dynamic_voronoi_site_is_initialised(self):
-		reference_indices = [1, 2, 3, 4]
-		site = DynamicVoronoiSite(reference_indices=reference_indices)
-		
-		# Check that reference indices are stored correctly
-		self.assertEqual(site.reference_indices, reference_indices)
-		
-		# Check that the centre is initially None (calculated on demand)
-		self.assertEqual(site._centre_coords, None)
-		
-		# Check inherited attributes from Site
-		self.assertEqual(site.index, 1)
-		self.assertEqual(site.label, None)
-		self.assertEqual(site.contains_atoms, [])
-		self.assertEqual(site.trajectory, [])
-		self.assertEqual(site.points, [])
-		self.assertEqual(site.transitions, Counter())
-		
-	def test_dynamic_voronoi_site_is_initialised_with_label(self):
-		"""Test that a DynamicVoronoiSite can be initialised with a label."""
-		reference_indices = [1, 2, 3, 4]
-		label = "tetrahedral_site"
-		site = DynamicVoronoiSite(reference_indices=reference_indices, label=label)
-		
-		self.assertEqual(site.reference_indices, reference_indices)
-		self.assertEqual(site.label, label)
-		
-	def test_reset(self):
-		"""Test that reset clears the calculated centre and site occupation data."""
-		reference_indices = [1, 2, 3, 4]
-		site = DynamicVoronoiSite(reference_indices=reference_indices)
-		
-		# Mock some data that should be cleared on reset
-		site._centre_coords = np.array([0.5, 0.5, 0.5])
-		site.contains_atoms = [10, 11]
-		site.trajectory = [[10], [10, 11]]
-		site.transitions = Counter([3, 3, 2])
-		
-		# Reset the site
-		site.reset()
-		
-		# Check that all appropriate attributes are reset
-		self.assertEqual(site._centre_coords, None)
-		self.assertEqual(site.contains_atoms, [])
-		self.assertEqual(site.trajectory, [])
-		self.assertEqual(site.transitions, Counter())
-		
-	def test_calculate_centre(self):
-		"""Test that the centre is correctly calculated from reference atoms."""
-		reference_indices = [0, 1, 2, 3]
-		site = DynamicVoronoiSite(reference_indices=reference_indices)
-		
-		# Create a test structure
-		lattice = Lattice.cubic(10.0)
-		coords = [[0.1, 0.1, 0.1],   # atom 0
-				  [0.2, 0.1, 0.1],   # atom 1 
-				  [0.1, 0.2, 0.1],   # atom 2
-				  [0.1, 0.1, 0.2]]   # atom 3
-		structure = Structure(lattice, ["Li", "Li", "Li", "Li"], coords)
-		
-		# Calculate centre
-		site.calculate_centre(structure)
-		
-		# Expected centre is the mean of the reference atom positions
-		expected_centre = np.mean(np.array(coords), axis=0)
-		np.testing.assert_array_almost_equal(site._centre_coords, expected_centre)
-	
-	def test_calculate_centre_with_pbc(self):
-		"""Test centre calculation with atoms that cross periodic boundaries."""
-		reference_indices = [0, 1, 2, 3]
[…]
