### 2025-06-03 68e5b7a
Added reference-centre PBC method
