### 2025-06-03 0dbfdb5
trying to vectorise PBCs
