+++ tests/test_pbc_utils.py
@@ -5,7 +5,7 @@ from pymatgen.core import Structure, Lattice
 from site_analysis.polyhedral_site import PolyhedralSite
 from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.site import Site
-from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_centre
+from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_centre, unwrap_vertices_vectorised
 
 
 class TestLegacyPBCCorrection(unittest.TestCase):
@@ -313,6 +313,250 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		# The centre of the unwrapped vertices should be close to the reference centre
 		result_centre = np.mean(result, axis=0)
 		np.testing.assert_array_almost_equal(result_centre, expected_result_centre, decimal=2)
+		
+class TestVectorisedUnwrapping(unittest.TestCase):
+	"""Tests for vectorised reference centre-based vertex unwrapping."""
+	
+	def setUp(self):
+		self.lattice = Lattice.cubic(10.0)
+	
+	def test_unwrap_vertices_vectorised_no_unwrapping_needed(self):
+		"""Test vectorised case where vertices form a tetrahedron around the reference centre."""
+		# Simple tetrahedron centered around [0.5, 0.5, 0.5]
+		vertex_coords = np.array([
+			[0.4, 0.4, 0.4],
+			[0.6, 0.6, 0.4],
+			[0.6, 0.4, 0.6],
+			[0.4, 0.6, 0.6]
+		])
+		reference_centre = np.array([0.5, 0.5, 0.5])
+		
+		result = unwrap_vertices_vectorised(vertex_coords, reference_centre, self.lattice)
+		
+		# Should remain unchanged since all vertices are already close
+		np.testing.assert_array_almost_equal(result, vertex_coords, decimal=7)
+		
+	def test_unwrap_vertices_vectorised_simple_case(self):
+		"""Test vectorised unwrapping for an octahedron spanning the boundary at x=0."""
+		# Octahedron centered at [0.05, 0.5, 0.5] spanning the x=0 boundary
+		vertex_coords = np.array([
+			[0.1, 0.5, 0.5],   # +x vertex
+			[-0.05, 0.5, 0.5], # -x vertex
+			[0.05, 0.6, 0.5],  # +y vertex  
+			[0.05, 0.4, 0.5],  # -y vertex
+			[0.05, 0.5, 0.6],  # +z vertex
+			[0.05, 0.5, 0.4]   # -z vertex
+		])
+		reference_centre = np.array([0.05, 0.5, 0.5])
+		
+		result = unwrap_vertices_vectorised(vertex_coords, reference_centre, self.lattice)
+		
+		# After unwrapping and shifting to ensure non-negative coordinates
+		expected = np.array([
+			[1.1, 0.5, 0.5],   # 0.1 + 1.0 (shifted)
+			[0.95, 0.5, 0.5],  # -0.05 + 1.0 (unwrapped and shifted)
+			[1.05, 0.6, 0.5],  # 0.05 + 1.0 (shifted)
+			[1.05, 0.4, 0.5],  # 0.05 + 1.0 (shifted)
+			[1.05, 0.5, 0.6],  # 0.05 + 1.0 (shifted)
+			[1.05, 0.5, 0.4]   # 0.05 + 1.0 (shifted)
+		])
+		np.testing.assert_array_almost_equal(result, expected, decimal=7)
+	
+	def test_unwrap_vertices_vectorised_pathological_case(self):
[…]
