+++ tests/test_pbc_utils.py
@@ -5,7 +5,7 @@ from pymatgen.core import Structure, Lattice
 from site_analysis.polyhedral_site import PolyhedralSite
 from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.site import Site
-from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_centre, unwrap_vertices_vectorised
+from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_centre
 
 
 class TestLegacyPBCCorrection(unittest.TestCase):
@@ -117,8 +117,8 @@ class TestLegacyPBCCorrection(unittest.TestCase):
 		result = apply_legacy_pbc_correction(frac_coords)
 		
 		np.testing.assert_array_equal(result, frac_coords)
-	
-	
+
+
 class TestCurrentPBCBehaviorRegression(unittest.TestCase):
 	"""Regression tests to preserve correct current PBC behavior during refactoring."""
 
@@ -246,7 +246,8 @@ class TestCurrentPBCBehaviorRegression(unittest.TestCase):
 		# Should calculate centre correctly (no correction needed)
 		expected_centre = np.mean(coords, axis=0) % 1.0
 		np.testing.assert_array_almost_equal(site.centre, expected_centre, decimal=7)
-		
+
+
 class TestReferenceBasedUnwrapping(unittest.TestCase):
 	"""Tests for reference centre-based vertex unwrapping."""
 	
@@ -313,16 +314,9 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		# The centre of the unwrapped vertices should be close to the reference centre
 		result_centre = np.mean(result, axis=0)
 		np.testing.assert_array_almost_equal(result_centre, expected_result_centre, decimal=2)
-		
-class TestVectorisedUnwrapping(unittest.TestCase):
-	"""Tests for vectorised reference centre-based vertex unwrapping."""
 	
-	def setUp(self):
-		self.lattice = Lattice.cubic(10.0)
-	
-	def test_unwrap_vertices_vectorised_no_unwrapping_needed(self):
-		"""Test vectorised case where vertices form a tetrahedron around the reference centre."""
-		# Simple tetrahedron centered around [0.5, 0.5, 0.5]
+	def test_unwrap_vertices_tetrahedral_coordination(self):
+		"""Test 4-coordinate tetrahedral case."""
 		vertex_coords = np.array([
 			[0.4, 0.4, 0.4],
 			[0.6, 0.6, 0.4],
@@ -331,233 +325,20 @@ class TestVectorisedUnwrapping(unittest.TestCase):
 		])
 		reference_centre = np.array([0.5, 0.5, 0.5])
 		
-		result = unwrap_vertices_vectorised(vertex_coords, reference_centre, self.lattice)
+		result = unwrap_vertices_to_reference_centre(vertex_coords, reference_centre, self.lattice)
 		
 		# Should remain unchanged since all vertices are already close
 		np.testing.assert_array_almost_equal(result, vertex_coords, decimal=7)
-		
-	def test_unwrap_vertices_vectorised_simple_case(self):
-		"""Test vectorised unwrapping for an octahedron spanning the boundary at x=0."""
-		# Octahedron centered at [0.05, 0.5, 0.5] spanning the x=0 boundary
-		vertex_coords = np.array([
-			[0.1, 0.5, 0.5],   # +x vertex
-			[-0.05, 0.5, 0.5], # -x vertex
-			[0.05, 0.6, 0.5],  # +y vertex  
-			[0.05, 0.4, 0.5],  # -y vertex
[…]
