+++ site_analysis/pbc_utils.py
@@ -1,6 +1,8 @@
 """Utilities for handling periodic boundary conditions."""
 
 import numpy as np
+from pymatgen.core import Lattice
+from typing import cast
 
 def apply_legacy_pbc_correction(frac_coords: np.ndarray) -> np.ndarray:
 	"""Apply the legacy spread-based periodic boundary condition handling.
@@ -31,37 +33,17 @@ def apply_legacy_pbc_correction(frac_coords: np.ndarray) -> np.ndarray:
 			corrected_coords[corrected_coords[:, dim] < 0.5, dim] += 1.0
 	
 	return corrected_coords
-	
-def unwrap_vertices_to_reference_centre(frac_coords: np.ndarray, reference_centre: np.ndarray, lattice) -> np.ndarray:
-	"""Unwrap vertices to their closest periodic images relative to a reference centre.
-	
-	Args:
-		frac_coords: Array of fractional coordinates with shape (n, 3).
-		reference_centre: Reference centre position for unwrapping.
-		lattice: Lattice object for distance calculations.
-		
-	Returns:
-		Unwrapped fractional coordinates with the same shape, shifted to ensure all coordinates >= 0.
-	"""
-	# Handle empty input
-	if len(frac_coords) == 0:
-		return frac_coords.copy()
-	
-	result = np.zeros_like(frac_coords)
-	
-	for i, vertex in enumerate(frac_coords):
-		_, image = lattice.get_distance_and_image(reference_centre, vertex)
-		result[i] = vertex + image
-	
-	# Apply uniform shift to ensure all coordinates are non-negative
-	min_coords = np.min(result, axis=0)
-	shift = np.maximum(0, np.ceil(-min_coords))
-	result += shift
-	
-	return result
-	
-	
-def unwrap_vertices_vectorised(frac_coords: np.ndarray, reference_centre: np.ndarray, lattice) -> np.ndarray:
+
+# Generate all 27 possible shifts: [-1,0,1] for each dimension
+_PERIODIC_SHIFTS = np.array([[dx, dy, dz] for dx in [-1, 0, 1] 
+  										  for dy in [-1, 0, 1] 
+  										  for dz in [-1, 0, 1]])
+										
+def unwrap_vertices_to_reference_centre(
+	frac_coords: np.ndarray,
+	reference_centre: np.ndarray,
+	lattice: Lattice
+) -> np.ndarray:
 	"""Vectorised unwrapping of vertices to their closest periodic images relative to a reference centre.
 	
 	Args:
@@ -72,17 +54,8 @@ def unwrap_vertices_vectorised(frac_coords: np.ndarray, reference_centre: np.nda
 	Returns:
 		Unwrapped fractional coordinates with the same shape, shifted to ensure all coordinates >= 0.
 	"""
-	# Handle empty input
-	if len(frac_coords) == 0:
-		return frac_coords.copy()
-	
-	# Generate all 27 possible shifts: [-1,0,1] for each dimension
-	shifts = np.array([[dx, dy, dz] for dx in [-1, 0, 1] 
-									for dy in [-1, 0, 1] 
-									for dz in [-1, 0, 1]])  # (27, 3)
-	
 	# Apply all shifts using broadcasting: (n_vertices, 27, 3)
-	vertex_images = frac_coords[:, np.newaxis, :] + shifts[np.newaxis, :, :]
+	vertex_images = frac_coords[:, np.newaxis, :] + _PERIODIC_SHIFTS[np.newaxis, :, :]
 	
 	# Convert to Cartesian coordinates for true distance calculation
 	ref_cart = lattice.get_cartesian_coords(reference_centre)  # (3,)
[…]
