### 2025-06-03 9b95d9b
Reference PBC method
