### 2025-06-05 3f04e48
WIP: reference sites PBC handling
