+++ tests/test_reference_workflow/test_site_factory.py
@@ -25,42 +25,40 @@ class TestSiteFactory(unittest.TestCase):
 		
 		# Create a simple test structure
 		lattice = Lattice.cubic(5.0)
-		species = ["Na", "Cl", "Na", "Cl", "Na"]
+		species = ["Na", "Cl", "Na", "Cl", "Na", "Cl"]
 		coords = [
-			[0.0, 0.0, 0.0],  # Na0
-			[0.2, 0.2, 0.2],  # Cl1
-			[0.5, 0.0, 0.0],  # Na2
-			[0.7, 0.2, 0.2],  # Cl3
-			[0.0, 0.5, 0.0],  # Na4
+			[0.0, 0.0, 0.0], 
+			[0.2, 0.2, 0.2], 
+			[0.5, 0.0, 0.0], 
+			[0.7, 0.2, 0.2], 
+			[0.0, 0.5, 0.0],
+			[0.3, 0.3, 0.3] 
 		]
 		self.structure = Structure(lattice, species, coords)
+		self.site_factory = SiteFactory(self.structure)
 		
 		# Define some coordination environments as lists of atom indices
 		self.tetrahedral_env = [[1, 2, 3, 4]]  # A tetrahedral environment
 		self.multiple_envs = [
-			[1, 2, 4],     # First environment (3 atoms)
-			[0, 1, 3, 4]   # Second environment (4 atoms)
+			[1, 2, 4, 5],  # First environment (4 atoms)
+			[0, 1, 3, 4, 5]   # Second environment (5 atoms)
 		]
 	
 	def test_validate_environments_valid(self):
 		"""Test environment validation with valid environments."""
-		factory = SiteFactory(self.structure)
-		
 		# This should not raise any exception
-		factory._validate_environments(self.tetrahedral_env)
-		factory._validate_environments(self.multiple_envs)
+		self.site_factory._validate_environments(self.tetrahedral_env)
+		self.site_factory._validate_environments(self.multiple_envs)
 	
 	def test_validate_environments_invalid_type(self):
 		"""Test environment validation with invalid environment type."""
-		factory = SiteFactory(self.structure)
-		
 		# Not a list
 		with self.assertRaises(ValueError):
-			factory._validate_environments(123)
+			self.site_factory._validate_environments(123)
 		
 		# Not a list of lists
 		with self.assertRaises(ValueError):
-			factory._validate_environments([1, 2, 3])
+			self.site_factory._validate_environments([1, 2, 3])
 	
 	def test_validate_environments_invalid_indices(self):
 		"""Test environment validation with invalid indices."""
@@ -76,32 +74,26 @@ class TestSiteFactory(unittest.TestCase):
 	
 	def test_validate_labels_valid(self):
 		"""Test label validation with valid labels."""
-		factory = SiteFactory(self.structure)
-		
 		# Valid cases
-		factory._validate_labels(None, None, 2)  # No labels
-		factory._validate_labels("label", None, 2)  # Single label
-		factory._validate_labels(None, ["label1", "label2"], 2)  # Multiple labels
+		self.site_factory._validate_labels(None, None, 2)  # No labels
+		self.site_factory._validate_labels("label", None, 2)  # Single label
+		self.site_factory._validate_labels(None, ["label1", "label2"], 2)  # Multiple labels
 	
 	def test_validate_labels_invalid(self):
 		"""Test label validation with invalid labels."""
-		factory = SiteFactory(self.structure)
-		
 		# Both label and labels provided
 		with self.assertRaises(ValueError):
-			factory._validate_labels("label", ["label1", "label2"], 2)
+			self.site_factory._validate_labels("label", ["label1", "label2"], 2)
 		
 		# Wrong number of labels
 		with self.assertRaises(ValueError):
-			factory._validate_labels(None, ["label1"], 2)  # Too few
[…]

[…]
