### 2025-06-05 f507e7a
WIP: reference sites PBC handling
