+++ tests/test_reference_workflow/test_reference_based_sites.py
@@ -251,118 +251,101 @@ class TestReferenceBasedSites(unittest.TestCase):
         with self.coord_finder_patch, self.index_mapper_patch, self.site_factory_patch:
             rbs = ReferenceBasedSites(self.reference, self.target, align=False)
             
-            # Create polyhedral sites
-            sites = rbs.create_polyhedral_sites(
-                center_species='Na',
-                vertex_species='Cl',
-                cutoff=3.0,
-                n_vertices=2,
-                label='test_label',
-                target_species='Cl'
-            )
-            
-            # Check that find_coordination_environments was called
-            self.mock_coord_finder.find_environments.assert_called_with(
-                center_species='Na',
-                coordination_species='Cl',
-                n_coord=2,
-                cutoff=3.0
-            )
-            
-            # Check that map_coordinating_atoms was called
-            self.mock_index_mapper.map_coordinating_atoms.assert_called_with(
-                self.reference,
-                self.target,
-                self.mapped_environments,
-                target_species='Cl'
-            )
-            
-            # Check that create_polyhedral_sites was called
-            self.mock_site_factory.create_polyhedral_sites.assert_called_with(
-                self.mapped_environments,
-                label='test_label',
-                labels=None
-            )
-            
-            # Check returned sites
-            self.assertEqual(sites, self.mock_site_factory.create_polyhedral_sites.return_value)
+            # Mock the _calculate_reference_centers_from_indices method
+            mock_reference_centers = [np.array([0.1, 0.1, 0.1])]
+            with patch.object(rbs, '_calculate_reference_centers_from_indices') as mock_calc_centers:
+                mock_calc_centers.return_value = mock_reference_centers
+                
+                # Create polyhedral sites
+                sites = rbs.create_polyhedral_sites(
+                    center_species='Na',
+                    vertex_species='Cl',
+                    cutoff=3.0,
+                    n_vertices=2,
+                    label='test_label',
+                    target_species='Cl'
+                )
+                
+                # Check that find_coordination_environments was called
+                self.mock_coord_finder.find_environments.assert_called_with(
+                    center_species='Na',
+                    coordination_species='Cl',
+                    n_coord=2,
+                    cutoff=3.0
+                )
+                
+                # Check that _calculate_reference_centers_from_indices was called
+                mock_calc_centers.assert_called_once_with([0])  # center indices from ref_environments.keys()
+                
+                # Check that map_coordinating_atoms was called
+                self.mock_index_mapper.map_coordinating_atoms.assert_called_with(
+                    self.reference,
+                    self.target,
+                    self.mapped_environments,
+                    target_species='Cl'
+                )
+                
+                # Check that create_polyhedral_sites was called with reference_centers
+                self.mock_site_factory.create_polyhedral_sites.assert_called_with(
+                    self.mapped_environments,
+                    reference_centers=mock_reference_centers,
+                    label='test_label',
+                    labels=None
+                )
+                
+                # Check returned sites
+                self.assertEqual(sites, self.mock_site_factory.create_polyhedral_sites.return_value)
     
     def test_create_dynamic_voronoi_sites(self):
         """Test the create_dynamic_voronoi_sites method."""
         with self.coord_finder_patch, self.index_mapper_patch, self.site_factory_patch:
             rbs = ReferenceBasedSites(self.reference, self.target, align=False)
             
-            # Create dynamic voronoi sites
-            sites = rbs.create_dynamic_voronoi_sites(
-                center_species='Na',
-                reference_species='Cl',
-                cutoff=3.0,
-                n_reference=2,
-                labels=['site1'],
-                target_species='Cl'
-            )
[…]
