+++ site_analysis/builders.py
@@ -56,7 +56,7 @@ from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.trajectory import Trajectory
 from site_analysis.reference_workflow.reference_based_sites import ReferenceBasedSites
 import numpy as np
-from typing import Union, Optional, cast, Callable, Sequence, Any
+from typing import cast, Callable, Sequence, Any
 
 
 class TrajectoryBuilder:
@@ -124,21 +124,21 @@ class TrajectoryBuilder:
 		Returns:
 			self: For method chaining
 		"""
-		self._structure: Optional[Structure] = None
-		self._reference_structure: Optional[Structure]= None
-		self._mobile_species: Optional[str|list[str]] = None
-		self._atoms: Optional[list[Atom]] = None
+		self._structure: Structure | None = None
+		self._reference_structure: Structure | None= None
+		self._mobile_species: str | list[str] | None = None
+		self._atoms: list[Atom] | None = None
 		
 		# Alignment options
 		self._align = True
-		self._align_species: Optional[list[str]] = None
+		self._align_species: list[str] | None = None
 		self._align_metric = 'rmsd'
 		self._align_algorithm = 'Nelder-Mead'
-		self._align_minimizer_options: Optional[dict[str, Any]] = None
+		self._align_minimizer_options: dict[str, Any] | None = None
 		self._align_tolerance = 1e-4  # Default tolerance value
 		
 		# Mapping options
-		self._mapping_species: Optional[list[str]] = None
+		self._mapping_species: list[str] | None = None
 		
 		# Functions to be called during build() to create sites
 		self._site_generators: list[Callable] = []
@@ -170,7 +170,7 @@ class TrajectoryBuilder:
 		self._reference_structure = reference_structure
 		return self
 		
-	def with_mobile_species(self, species: Union[str, list[str]]) -> TrajectoryBuilder:
+	def with_mobile_species(self, species: str | list[str]) -> TrajectoryBuilder:
 		"""Set the mobile species to track.
 		
 		Args:
@@ -184,10 +184,10 @@ class TrajectoryBuilder:
 	
 	def with_structure_alignment(self, 
 				align: bool = True, 
-				align_species: Optional[Union[str, list[str]]] = None, 
+				align_species: str | list[str] | None = None, 
 				align_metric: str = 'rmsd',
 				align_algorithm: str = 'Nelder-Mead',
-				align_minimizer_options: Optional[dict[str, Any]] = None,
+				align_minimizer_options: dict[str, Any] | None = None,
 				align_tolerance: float = 1e-4) -> 'TrajectoryBuilder':
 		"""Set options for aligning reference and target structures.
 		
@@ -255,7 +255,7 @@ class TrajectoryBuilder:
 		self._align_tolerance = align_tolerance
 		return self
 		
-	def with_site_mapping(self, mapping_species: Optional[Union[str, list[str]]]) -> TrajectoryBuilder:
+	def with_site_mapping(self, mapping_species: str | list[str] | None) -> TrajectoryBuilder:
 		"""Set the species to use for mapping sites between reference and target structures.
 		
 		Site mapping identifies corresponding sites between structures even when
@@ -284,8 +284,8 @@ class TrajectoryBuilder:
 		
 	def with_spherical_sites(self, 
 					centres: list[list[float]], 
-					radii: Union[float, list[float]], 
-					labels: Optional[Union[str, list[str]]] = None) -> TrajectoryBuilder:
+					radii: float | list[float], 
+					labels: str | list[str] | None = None) -> TrajectoryBuilder:
 		"""Define spherical sites.
 		
 		Note: Sites will be generated when build() is called.
@@ -334,7 +334,7 @@ class TrajectoryBuilder:
 		
 	def with_voronoi_sites(self, 
[…]
