### 2025-06-05 f042af0
Fixed bug in mapping central atom indices to coordination environments
