### 2025-06-05 be60dd8
Removed debug print statement
