### 2025-06-05 59b4121
Updated builder docstrings
