### 2025-06-05 d72f754
Improved docstrings
