### 2025-06-06 5b8739d
Added PBC options documentation
