### 2025-06-06 ff4cea4
Bumped to v1.1.0
