### 2025-06-06 8f0236f
Fixed error in PBC handling docs
