### 2025-06-06 b63bc21
Bumped to v1.1.2
