### 2025-06-07 972c4c5
Polyhedral site assignment optimisation
