### 2025-06-07 8592e91
Bumped to v1.2.0
