### 2025-06-07 788f2ed
Fixing mypy errors
