### 2025-06-07 3d1e6a8
Implemented priority site checking for Spherical Sites
