+++ site_analysis/spherical_site_collection.py
@@ -21,16 +21,114 @@ without generating spurious site transitions.
 """
 
 import numpy as np
-from typing import List
-from pymatgen.core import Structure
+from typing import Iterator
+from pymatgen.core import Structure, Lattice
 from site_analysis.atom import Atom
+from site_analysis.spherical_site import SphericalSite
 from site_analysis.site_collection import SiteCollection
 
 class SphericalSiteCollection(SiteCollection):
 
 
+    def __init__(self,
+        sites: list[SphericalSite],
+        neighbour_cutoff: float = 10.0) -> None:
+        """Create a SphericalSiteCollection instance.
+        
+        Args:
+            sites: List of SphericalSite objects.
+            neighbour_cutoff: Distance cutoff in Ångström for 
+                determining neighbouring sites. Default is 10.0 Å.
+        
+        Returns:
+            None
+        
+        """
+        for s in sites:
+            if not isinstance(s, SphericalSite):
+                raise TypeError
+        super(SphericalSiteCollection, self).__init__(sites)
+        self.sites = self.sites  # type: list[SphericalSite]
+        self._neighbouring_sites: dict[int, list[SphericalSite]] | None = None
+        self._current_lattice: Lattice | None = None
+        self._neighbour_cutoff = neighbour_cutoff
+        
+    def neighbouring_sites(self, index: int) -> list[SphericalSite]:
+        """Returns the neighbouring sites for a given site index.
+        
+        Neighbours are defined as sites within 10.0 Å of the given site.
+        The neighbour calculation is performed lazily on first access and cached.
+        
+        Args:
+            index (int): Index of the site to return neighbours for.
+            
+        Returns:
+            list[SphericalSite]: List of neighbouring SphericalSite objects.
+            
+        Raises:
+            RuntimeError: If no lattice is available for distance calculations.
+        """
+        if self._neighbouring_sites is None:
+            lattice = self._get_current_lattice()
+            if lattice is None:
+                raise RuntimeError("No lattice available for neighbour calculation. "
+                                "Call assign_site_occupations or analyse_structure first.")
+            self._neighbouring_sites = self._calculate_all_neighbouring_sites(lattice)
+        
+        return self._neighbouring_sites[index]
+        
+    def _get_current_lattice(self) -> Lattice | None:
+        """Get the current lattice for distance calculations.
+        
+        Returns:
+            Lattice or None: The current lattice if available, None otherwise.
+        """
+        return self._current_lattice
+        
+    def _calculate_all_neighbouring_sites(self,
+        lattice: Lattice
+    ) -> dict[int, list[SphericalSite]]:
+        """Calculate all site-site distances and determine neighbours within 10.0 Å.
+        
+        Args:
+            lattice : Pymatgen Lattice object for distance calculations.
+            
+        Returns:
+            dict: Dictionary mapping site indices to lists
+                of neighbouring SphericalSite objects within 10.0 Å, sorted by 
+                increasing distance.
+        """
+        if not self.sites:
+            return {}
+        
+        # Get coordinates of all sites
+        site_coords = np.array([site.frac_coords for site in self.sites])
+        
+        # Calculate all pairwise distances
+        distance_matrix = lattice.get_all_distances(site_coords, site_coords)
+        
+        # Set diagonal to infinity to exclude self-references
+        np.fill_diagonal(distance_matrix, np.inf)
+        
+        # Create mask for valid neighbors
+        neighbor_mask = distance_matrix <= self._neighbour_cutoff
+        
+        # Build neighbour dictionary
+        neighbours = {}
+        for i, site_i in enumerate(self.sites):
+            neighbor_indices = np.where(neighbor_mask[i])[0]
[…]
