### 2025-06-07 7804c67
Improved docstrings
