### 2025-06-07 044e70c
Closes https://github.com/bjmorgan/site-analysis/issues/10
