### 2025-06-07 63ec84f
Improved quickstart example
