+++ .github/workflows/python-publish.yml
@@ -13,7 +13,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v3
         with:
-          ref: ${{ github.event.release.tag_name }}  # ← Add this line
+          ref: ${{ github.event.release.tag_name }}
           
       - name: Set up Python
         uses: actions/setup-python@v4
