### 2025-06-09 6bfefca
Added JOSS badge to docs
