### 2025-08-05 2528162
Updated minimum setuptools to v64
