### 2025-08-08 450affc
Adding numpy array inputs for PolyhedralSite
