### 2025-09-01 25a2828
Added description to tqdm output
