+++ .github/workflows/python-publish.yml
@@ -1,37 +1,52 @@
-name: Upload Python Package
+name: Build and publish to PyPI
 
 on:
   release:
-    types:
-      - created
+    types: [published]
 
 jobs:
-  publish:
+  build-and-publish:
+    name: Build and publish Python distributions to PyPI
     runs-on: ubuntu-latest
+    environment: release
+    permissions:
+      id-token: write
+      contents: read
     
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-        with:
-          ref: ${{ github.event.release.tag_name }}  # ← Add this line
-          
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.11'
-          
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install build twine
-          
-      - name: Build package
-        run: python -m build
-        
-      - name: List distribution files
-        run: ls -la dist/
-        
-      - name: Publish package
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          password: ${{ secrets.PYPI_API_TOKEN }}
+    - name: Checkout code
+      uses: actions/checkout@v4
+      with:
+        ref: ${{ github.event.release.tag_name }}
+    
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: '3.11'
+    
+    - name: Install build dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install build twine
+    
+    - name: Build wheel and source distribution
+      run: python -m build
+    
+    - name: Check distribution
+      run: |
+        ls -la dist/
+        python -m twine check dist/*
+    
+    - name: Test installation
+      run: |
+        python -m pip install dist/*.whl
+        python -c "import site_analysis; print(f'✅ site-analysis {site_analysis.__version__} imported successfully')"
+    
+    - name: Upload artifacts
+      uses: actions/upload-artifact@v4
+      with:
+        name: python-package-distributions
+        path: dist/
+    
+    - name: Publish to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
