### 2025-10-26 e68a39a
Refactor build workflow for Python dependencies and tests
Updated Python dependencies installation and test commands in the build workflow.
