+++ .github/workflows/build.yml
@@ -1,6 +1,6 @@
 name: build
 
-# Controls when the action will run. 
+# Controls when the action will run
 on: [push, pull_request]
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
@@ -23,44 +23,20 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install -U numpy
-          pip install mypy
-          pip install types-PyYAML
-          pip install pytest
-          pip install pytest-cov
-          pip install pylint
-          pip install .
+          # Install package with dev dependencies
+          pip install -e ".[dev]"
           pip list
       
       - name: Run tests
         run: |
-          pytest --cov-config=.coveragerc --cov=vasppy --cov-report lcov
+          pytest --cov-config=.coveragerc --cov=site_analysis --cov-report lcov
       
       - name: Coveralls GitHub Action
-        uses: coverallsapp/github-action@v2.3.6
+        uses: coverallsapp/github-action@v2
         with:
           path-to-lcov: ./coverage.lcov
           github-token: ${{ secrets.GITHUB_TOKEN }}
-          parallel: true
-          flag-name: python-${{ matrix.python-version }}
       
       - name: Static type checking
         run: |
-          mypy vasppy
-      
-      - name: PyLint
-        run: |
-          pylint --errors-only vasppy
-
-  # This job finalizes the parallel coverage reports
-  finish:
-    needs: tests
-    if: ${{ always() }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Coveralls Finished
-        uses: coverallsapp/github-action@v2.3.6
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          parallel-finished: true
+          mypy site_analysis
