### 2025-11-16 #8279 @Kevin-Mattheus-Moerman
@gezelter, @orionarcher, @amritagos, @arose, @hmacdope, @dengzeyu, @cxhernandez, @anotherjoshsmith, @marjanalbooyeh, @LIVazquezS, @hejung, @colinbousige, @helmutcarter, @daico007 would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org/)? The work is titled site-analysis: A Python package for site-projection analysis of molecular dynamics trajectories.

JOSS reviews take place on GitHub and focus on [the software 💾 ](https://github.com/bjmorgan/site-analysis)as well as [a short paper 📜 ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08279/joss.08279/10.21105.joss.08279.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Here is an example of an finished review issue if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7004.

If you are interested you can let me know by responding here.

Thanks!
