### 2025-11-25 #8279 @jkrajniak
> [@gezelter](https://github.com/gezelter), [@orionarcher](https://github.com/orionarcher), [@amritagos](https://github.com/amritagos), [@arose](https://github.com/arose), [@hmacdope](https://github.com/hmacdope), [@dengzeyu](https://github.com/dengzeyu), [@cxhernandez](https://github.com/cxhernandez), [@anotherjoshsmith](https://github.com/anotherjoshsmith), [@marjanalbooyeh](https://github.com/marjanalbooyeh), [@hejung](https://github.com/hejung), [@colinbousige](https://github.com/colinbousige), [@helmutcarter](https://github.com/helmutcarter), [@daico007](https://github.com/daico007), [@samwaseda](https://github.com/samwaseda), [@prgupta-cpfam](https://github.com/prgupta-cpfam), [@yw-fang](https://github.com/yw-fang), [@IonRepo](https://github.com/IonRepo), [@braniii](https://github.com/braniii), [@corettialessandro](https://github.com/corettialessandro), [@jkrajniak](https://github.com/jkrajniak), [@samuelmurail](https://github.com/samuelmurail), [@foeroyingur](https://github.com/foeroyingur), [@ryokbys](https://github.com/ryokbys), [@Xia-Yijie](https://github.com/Xia-Yijie), [@foeroyingur](https://github.com/foeroyingur) would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org/)? The work is titled site-analysis: A Python package for site-projection analysis of molecular dynamics trajectories.
>
> JOSS reviews take place on GitHub and focus on [the software 💾 ](https://github.com/bjmorgan/site-analysis)as well as [a short paper 📜 ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08279/joss.08279/10.21105.joss.08279.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Here is an example of an finished review issue if you want to see what it is like: [#7004](https://github.com/openjournals/joss-reviews/issues/7004).
>
> If you are interested you can let me know by responding here.
>
> Thanks!

sure!
