### 2025-12-11 #9606 @editorialbot
[REVIEW]: site-analysis: A Python package for site-projection analysis of molecular dynamics trajectories
**Submitting author:** @bjmorgan (<a href="http://orcid.org/0000-0002-3056-8233">Benjamin J. Morgan</a>)
**Repository:** https://github.com/bjmorgan/site-analysis
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
