### 2026-01-17 bjmorgan/site-analysis#31 @jkrajniak
Quick start / missing xdatcar
The example provided in README.md does not work out-of-box
```
Traceback (most recent call last):
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/ex.py", line 5, in <module>
    xdatcar = Xdatcar("XDATCAR")
              ^^^^^^^^^^^^^^^^^^
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/.venv/lib/python3.11/site-packages/pymatgen/io/vasp/outputs.py", line 4678, in __init__
    file_len = sum(1 for _ in zopen(filename, mode="rt", encoding="utf-8"))
                              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/.venv/lib/python3.11/site-packages/monty/io.py", line 116, in zopen
    return open(filename, mode, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: 'XDATCAR'
```

I suppose I should get XDATCAR file from somewhere. I tried from examples, but then another issue
```
Traceback (most recent call last):
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/ex.py", line 16, in <module>
    .build())
     ^^^^^^^
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/.venv/lib/python3.11/site-packages/site_analysis/builders.py", line 678, in build
    trajectory = Trajectory(sites=sites, atoms=self._atoms)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/jakubkrajniak/Work/Science/joss/ex-site-analysis/.venv/lib/python3.11/site-packages/site_analysis/trajectory.py", line 71, in __init__
    raise ValueError("Cannot initialize Trajectory with empty atoms list")
ValueError: Cannot initialize Trajectory with empty atoms list
```

here is the code
```py
from site_analysis.builders import TrajectoryBuilder
from pymatgen.io.vasp import Xdatcar

# Load MD trajectory from VASP XDATCAR file
xdatcar = Xdatcar("XDATCAR")
md_structures = xdatcar.structures

# Define multiple sites and track Li+ ion movements between them
trajectory = (TrajectoryBuilder()
              .with_structure(md_structures[0])  # Use first frame as reference
              .with_mobile_species("Li")
              .with_spherical_sites(centres=[[0.0, 0.0, 0.0],
                                             [0.5, 0.5, 0.5],
                                             [0.0, 0.5, 0.0]],
                                    radii=1.5)
              .build())

trajectory.trajectory_from_structures(md_structures)

# Get site occupancies over time
print(trajectory.atoms_trajectory)  # Which site each atom occupies
print(trajectory.sites_trajectory)  # Which atoms in each site
```
