### 2026-01-17 4b5066f
Fix quickstart example to work out-of-box (fixes #31)
- Add example_data/XDATCAR with minimal Li trajectory for quickstart
- Update README quickstart to use example_data/XDATCAR with matching
  site centres and explanatory comments
- Add helpful error message when mobile species not found in structure,
  listing available species
- Fix docs quickstart notebook path case (xdatcar -> XDATCAR) for Linux
- Remove obsolete example/ directory (notebooks superseded by formal docs)
