### 2026-01-17 89a2247
Bump version to 1.2.5
