### 2026-01-17 #9606 @bjmorgan
> One problem with README and Quickstart, reported here: [bjmorgan/site-analysis#31](https://github.com/bjmorgan/site-analysis/issues/31)

Thanks for flagging this. It should be addressed by https://github.com/bjmorgan/site-analysis/pull/32
