### 2026-02-21 9923b76
Remove debug print statement from PolyhedralSite.__init__
