### 2026-02-21 998df54
Fix float distance truncation in get_coordination_indices
The neighbour distance was being cast to int, which truncated distances
like 1.8 A to 1. This broke the distance-based sorting of returned
neighbour indices.
