### 2026-02-21 c63203d
Fix append_timestep dropping timestep t=0
The truthiness check `if t:` treated t=0 as falsy, so timestep 0 was
silently not recorded. Changed to `if t is not None:`.
