### 2026-02-21 8fa7caa
Fix transitions from site index 0 not being recorded
The truthiness check `if previous_site_index:` treated index 0 as
falsy, so transitions from site 0 were silently ignored. Changed to
`if previous_site_index is not None:`.
