### 2026-02-21 0b8b191
Fix contains_point_sn appending inside the wrong loop
The inside.append() call was indented inside the inner face loop
instead of the outer point loop, causing it to check partial dotsum
values after each face rather than the final sum after all faces.
