### 2026-02-21 957d9c7
Remove dead code and variable shadowing in Trajectory.__init__
The variable site_type was assigned on one line then immediately
shadowed by the for-loop variable. Removed the dead assignment and
renamed the loop variable to registered_type for clarity.
