### 2026-02-21 1b36568
Fix x_pbc docstring: returns (8,3) not (9,3)
