### 2026-02-21 7d30a56
Bump version to 1.2.6
