### 2026-02-21 25261a9
Replace type narrowing asserts with local variables in reference_based_sites.py
Use local variable assignment instead of assert statements for mypy
type narrowing after lazy initialisation of _site_factory and
_coord_finder.
