### 2026-02-21 3782f58
Remove redundant assert statements in tools.py and trajectory.py
Remove dead-code asserts in site_index_mapping where the preceding
coercion logic already guarantees the variables are lists. Also remove
a commented-out assert in Trajectory.append_timestep.
