### 2026-02-21 db4c69e
Replace defensive assert statements with proper exceptions
Use RuntimeError when vertex coordinates are unset in
PolyhedralSite.centre, and TypeError when a non-Structure argument
is passed to sites_contain_points. Update corresponding tests to
expect the new exception types.
