### 2026-02-21 c612e7e
Fix incorrect docstrings
- Fix "Retuns" typo in Site.__init__
- Fix reset_index docstring claiming default is 1 (actual default is 0)
- Remove erroneous "Returns: None" from Atom.__init__
- Document sites parameter in construct_neighbouring_sites
- Fix parameter name x -> x_list in contains_point_sn docstring
