### 2026-02-21 829c27d
Replace deprecated typing imports with built-in equivalents
