### 2026-02-21 403352e
Use public APIs instead of accessing private attributes
- atoms_from_structure: use assign_coords() instead of setting
  atom._frac_coords directly
- species_string_from_site: use site.species_string instead of
  accessing site._species
