### 2026-02-21 5f7fd4a
Avoid mutating caller's minimizer_options dict in StructureAligner
Pop 'bounds' from the local options copy instead of from the caller's
minimizer_options dict.
