+++ site_analysis/reference_workflow/index_mapper.py
@@ -24,157 +24,157 @@ from pymatgen.core import Structure
 
 
 class IndexMapper:
-	"""Maps coordinating atom indices between reference and target crystal structures.
-	
-	Used to translate coordination environments defined in an ideal reference
-	structure to corresponding atoms in a target structure, handling permutations
-	and structural distortions via distance-based matching.
-	
-	The mapper verifies 1:1 correspondence between reference and target atoms,
-	ensuring each coordinating position maps to exactly one target atom. If
-	this constraint is violated, a ValueError is raised.
-	"""
-	
-	def map_coordinating_atoms(
-		self,
-		reference: Structure,
-		target: Structure,
-		ref_coordinating: list[list[int]],
-		target_species: str | list[str] | None = None
-	) -> list[list[int]]:
-		"""Map coordinating atom indices from reference to target structure.
-		
-		Args:
-			reference: Reference structure containing ideal coordination environments.
-			target: target structure with potentially distorted or permuted atoms.
-			ref_coordinating: List of coordinating atom index lists from reference.
-				Each sublist contains indices of atoms that define a site.
-			target_species: Optional filter for target atom species to map to.
-				If specified, only maps to atoms of these species in the target structure.
-				
-		Returns:
-			List of coordinating atom index lists mapped to the target structure.
-			Maintains the same structure as input but with updated indices.
-			
-		Raises:
-			ValueError: If 1:1 mapping cannot be achieved (e.g., missing atoms,
-				ambiguous distances, or insufficient target atoms in target structure).
-		"""
-		# Extract all unique coordinating atoms from reference structure
-		unique_indices = self._extract_unique_coordinating_atoms(ref_coordinating)
-		
-		# Create mapping from reference to target structure
-		index_mapping = self._find_closest_atom_mapping(
-			reference, target, unique_indices, target_species
-		)
-		
-		# Map the coordination lists using the established mapping
-		mapped_coordinating = self._apply_mapping(ref_coordinating, index_mapping)
-		
-		return mapped_coordinating
-	
-	def _extract_unique_coordinating_atoms(
-		self,
-		ref_coordinating: list[list[int]]
-	) -> list[int]:
-		"""Extract unique coordinating atom indices from coordination lists.
-		
-		Args:
-			ref_coordinating: List of coordinating atom index lists.
-			
-		Returns:
-			Sorted list of unique coordinating atom indices.
-		"""
-		unique_indices = set()
-		for coord_list in ref_coordinating:
-			unique_indices.update(coord_list)
-		return sorted(list(unique_indices))
-	
-	def _find_closest_atom_mapping(
-		self,
-		reference: Structure,
-		target: Structure,
-		ref_indices: list[int],
-		target_species: str | list[str] | None
-	) -> dict[int, int]:
-		"""Find closest atom in target structure for each reference atom.
-		
-		Args:
-			reference: Reference structure.
-			target: Target structure.
-			ref_indices: List of reference atom indices to map.
-			target_species: Optional species filter for the target structure.
-			
-		Returns:
-			Dictionary mapping reference indices to target indices.
-			
-		Raises:
-			ValueError: If 1:1 mapping cannot be achieved.
-		"""
-		# Ensure that target_species is a list if it is specified
-		if isinstance(target_species, str):
-			target_species = [target_species]
-			
-		# Create a filtered list of target atoms in the target structure
-		if target_species is not None:
-			target_mask = np.array([site.species_string in target_species for site in target])
-			if not np.any(target_mask):
-				raise ValueError(f"No atoms of species {target_species} found in target structure")
-		else:
-			target_mask = np.ones(len(target), dtype=bool)
-			
[…]
