+++ site_analysis/dynamic_voronoi_site.py
@@ -14,217 +14,217 @@ from pymatgen.core import Structure
 from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_center
 
 class DynamicVoronoiSite(Site):
-	"""Site subclass corresponding to Voronoi cells with centres dynamically 
-	calculated from the positions of sets of reference atoms.
-	
-	Unlike standard VoronoiSite objects, which have fixed centers, the positions
-	of DynamicVoronoiSite objects adapt to structural changes as the reference 
-	atoms move. The site center is calculated as the mean position of the reference 
-	atoms, with special handling for periodic boundary conditions.
-	
-	This makes DynamicVoronoiSite particularly useful for tracking sites in mobile 
-	frameworks where the crystal structure deforms during simulation.
-	
-	Similar to standard Voronoi sites, a single DynamicVoronoiSite cannot determine
-	whether an atom is contained within it, as this depends on the positions of all
-	other sites. The coordination number of a DynamicVoronoiSite is defined as the 
-	number of reference atoms used to calculate its center.
-	
-	Attributes:
-		reference_indices (list[int]): Indices of atoms used as reference to calculate
-			the dynamic centre of the site.
-		centre_coords (np.ndarray or None): Fractional coordinates of the dynamically 
-			calculated site centre. This is None initially and calculated on demand.
-			
-	See Also:
-		:class:`~site_analysis.site.Site`: Parent class documenting inherited attributes
-			(index, label, contains_atoms, trajectory, points, transitions, average_occupation).
-			
-	Important:
-		This class is designed to be used with 
-		:class:`~site_analysis.dynamic_voronoi_site_collection.DynamicVoronoiSiteCollection`,
-		which handles the Voronoi tessellation logic.
-	"""
-		
-	def __init__(self,
-		reference_indices: list[int],
-		label: str | None = None,
-		reference_center: np.ndarray | None = None) -> None:
-		"""Create a ``DynamicVoronoiSite`` instance.
-			
-		Args:
-			reference_indices: List of atom indices whose positions will be used to dynamically calculate the centre of this site.
-			label: Optional label for this site.
-			reference_center: Optional reference centre for PBC handling.
-		
-		Returns:
-			None
-		"""
-		super(DynamicVoronoiSite, self).__init__(label=label)
-		self.reference_indices = reference_indices
-		self._centre_coords: np.ndarray | None = None
-		self.reference_center = reference_center
-		
-	def __repr__(self) -> str:
-		string = ('site_analysis.DynamicVoronoiSite('
-				f'index={self.index}, '
-				f'label={self.label}, '
-				f'reference_indices={self.reference_indices}, '
-				f'contains_atoms={self.contains_atoms})')
-		return string
+    """Site subclass corresponding to Voronoi cells with centres dynamically 
+    calculated from the positions of sets of reference atoms.
+    
+    Unlike standard VoronoiSite objects, which have fixed centers, the positions
+    of DynamicVoronoiSite objects adapt to structural changes as the reference 
+    atoms move. The site center is calculated as the mean position of the reference 
+    atoms, with special handling for periodic boundary conditions.
+    
+    This makes DynamicVoronoiSite particularly useful for tracking sites in mobile 
+    frameworks where the crystal structure deforms during simulation.
+    
+    Similar to standard Voronoi sites, a single DynamicVoronoiSite cannot determine
+    whether an atom is contained within it, as this depends on the positions of all
+    other sites. The coordination number of a DynamicVoronoiSite is defined as the 
+    number of reference atoms used to calculate its center.
+    
+    Attributes:
+        reference_indices (list[int]): Indices of atoms used as reference to calculate
+            the dynamic centre of the site.
[…]

[…]
