### 2026-02-21 318c565
Remove redundant object inheritance from Atom and Trajectory
