### 2026-02-21 9362d1e
Replace self-assignment type narrowing hacks with annotations
Use `self.sites: list[ConcreteType]` annotation instead of the no-op
`self.sites = self.sites  # type: list[ConcreteType]` pattern in all
four SiteCollection subclasses.
