### 2026-02-21 b0af196
Address PR review comments
- Change one_to_one_mapping and return_mapping_distances from
  bool | None to plain bool in site_index_mapping
- Add test that _run_differential_evolution does not mutate the
  caller's minimizer_options dict
