### 2026-02-21 d604561
Consolidate lazy initialisation in ReferenceBasedSites
Use _initialise_site_factory() at both call sites instead of
duplicating the None-check-and-init pattern. Extract a matching
_initialise_coord_finder() helper for consistency.
