### 2026-02-21 c162dbe
Only upload coverage from one matrix entry to avoid Coveralls conflict
