+++ .github/workflows/build.yml
@@ -1,9 +1,7 @@
 name: build
 
-# Controls when the action will run
 on: [push, pull_request]
 
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
   tests:
     runs-on: ubuntu-latest
@@ -13,30 +11,41 @@ jobs:
         python-version: ['3.10', '3.11', '3.12', '3.13']
     steps:
       - uses: actions/checkout@v4
-      
+
       - name: Set up Python ${{matrix.python-version}}
         uses: actions/setup-python@v5
         with:
           python-version: ${{matrix.python-version}}
           cache: 'pip'
-      
+
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          # Install package with dev dependencies
           pip install -e ".[dev]"
           pip list
-      
+
       - name: Run tests
         run: |
           pytest --cov-config=.coveragerc --cov=site_analysis --cov-report lcov
-      
-      - name: Coveralls GitHub Action
+
+      - name: Upload coverage to Coveralls
         uses: coverallsapp/github-action@v2
         with:
           path-to-lcov: ./coverage.lcov
           github-token: ${{ secrets.GITHUB_TOKEN }}
-      
+          parallel: true
+          flag-name: python-${{matrix.python-version}}
+
       - name: Static type checking
         run: |
           mypy site_analysis
+
+  coverage:
+    needs: tests
+    if: ${{ always() }}
+    runs-on: ubuntu-latest
+    steps:
+      - name: Close parallel build
+        uses: coverallsapp/github-action@v2
+        with:
+          parallel-finished: true
