### 2026-02-21 dd5f33d
Fix mypy errors introduced by indentation and typing changes
- Add None guards for vertex_coords in delaunay property and
  contains_point_sn to satisfy mypy type narrowing
- Add return type annotation to _initialise_site_factory
- Add type: ignore comments for scipy minimize/differential_evolution
  calls where type stubs are incomplete
- Update contains_point_sn tests to set vertex_coords on the site
