### 2026-02-21 b32421b
Add scipy-stubs as a dev dependency
Ensures CI has the same scipy type stubs available as local
development. The type: ignore comments in structure_aligner.py
are needed because the **options dict-unpacking pattern is
incompatible with scipy-stubs' strict per-parameter types.
