+++ pyproject.toml
@@ -31,6 +31,7 @@ dev = [
     "pytest-cov",           # Pytest coverage plugin
     "mypy",                 # Type checking
     "types-PyYAML",         # Type stubs for PyYAML
+    "scipy-stubs",          # Type stubs for SciPy
     "sphinx",               # Documentation generation
     "nbsphinx",             # For notebook integration in docs
     "sphinx-rtd-theme",     # Documentation theme
