### 2026-02-21 238d6a9
Fix docstring indentation in create_dynamic_voronoi_sites
