### 2026-02-21 b981320
Bump version to 1.2.7
