### 2026-02-24 #9606 @Xia-Yijie
Happy Chinese New Year everyone!

I have thoroughly reviewed the package, and overall, it is an excellent piece of work. The completion level is high, and the documentation and tutorials are very detailed and easy to follow.

I do have three minor suggestions to further improve the project:

**1. Visibility of Contribution Guidelines:**
While I noticed there is a CONTRIBUTING.md file in the repository, its presence is not very prominent. I recommend adding a clear "Contributing" section directly in the README.md or  the documentation, linking to the CONTRIBUTING.md file. This will make it much easier for the community to know how to engage with the project.

**2. Highlighting Generality of MD Engine Support:**
Currently, the quick start and documentation heavily emphasize VASP (XDATCAR). However, looking at the code itself, since the trajectory input relies on pymatgen, it seems that trajectories from any MD engine supported by pymatgen (such as LAMMPS) can be analyzed using your package. Explicitly mentioning this compatibility could significantly broaden your potential user base and prevent you from losing users of other MD engines.

**3. Details on Atom-in-Site Assignment Criteria:**
The documentation currently provides a somewhat brief description of the exact criteria used for atom-in-site assignments. It would be highly beneficial to include more detailed guidance on how to adjust the underlying parameters that dictate whether an atom is considered strictly inside or outside a site. I think elaborating on this is particularly important for edge cases. Furthermore, as this is academic software, explicitly citing relevant literature or theoretical background for these assignment algorithms (e.g., Voronoi or polyhedral decomposition) in the documentation would greatly help users understand the physical and mathematical justifications behind these criteria and default parameters.

Overall, I think the software and the paper are in great shape , and are fully acceptable. Excellent work again!
