### 2026-02-26 80cfebd
Clarify input generality and add contributing section to README
Remove VASP-specific framing from README, documentation, and paper.
The package operates on lists of pymatgen Structure objects and has
no VASP-specific code, so the description now reflects this generality.
Add a Contributing section to README linking to CONTRIBUTING.md.

Addresses reviewer comments from JOSS review.
