+++ paper/paper.md
@@ -45,7 +45,7 @@ The package provides four site-definition methods—spherical, polyhedral, Voron
 
 To assist with analyses using dynamic Voronoi and polyhedral sites, the package provides a reference-based workflow that automates site generation: users provide an ideal reference structure and specify coordination criteria (e.g., "Li ions coordinated by 4 O atoms"), and the software identifies all matching environments and maps them to target structures. This automated approach allows analyses using these dynamic sites without requiring the user to manually identify all relevant coordination environments.
 
-The package uses a builder pattern interface for configuration and parameter validation. Analyses can be set up and executed with a few lines of code, with results returned as Python lists that integrate simply into downstream analysis workflows. The software natively supports VASP XDATCAR files and integrates with the pymatgen ecosystem.
+The package uses a builder pattern interface for configuration and parameter validation. Analyses can be set up and executed with a few lines of code, with results returned as Python lists that integrate simply into downstream analysis workflows. The software operates on lists of pymatgen `Structure` objects, making it compatible with any trajectory source that can produce pymatgen structures.
 
 The materials modelling community has developed various tools for site-based trajectory analysis, including `pymatgen-analysis-diffusion` [@pymatgen-analysis-diffusion; @DengEtAl_ChemMater2016], `SITATOR` [@sitator; @KahleEtAl_PhysRevMater2019], `IonDiff` [@LopezEtAl_JOpenSourceSoftw2024], and `gemdat` [@GEMDAT]. These packages each implement specific schemes for defining sites and assigning occupations, and are often tightly integrated with downstream workflows for particular analysis tasks. `site-analysis` complements these tools by focusing on generality: it provides multiple site-definition methods within a consistent interface and produces output using simple data structures that can feed into any downstream analysis. To our knowledge, `site-analysis` is unique in implementing geometric sites that dynamically update during the simulation (polyhedral sites and dynamic Voronoi sites) in a publicly available package. We note that the use of dynamic polyhedral sites is particularly valuable for materials with close-packed host-framework structures, where mobile ion coordination environments provide a natural and intuitive basis for describing transport mechanisms [@Burbano2016; @Morgan2021; @Mercadier2023].
 
