+++ docs/source/index.md
@@ -2,7 +2,7 @@
 
 `site-analysis` is a Python module for analysing molecular dynamics simulations of solid-state ion transport, by assigning positions of mobile ions to specific "sites" within the host structure.
 
-The code is built on top of [pymatgen](https://pymatgen.org) and processes molecular dynamics trajectory data, with particular support for VASP XDATCAR format files.
+The code is built on top of [pymatgen](https://pymatgen.org) and operates on molecular dynamics trajectories represented as lists of pymatgen `Structure` objects. Any trajectory source that can produce pymatgen structures can be used as input.
 
 [![Build Status](https://github.com/bjmorgan/site-analysis/actions/workflows/build.yml/badge.svg)](https://github.com/bjmorgan/site-analysis/actions/workflows/build.yml)
 [![Coverage Status](https://coveralls.io/repos/github/bjmorgan/site-analysis/badge.svg?branch=main)](https://coveralls.io/github/bjmorgan/site-analysis?branch=main)
