### 2026-02-26 a7ae9ad
Simplify CI: run on PRs only and remove Coveralls
Remove coverage reporting and Coveralls integration from the build
workflow. Restrict CI trigger to pull requests only.
