### 2026-02-26 #9606 @bjmorgan
@Xia-Yijie Thank you for the review.

Points 1 & 2: I have addressed these in https://github.com/bjmorgan/site-analysis/pull/35. The README now includes a Contributing section linking to CONTRIBUTING.md, and the documentation and paper have been updated to clarify that the package operates on lists of pymatgen Structure objects and is not specific to any particular MD engine.

Point 3: Could you expand on what edge cases you have in mind? For spherical sites, a site contains an atom if the distance to the site centre is less than the sphere radius. How assignment is handled when sites overlap, or when atoms fall outside all sites, is already discussed in the documentation. For polyhedral and Voronoi sites, the containment criteria are deterministic with no tuneable parameters. A point is either inside the volume or it is not. If there are specific scenarios you're thinking of where the documentation could be clearer, I'd be happy to address those.
