### 2026-02-27 95bc002
Optimise polyhedral containment with numba surface normal method
Add a cached surface normal containment algorithm with optional numba
JIT compilation. Face topology is computed once from a ConvexHull and
cached; per-timestep only face normals are recomputed from the cached
topology. The cache is built lazily on first contains_point() call,
avoiding unnecessary ConvexHull construction for sites that are never
queried.

When numba is not available, falls back to the existing Delaunay
tessellation. The algo parameter on contains_point() and
contains_atom() is deprecated.

Points exactly on a pol
[…]
