### 2026-02-27 bd2c96d
Lazy vertex coordinate assignment for polyhedral sites
Defer PBC-corrected coordinate extraction until contains_point is
actually called, rather than eagerly computing for all sites each frame.
The collection pre-extracts structure.frac_coords once and passes the
shared array to each site via notify_structure_changed; each site then
indexes into this array on demand. This avoids both redundant
PeriodicSite object creation (192 unique vertices were previously
accessed 5760 times) and PBC correction for the ~1200 sites never
queried per frame.

Benchmark on argyrodite Li6PS5Cl (1440 sites, 192 a
[…]
