+++ site_analysis/polyhedral_site_collection.py
@@ -72,8 +72,10 @@ class PolyhedralSiteCollection(SiteCollection):
             structure: Structure):
         for a in atoms:
             a.assign_coords(structure)
+        all_frac_coords = structure.frac_coords
+        lattice = structure.lattice
         for s in self.sites:
-            s.assign_vertex_coords(structure)
+            s.notify_structure_changed(all_frac_coords, lattice)
         self.assign_site_occupations(atoms, structure)
                     
     def assign_site_occupations(self, atoms, structure) -> None:
