+++ tests/benchmark_containment.py
@@ -1,0 +1,167 @@
+"""Benchmark comparing containment strategies on real MD data.
+
+Isolates the site assignment cost (assign_vertex_coords + contains_point)
+by running the argyrodite trajectory multiple times through the same
+trajectory object, emulating a longer trajectory.
+
+Usage:
+    python tests/benchmark_containment.py [--repeats N]
+"""
+
+import argparse
+import sys
+import time
+from pathlib import Path
+from unittest.mock import patch
+
+import numpy as np
+from pymatgen.core import Lattice, Structure
+from pymatgen.io.vasp import Xdatcar
+
+from site_analysis import TrajectoryBuilder
+from site_analysis.containment import HAS_NUMBA
+from site_analysis.polyhedral_site_collection import PolyhedralSiteCollection
+from site_analysis.site import Site
+
+DATA_DIR = Path(__file__).resolve().parent.parent / "docs" / "source"
+ARGYRODITE_XDATCAR = DATA_DIR / "tutorials" / "data" / "Li6PS5Cl_0p_XDATCAR.gz"
+
+
+def build_argyrodite_trajectory(structures):
+    """Build a trajectory object for the argyrodite analysis."""
+    Site._newid = 0
+
+    lattice = Lattice.cubic(a=10.155)
+    coords = np.array([
+        [0.5, 0.5, 0.5],
+        [0.9, 0.9, 0.6],
+        [0.23, 0.92, 0.09],
+        [0.25, 0.25, 0.25],
+        [0.15, 0.15, 0.15],
+        [0.0, 0.183, 0.183],
+        [0.0, 0.0, 0.0],
+        [0.75, 0.25, 0.25],
+        [0.11824, 0.11824, 0.38176],
+    ])
+    reference_structure = Structure.from_spacegroup(
+        sg="F-43m",
+        lattice=lattice,
+        species=["P", "Li", "Mg", "Na", "Be", "K", "S", "S", "S"],
+        coords=coords,
+    ) * [2, 2, 2]
+
+    builder = (
+        TrajectoryBuilder()
+        .with_structure(structures[0])
+        .with_reference_structure(reference_structure)
+        .with_mobile_species("Li")
+        .with_structure_alignment(align_species="P")
+        .with_site_mapping(mapping_species=["S", "Cl"])
+    )
+
+    for centre_species, label in [
+        ("Li", "type 1"), ("Mg", "type 2"), ("Na", "type 3"),
+        ("Be", "type 4"), ("K", "type 5"),
+    ]:
+        builder.with_polyhedral_sites(
+            centre_species=centre_species,
+            vertex_species="S",
+            cutoff=3.0,
+            n_vertices=4,
+            label=label,
+        )
+
+    return builder.build()
+
+
+def _eager_analyse_structure(self, atoms, structure):
+    """Original eager analysis: assign all vertex coords upfront."""
+    for a in atoms:
+        a.assign_coords(structure)
+    for s in self.sites:
+        s.assign_vertex_coords(structure)
+    self.assign_site_occupations(atoms, structure)
+
+
+def benchmark_assignment(trajectory, structures, repeats):
+    """Benchmark just the per-frame analysis loop."""
+    # Warm up: run one frame to trigger lazy cache creation
+    trajectory.append_timestep(structures[0], t=0)
+
+    # Now time repeated passes over the trajectory
+    all_structures = structures * repeats
+    start = time.perf_counter()
+    for i, s in enumerate(all_structures):
+        trajectory.analyse_structure(s)
[…]
