### 2026-02-27 15ed9e4
Cache PBC image shifts to avoid redundant unwrapping
The per-vertex integer image shifts from
unwrap_vertices_to_reference_center are stable across most consecutive
frames. Cache these shifts on first computation and reapply on
subsequent frames. If any vertex coordinate changes by > 0.5 fractional
(indicating a periodic boundary crossing), the full 27-image distance
search is recomputed and the cache updated.

The uniform non-negative shift is recomputed each frame (trivial cost).

Also clean up over-commented code in pbc_utils.

Benchmark on argyrodite Li6PS5Cl (1440 sites, 192 atoms, 700 fr
[…]
