### 2026-02-27 1a244e2
JIT-compile face normal update and fix PBC shift cache wrapping
Move FaceTopologyCache.update computation into a numba-compiled
function, eliminating numpy dispatch overhead for small arrays (4 faces).
This was the dominant bottleneck at 14.3s cumulative.

Also fix PBC shift cache invalidation: separate coordinate wrapping
(integer jumps from VASP wrapping to [0,1)) from physical displacement.
Cached shifts are adjusted for wrapping rather than triggering a full
recomputation, reducing unwrap calls from 61k to 0 after warm-up.

Benchmark on argyrodite Li6PS5Cl (1440 sites, 192 atoms, 700 frame
[…]
