### 2026-02-27 b7803b6
JIT-compile PBC shift update and add missing unit tests
Move the cached PBC shift update to a numba JIT function with
pure-numpy fallback, dispatched once at import time via
update_pbc_shifts.  This eliminates numpy dispatch overhead on
small (4x3) arrays in the hot path.

Add unit tests for: update_pbc_shifts (cache hit, wrapping,
miss), notify_structure_changed / _assign_from_pending (lazy
assignment), _store_vertex_coords (PBC shift caching), and
unwrap_vertices_to_reference_center return_image_shifts parameter.

Update stale docstrings in containment.py and polyhedral_site.py.
