+++ tests/test_containment.py
@@ -8,6 +8,8 @@ from numpy.testing import assert_array_equal
 from site_analysis.containment import (
     FaceTopologyCache,
     HAS_NUMBA,
+    _numpy_update_pbc_shifts,
+    update_pbc_shifts,
 )
 
 # Tetrahedron centred at [0.5, 0.5, 0.5]
@@ -154,6 +156,100 @@ class TestNumbaQuery(unittest.TestCase):
             cache._centre_signs,
         )
         self.assertFalse(result)
+
+
+class TestNumpyUpdatePbcShifts(unittest.TestCase):
+    """Tests for _numpy_update_pbc_shifts."""
+
+    def test_cache_hit_small_displacement(self):
+        """Small physical displacement returns valid cache."""
+        cached = np.array([[0.1, 0.2, 0.3],
+                           [0.4, 0.5, 0.6]])
+        new = cached + 0.01  # small vibration
+        shifts = np.array([[0, 0, 1],
+                           [0, 1, 0]], dtype=np.int64)
+        valid, coords, new_shifts = _numpy_update_pbc_shifts(new, cached, shifts)
+        self.assertTrue(valid)
+        assert_array_equal(new_shifts, shifts)
+        expected = new + shifts
+        min_coords = np.min(expected, axis=0)
+        uniform = np.maximum(0, np.ceil(-min_coords))
+        np.testing.assert_array_almost_equal(coords, expected + uniform)
+
+    def test_cache_hit_with_wrapping(self):
+        """Vertex wrapping from 0.99 to 0.01 is detected and shifts adjusted."""
+        cached = np.array([[0.99, 0.5, 0.5],
+                           [0.5, 0.5, 0.5]])
+        new = np.array([[0.01, 0.5, 0.5],   # wrapped across boundary
+                        [0.5, 0.5, 0.5]])
+        shifts = np.array([[0, 0, 0],
+                           [0, 0, 0]], dtype=np.int64)
+        valid, coords, new_shifts = _numpy_update_pbc_shifts(new, cached, shifts)
+        self.assertTrue(valid)
+        # Wrapping of -0.98 rounds to -1, so shift gains +1
+        assert_array_equal(new_shifts[0], [1, 0, 0])
+        assert_array_equal(new_shifts[1], [0, 0, 0])
+
+    def test_cache_miss_large_displacement(self):
+        """Large physical displacement invalidates cache."""
+        cached = np.array([[0.1, 0.2, 0.3],
+                           [0.4, 0.5, 0.6]])
+        new = cached + 0.4  # too large
+        shifts = np.zeros((2, 3), dtype=np.int64)
+        valid, _, _ = _numpy_update_pbc_shifts(new, cached, shifts)
+        self.assertFalse(valid)
+
+    def test_non_negative_shift_applied(self):
+        """Uniform shift ensures all output coordinates are non-negative."""
+        cached = np.array([[0.05, 0.05, 0.05],
+                           [0.1, 0.1, 0.1]])
+        new = cached + 0.001
+        shifts = np.array([[0, 0, -1],
+                           [0, 0, -1]], dtype=np.int64)
+        valid, coords, _ = _numpy_update_pbc_shifts(new, cached, shifts)
+        self.assertTrue(valid)
+        self.assertTrue(np.all(coords >= 0))
+
+
+@pytest.mark.skipif(not HAS_NUMBA, reason="numba not installed")
+class TestNumbaUpdatePbcShifts(unittest.TestCase):
+    """Tests that numba and numpy PBC shift implementations agree."""
+
+    def test_agrees_with_numpy_small_displacement(self):
[…]
