### 2026-02-27 8477a4c
Compute PBC images once per atom instead of once per site check
Move x_pbc call from inside contains_point (called per atom-site pair)
to the per-atom loop in assign_site_occupations, passing the result
through via a new pbc_images keyword argument.
