### 2026-02-27 d944670
Update docs for automatic containment algorithm selection
- Replace deprecated algo parameter examples with numba installation guidance
- Fix use_reference_centers typo in PBC handling docs
