+++ docs/source/concepts/pbc_handling.md
@@ -38,7 +38,7 @@ The reference centre for each site is the fractional coordinates of the central 
 
 ### Performance Considerations
 
-For some systems the spread-based method can be up to 20% faster than the reference centre method, which may be advantageous when analysing large simulation trajectories. The specific performance characteristics depend on your system&mdash;the number of sites, system size, and coordination environments all influence computational cost. The reference-center method is the default approach when using the {class}`~site_analysis.builders.TrajectoryBuilder` or {class}`~site_analysis.reference_workflow.ReferenceBasedSites` workflows, due to its correctness in small simulation cells. If you are certain that all sites have spreads well below &frac12; the simulation cell dimensions, even under dynamic distortions, you may wish to try the spread-based method (e.g., by setting `use_reference_centers=True`) to compare the relative performance of the two methods.
+For some systems the spread-based method can be up to 20% faster than the reference centre method, which may be advantageous when analysing large simulation trajectories. The specific performance characteristics depend on your system&mdash;the number of sites, system size, and coordination environments all influence computational cost. The reference-center method is the default approach when using the {class}`~site_analysis.builders.TrajectoryBuilder` or {class}`~site_analysis.reference_workflow.ReferenceBasedSites` workflows, due to its correctness in small simulation cells. If you are certain that all sites have spreads well below &frac12; the simulation cell dimensions, even under dynamic distortions, you may wish to try the spread-based method (e.g., by setting `use_reference_centers=False`) to compare the relative performance of the two methods.
 
 ## Controlling PBC Handling in Your Code
 
