### 2026-02-27 b603adf
Fix mypy errors from CI
Add overloads to unwrap_vertices_to_reference_center for return type
narrowing, and assert pending state in _assign_from_pending.
