### 2026-02-27 a179acb
Address PR review comments
- Fix unwrap_vertices_to_reference_center returning wrong type for
  empty input when return_image_shifts=True
- Add numba guard to FaceTopologyCache.__init__ for clearer error
- Use `is not None` instead of truthiness check for Structure argument
