### 2026-02-27 a882133
Document [fast] install option in README and installation guide
