### 2026-02-27 4558c41
Update PBC performance docs to reflect cached shift incremental updates
