### 2026-02-27 1d3d303
Bump version to 1.3.0
