### 2026-02-27 c70cbba
Address additional review feedback
- Clear pending state and PBC shift caches in reset()
- Wrap QhullError with site context for easier debugging
- Improve _store_vertex_coords docstring
- Comment the dot != 0.0 boundary correctness guard
- Annotate np.asarray as mypy workaround
- Add tests for reset() clearing and _cache_stale transitions
