### 2026-02-27 5b077c5
Tighten type narrowing for paired cache fields
- Guard _pbc_cached_raw_frac alongside _pbc_image_shifts in
  _store_vertex_coords
- Remove unnecessary None annotations from FaceTopologyCache
  internal attributes (always set by update() in __init__)
