### 2026-02-27 da8721d
Optimise polyhedral containment with optional numba acceleration (#37)
Add optional numba acceleration for polyhedral site containment (~7x
speedup) with automatic fallback to Delaunay tessellation.

Optimisations: cached face topology with JIT-compiled surface normal
queries, lazy vertex coordinate assignment, PBC image shift caching
with JIT-compiled incremental updates, and per-atom (rather than
per-site) PBC image computation.

New `[fast]` install extra: `pip install site-analysis[fast]`

The `algo` parameter on `contains_point`/`contains_atom` is deprecated;
the best available algorithm 
[…]
