### 2026-02-28 #9606 @Xia-Yijie
@bjmorgan Thanks for the detailed response and for updating the PR.

For Point 3, my initial question about edge cases was driven by my general MD background. I wasn't entirely familiar with the specific geometric site definitions used here and didn't realize upfront how strictly deterministic they are. Your explanation clears this up perfectly.

The revisions to the documentation and the paper are very helpful. I am completely satisfied with these updates, and have no more questions.
