### 2026-02-28 3bf3f53
Add changelog, backfilled from release history
Include changelog in documentation via MyST include directive.
