+++ CHANGELOG.md
@@ -1,0 +1,169 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+
+## [1.3.0] - 2026-02-27
+
+### Added
+
+- Optional Numba acceleration for polyhedral site containment testing (~7x speedup). Install with `pip install site-analysis[fast]`.
+
+### Changed
+
+- Cached face topology with JIT-compiled surface normal queries, replacing per-timestep Delaunay construction.
+- Lazy vertex coordinate assignment defers PBC correction until a site is actually queried.
+- PBC image shift caching avoids the expensive 27-image distance search.
+- Per-atom PBC image computation (`x_pbc` computed once per atom per frame rather than once per atom-site pair).
+- `algo` parameter on `contains_point`/`contains_atom` is deprecated; algorithm is now selected automatically.
+
+### Fixed
+
+- Fixed `use_reference_centers` typo in PBC handling documentation.
+
+## [1.2.7] - 2026-02-21
+
+### Changed
+
+- Replaced deprecated `typing` imports (`Optional`, `Union`, `List`, etc.) with built-in equivalents.
+- Replaced defensive `assert` statements with proper exceptions (`ValueError`, `TypeError`, `RuntimeError`).
+- Standardised indentation to 4 spaces across all source files.
+- Removed redundant `object` inheritance from `Atom` and `Trajectory`.
+- Replaced type-narrowing hacks with annotations.
+- Consolidated lazy initialisation helpers in `ReferenceBasedSites`.
+- Use public `assign_coords()` API instead of accessing private `Atom._frac_coords`.
+
+### Fixed
+
+- Fixed `StructureAligner._run_differential_evolution` mutating the caller's `minimizer_options` dict.
+- Fixed incorrect docstrings (typos, wrong default values, mismatched parameter names).
+
+## [1.2.6] - 2026-02-21
+
+### Fixed
+
+- Fixed float distance truncation in `get_coordination_indices` (neighbour distances were cast to `int`).
+- Fixed `append_timestep` dropping timestep `t=0` (truthiness check treated 0 as falsy).
+- Fixed transitions from site index 0 not being recorded (truthiness check treated 0 as falsy).
+- Fixed `contains_point_sn` loop structure (`inside.append()` was inside the inner face loop).
+- Fixed `Atom.from_dict` crashing when `frac_coords` is absent.
+- Removed debug `print` statement from `PolyhedralSite.__init__`.
+- Removed dead code and variable shadowing in `Trajectory.__init__`.
+- Fixed `x_pbc` docstring (stated `(9, 3)` return shape; actual shape is `(8, 3)`).
+
+## [1.2.5] - 2026-01-17
+
+### Fixed
+
+- Fixed quickstart example to work out-of-the-box (#31).
+- Added `example_data/XDATCAR` with a minimal Li trajectory for the quickstart example.
+- Added helpful error message when mobile species is not found in structure.
+
+## [1.2.4] - 2025-09-01
+
+### Added
+
+- `Trajectory.site_summaries()` method returning structured summary statistics for all sites.
+- `Trajectory.write_site_summaries(filename)` method for exporting summaries to JSON.
+- Optional `metrics` parameter for selecting specific statistics (occupancy, transitions, etc.).
+- Numpy array inputs for `PolyhedralSite`.
+- Description label for `tqdm` progress output.
+
+## [1.2.1] - 2025-06-07
+
+### Changed
+
[…]
