### 2026-02-28 c2aa1ac
Optimise dynamic Voronoi site centre calculation
Extract structure.frac_coords once per timestep in the collection and
pass to all sites via calculate_centre_from_bulk, avoiding per-site
PeriodicSite object creation. Cache integer PBC image shifts between
timesteps and update incrementally by detecting coordinate wraps,
bypassing the expensive 27-image unwrap on subsequent frames.

Benchmarked on argyrodite (1440 sites, 192 atoms, 420 frames):
~2x speedup on the full trajectory analysis workflow.
