### 2026-02-28 15ef2f0
Batch centre calculation across sites in collection
Group sites by reference count and compute all centres per group in a
single vectorised operation, replacing 604,800 individual per-site
np.mean/PBC calls with ~2,100 batch operations (5 groups x 420 frames).

Centre calculation drops from 8.1s to 0.35s. Combined with the previous
per-site PBC shift caching, the full workflow speedup is ~4.3x
(45ms/frame to 10.5ms/frame on argyrodite with 1440 sites).
