+++ tests/test_dynamic_voronoi_site_collection.py
@@ -17,8 +17,10 @@ class DynamicVoronoiSiteCollectionTestCase(unittest.TestCase):
 	
 	def test_site_collection_is_initialised(self):
 		"""Test that a DynamicVoronoiSiteCollection is correctly initialised."""
-		sites = [Mock(spec=DynamicVoronoiSite, index=0),
-				 Mock(spec=DynamicVoronoiSite, index=1)]
+		sites = [Mock(spec=DynamicVoronoiSite, index=0,
+					  reference_indices=[0, 1], reference_center=None),
+				 Mock(spec=DynamicVoronoiSite, index=1,
+					  reference_indices=[2, 3], reference_center=None)]
 		site_collection = DynamicVoronoiSiteCollection(sites=sites)
 		self.assertEqual(site_collection.sites, sites)
 		
@@ -32,47 +34,37 @@ class DynamicVoronoiSiteCollectionTestCase(unittest.TestCase):
 			DynamicVoronoiSiteCollection(sites=sites)
 			
 	def test_analyse_structure(self):
-		"""Test that analyse_structure extracts coords once and calls calculate_centre_from_bulk."""
-		# Create mock sites and atoms with proper index attributes
-		sites = [
-			Mock(spec=DynamicVoronoiSite, reference_indices=[0, 1], index=0),
-			Mock(spec=DynamicVoronoiSite, reference_indices=[2, 3], index=1)
-		]
-		atoms = [Mock(spec=Atom) for _ in range(5)]
-		mock_frac_coords = np.array([[0.1, 0.2, 0.3]] * 5)
-		mock_lattice = Mock(spec=Lattice)
-		structure = Mock(spec=Structure)
-		type(structure).frac_coords = PropertyMock(return_value=mock_frac_coords)
-		structure.lattice = mock_lattice
+		"""Test that analyse_structure computes centres and assigns occupations."""
+		site1 = DynamicVoronoiSite(reference_indices=[0, 1])
+		site2 = DynamicVoronoiSite(reference_indices=[2, 3])
+		collection = DynamicVoronoiSiteCollection(sites=[site1, site2])
 
-		# Initialize the collection
-		site_collection = DynamicVoronoiSiteCollection(sites=sites)
+		lattice = Lattice.cubic(10.0)
+		# Atoms 0,1 are reference for site1; 2,3 for site2; 4 is mobile
+		coords = [[0.1, 0.1, 0.1], [0.2, 0.2, 0.2],
+				  [0.7, 0.7, 0.7], [0.8, 0.8, 0.8],
+				  [0.15, 0.15, 0.15]]
+		structure = Structure(lattice, ["Na"] * 5, coords)
 
-		# Mock the assign_site_occupations method to avoid actually calling it
-		site_collection.assign_site_occupations = Mock()
+		atom = Atom(index=4)
+		collection.analyse_structure([atom], structure)
 
-		# Call analyse_structure
-		site_collection.analyse_structure(atoms, structure)
+		# Centres should be computed from reference atoms
+		np.testing.assert_array_almost_equal(
+			site1.centre, np.array([0.15, 0.15, 0.15]))
+		np.testing.assert_array_almost_equal(
+			site2.centre, np.array([0.75, 0.75, 0.75]))
 
-		# Check that each atom's coordinates are assigned
-		for atom in atoms:
-			atom.assign_coords.assert_called_with(structure)
-
-		# Check that each site's centre is calculated via bulk extraction
-		for site in sites:
-			site.calculate_centre_from_bulk.assert_called_once()
-			args = site.calculate_centre_from_bulk.call_args[0]
[…]
