### 2026-02-28 6424219
Remove unused reference_centres field and fix reset cache consistency
Remove _CentreGroup.reference_centres — the first-frame fallback reads
site.reference_center directly, so the group-level copy was dead code
with a latent mixed-dtype bug.

Fix reset_centre_groups to also clear per-site PBC caches, preventing
stale site._pbc_cached_raw_frac from being used when the per-site
fallback path runs after a reset.
