### 2026-02-28 2009b18
Add test for correct centres after reset and reuse
