### 2026-02-28 b18b92d
Harden batch cache against asymmetric resets
Guard _batch_calculate_centres against per-site caches being cleared
independently of group state (e.g. by Trajectory.reset()). Clear
per-site caches before the fallback path to prevent stale data from
an earlier batch frame being reused by _compute_corrected_coords.

Add tests for batch cache invalidation fallback (large displacement),
PBC wrapping across periodic boundaries, and asymmetric reset via
site.reset() without group reset.
