+++ site_analysis/dynamic_voronoi_site_collection.py
@@ -125,7 +125,7 @@ class DynamicVoronoiSiteCollection(SiteCollection):
         for group in self._centre_groups:
             # (n_sites, n_ref, 3)
             batch_ref = all_frac_coords[group.ref_indices]
-            if group.initialised:
+            if group.initialised and self.sites[group.site_positions[0]]._pbc_image_shifts is not None:
                 diff = batch_ref - group.cached_raw_frac
                 wrapping = np.round(diff).astype(np.int64)
                 physical_diff = diff - wrapping
@@ -142,7 +142,12 @@ class DynamicVoronoiSiteCollection(SiteCollection):
                     for idx, pos in enumerate(group.site_positions):
                         self.sites[pos]._centre_coords = centres[idx]
                     continue
-            # First frame or cache invalidation — per-site full computation
+            # First frame or cache invalidation — per-site full computation.
+            # Clear per-site caches first to avoid stale data from an
+            # earlier batch frame being used by _compute_corrected_coords.
+            for pos in group.site_positions:
+                self.sites[pos]._pbc_image_shifts = None
+                self.sites[pos]._pbc_cached_raw_frac = None
             for idx, pos in enumerate(group.site_positions):
                 site = self.sites[pos]
                 ref_coords = batch_ref[idx]
