### 2026-02-28 864672c
Add tests for reference_center caching and mixed-group batching
Add TestComputeCorrectedCoordsWithReferenceCentre: tests that the
production reference_center path (unwrap_vertices_to_reference_center)
uses caching correctly across multiple frames.

Add test_mixed_n_reference_groups_produce_correct_centres: verifies
batch calculation with multiple groups of different n_reference values
produces identical results to per-site computation.

Fix benchmark docstring to match actual capabilities.
