### 2026-02-28 48e9af1
Remove benchmark script from tests
