### 2026-02-28 b8519ff
Remove per-site PBC caching from DynamicVoronoiSite
The batch path in the collection is the only production path for
centre calculation. Per-site PBC shift caching added complexity
(dual-layer cache synchronisation, asymmetric reset guards, stale
cache clearing) for no production benefit. Removing it eliminates
the synchronisation issues and simplifies the code.

- Strip _pbc_image_shifts and _pbc_cached_raw_frac from the site
- Remove dead calculate_centre_from_bulk method
- Simplify _batch_calculate_centres fallback to use return value
- Remove _CentreGroup.reset(); inline in reset_centre_gr
[…]
