+++ tests/test_dynamic_voronoi_site_collection.py
@@ -1,5 +1,5 @@
 import unittest
-from unittest.mock import Mock, patch, PropertyMock, call
+from unittest.mock import Mock, patch
 import numpy as np
 from pymatgen.core import Structure, Lattice
 
@@ -59,92 +59,30 @@ class DynamicVoronoiSiteCollectionTestCase(unittest.TestCase):
 		self.assertIn(atom.index, site1.contains_atoms)
 			
 	def test_assign_site_occupations(self):
-		"""Test that atoms are correctly assigned to sites based on Voronoi tessellation."""
-		# Create mock sites
-		site1 = Mock(spec=DynamicVoronoiSite,
-					 reference_indices=[0, 1], reference_center=None)
-		site2 = Mock(spec=DynamicVoronoiSite,
-					 reference_indices=[2, 3], reference_center=None)
-		site1.index = 0
-		site2.index = 1
-		# Set up centre method to return fixed coordinates
-		centre1 = PropertyMock(return_value=np.array([0.2, 0.2, 0.2]))
-		centre2 = PropertyMock(return_value=np.array([0.8, 0.8, 0.8]))
-		type(site1).centre = centre1
-		type(site2).centre = centre2
-		
-		sites = [site1, site2]
-		
-		# Create mock atoms
-		atoms = [Mock(spec=Atom) for _ in range(5)]
-		for i, atom in enumerate(atoms):
-			atom.index = i
-			# Set up frac_coords attribute
-			atom.frac_coords = np.array([0.1 + 0.2 * i,
-										 0.1 + 0.2 * i,
-										 0.1 + 0.2 * i])
-		
-		# Create mock structure and lattice
-		lattice = Mock(spec=Lattice)
-		structure = Mock(spec=Structure)
-		structure.lattice = lattice
-		
-		# Mock the distance calculation
-		# Return a matrix where:
-		# - atoms 0-1 are closer to site1
-		# - atoms 2-4 are closer to site2
-		mock_distances = np.array([
-			[2.0, 3.0, 8.0, 10.0, 7.0],  # Distances from site1 to atoms 0-4
-			[8.0, 7.0, 2.0, 1.0, 3.0]    # Distances from site2 to atoms 0-4
-		])
-		lattice.get_all_distances = Mock(return_value=mock_distances)
-		
-		# Create collection with mocked sites
-		site_collection = DynamicVoronoiSiteCollection(sites=sites)
-		
-		# Mock the reset_site_occupations method
-		site_collection.reset_site_occupations = Mock()
-		
-		# Mock the update_occupation method to track assignments
-		site_collection.update_occupation = Mock()
-		
-		# Call the method being tested
-		site_collection.assign_site_occupations(atoms, structure)
-		
-		# Verify reset_site_occupations was called
-		site_collection.reset_site_occupations.assert_called_once()
-		
-		# Verify centre property was accessed for each site
-		centre1.assert_called()
-		centre2.assert_called()
-		
-		# Verify get_all_distances was called with the correct parameters
-		# We can't use site1.centre here because it's a PropertyMock, so get the value directly
-		expected_site_coords = np.array([
-			centre1.return_value,
-			centre2.return_value
-		])
-		atom_coords = np.array([atom.frac_coords for atom in atoms])
-		
-		# We can't directly compare numpy arrays in the call args, so we need to extract them
-		args, kwargs = lattice.get_all_distances.call_args
-		np.testing.assert_array_equal(args[0], expected_site_coords)
-		np.testing.assert_array_equal(args[1], atom_coords)
-		
[…]
