### 2026-02-28 c780e9c
Move site reset responsibility into SiteCollection
Add reset() to base SiteCollection that resets all sites via
Site.reset(). DynamicVoronoiSiteCollection overrides to also clear
batch PBC group caches. Trajectory.reset() now delegates to
site_collection.reset() instead of iterating sites directly,
ensuring collection-level caches are properly cleared.

Also fixes _CentreGroup.site_positions docstring, renames
reset_centre_groups to reset, and fixes the asymmetric reset test
to use coordinates that actually trigger the fallback path.

Adds refactoring notes for future PBC correction unificatio
[…]
