### 2026-02-28 8e31e25
Add unit test for SiteCollection.reset()
