### 2026-02-28 71ff2fb
Move fast-path logic into _CentreGroup and fix review issues
Extract try_fast_update() and initialise() methods into _CentreGroup
so the dataclass owns its cache invariants and the vectorised
computation. _batch_calculate_centres becomes pure orchestration.

Also:
- Fix .astype(int) to .astype(np.int64) for cross-platform consistency
- Fix British English in docstrings (center -> centre)
- Document _compute_corrected_coords side-effect on _centre_coords
- Clarify misleading "uniform" comment
