### 2026-02-28 8a21ca2
Add integration tests for batch reset and reference centre path
Add test_reset_clears_dynamic_voronoi_batch_caches to trajectory
tests, verifying that trajectory.reset() clears group caches when
using DynamicVoronoiSiteCollection.

Add test_batch_with_reference_center_matches_per_site to collection
tests, confirming the batch path produces correct centres when sites
have reference_center set.
