### 2026-02-28 f66ec11
Update changelog for v1.4.0
