+++ docs/source/modules/dynamic_voronoi_site_collection.rst
@@ -1,0 +1,7 @@
+site\_analysis.dynamic\_voronoi\_site\_collection
+--------------------------------------------------
+
+.. automodule:: site_analysis.dynamic_voronoi_site_collection
+    :members:
+    :undoc-members:
+    :show-inheritance:
