### 2026-02-28 2a7d402
Optimise dynamic Voronoi centre calculation with batch vectorisation
- Batch vectorised centre calculation for dynamic Voronoi sites via `_CentreGroup`, with incremental PBC shift fast path and per-site fallback
- Move site reset responsibility into `SiteCollection` base class
- Add missing API reference pages for dynamic Voronoi modules
- Update changelog for v1.4.0
