+++ site_analysis/dynamic_voronoi_site_collection.py
@@ -6,30 +6,112 @@ atoms to these sites based on their positions in a crystal structure.
 
 The DynamicVoronoiSiteCollection extends the base SiteCollection class with
 specific functionality for dynamic Voronoi sites, including:
-1. Calculating the dynamic centers of sites based on reference atom positions
+1. Calculating the dynamic centres of sites based on reference atom positions
 2. Assigning atoms to sites using Voronoi tessellation principles
 
 For atom assignment, the collection:
-1. First updates each site's center by calculating the mean position of its
+1. First updates each site's centre by calculating the mean position of its
    reference atoms, with special handling for periodic boundary conditions
-2. Calculates distances from each (dynamically determined) site center to each atom
-3. Assigns each atom to the site with the nearest center
+2. Calculates distances from each (dynamically determined) site centre to each atom
+3. Assigns each atom to the site with the nearest centre
 4. Uses the structure's lattice to correctly handle distances across
    periodic boundaries
 
 This collection is particularly useful for tracking sites in frameworks
-that deform during simulation, as the site centers adapt to the changing
+that deform during simulation, as the site centres adapt to the changing
 positions of the reference atoms.
 """
 
 from __future__ import annotations
 
+from collections import defaultdict
+from dataclasses import dataclass, field
+
 import numpy as np
-from pymatgen.core import Structure
+from pymatgen.core import Lattice, Structure
 from site_analysis.site_collection import SiteCollection
 from site_analysis.site import Site
 from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.atom import Atom
+
+
+@dataclass
+class _CentreGroup:
+    """Batch arrays for a group of sites sharing the same n_reference.
+
+    Owns the cached PBC shift state and the vectorised fast-path
+    computation.  The collection orchestrates fallback (per-site)
+    computation and distributes computed centres back to individual
+    sites.
+
+    Attributes:
+        site_positions: Indices into the parent
+            ``DynamicVoronoiSiteCollection.sites`` list for this group.
+        ref_indices: ``(n_sites, n_ref)`` int array of reference atom
+            indices.
+        pbc_shifts: ``(n_sites, n_ref, 3)`` int, cached image shifts.
+        cached_raw_frac: ``(n_sites, n_ref, 3)`` float, previous raw
+            coords.
+        initialised: Whether the PBC caches have been populated.
+    """
+    site_positions: list[int]
+    ref_indices: np.ndarray
+    pbc_shifts: np.ndarray = field(init=False, repr=False)
+    cached_raw_frac: np.ndarray = field(init=False, repr=False)
+    initialised: bool = field(init=False, default=False)
+
+    def __post_init__(self) -> None:
+        n_sites, n_ref = self.ref_indices.shape
+        self.pbc_shifts = np.zeros((n_sites, n_ref, 3), dtype=np.int64)
+        self.cached_raw_frac = np.zeros((n_sites, n_ref, 3))
+
+    def try_fast_update(self, batch_ref: np.ndarray) -> np.ndarray | None:
+        """Try the vectorised incremental shift update.
+
+        If the group is initialised and all coordinate displacements
+        since the last frame are below 0.3 fractional units, updates
+        the cached shifts and returns the computed centres.  Otherwise
+        returns ``None`` to signal that the caller should fall back to
+        per-site full PBC computation.
+
+        Args:
+            batch_ref: Raw fractional coordinates for this group,
+                shape ``(n_sites, n_ref, 3)``.
+
+        Returns:
+            Site centres as ``(n_sites, 3)`` array, or ``None`` if the
+            fast path cannot be used.
+        """
+        if not self.initialised:
+            return None
[…]
