+++ docs/source/modules/dynamic_voronoi_site.rst
@@ -1,0 +1,7 @@
+site\_analysis.dynamic\_voronoi\_site
+--------------------------------------
+
+.. automodule:: site_analysis.dynamic_voronoi_site
+    :members:
+    :undoc-members:
+    :show-inheritance:

[…]
