### 2026-02-28 dc06926
Optimise polyhedral site priority heuristic with distance ranking
Track the two most recent distinct sites per atom (O(1) lookup instead
of trajectory walking) and use precomputed distance-ranked site ordering
as fallback when reference centres are available. Falls back to
face-sharing neighbours then arbitrary order without reference centres.

Reduces containment checks by ~68% and gives ~1.4x wall-clock speedup
on argyrodite Li6PS5Cl trajectories (4.70 -> 1.49 checks/atom).
