+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Changed
+
+- Optimised polyhedral site assignment priority heuristic: tracks the two most recent distinct sites (catches bounce-back), and uses precomputed distance-ranked site ordering as fallback when reference centres are available. Reduces containment checks by ~68% and gives ~1.4x wall-clock speedup on argyrodite Li6PS5Cl trajectories.
+
 ## [1.4.0] - 2026-02-28
 
 ### Changed

[…]
