### 2026-02-28 b4c2c5b
Address PR review: fix docstrings, type hints, and add missing tests
- Fix stale most_recent_site docstring (no longer reads from trajectory)
- Fix Atom.reset() docstring to mention _recent_sites
- Fix American spellings in class docstring (neighbouring, optimised)
- Fix Unicode em-dashes in test comments
- Add type hints to _get_priority_sites
- Fix docstrings that omitted transitions step for no-history case
- Document fractional-space distance limitation
- Add checked_indices.add(index) in distance-ranked branch for consistency
- Add test for update_occupation calling update_recent_site
- A
[…]
