### 2026-02-28 75290f2
Bundle reference centre data into _ReferenceData NamedTuple
Replace separate _reference_centres and _site_indices fields (both
optional, always set together) with a single _reference_data:
_ReferenceData | None. One None check now narrows both fields,
eliminating the assert-based type narrowing.

_nearest_site_index becomes a static method taking _ReferenceData
and always returns int.
