### 2026-02-28 6cd1712
Move nearest_site_index onto _ReferenceData
The method only operates on reference centre data, so it belongs on
the type that holds that data rather than as a static method on
PolyhedralSiteCollection.
