### 2026-02-28 75799e7
Rename _ReferenceData to _NearestSiteLookup
Name describes what the type does (spatial lookup) rather than what
it is built from (reference data).
