+++ site_analysis/site_collection.py
@@ -19,10 +19,37 @@ subclasses instead.
 """
 
 from abc import ABC, abstractmethod
+from typing import NamedTuple, Sequence
+
 import numpy as np
-from typing import Sequence
 from pymatgen.core import Structure # type: ignore
 from .site import Site
+
+
+class _NearestSiteLookup(NamedTuple):
+    """Precomputed lookup for finding the nearest site to a given position.
+
+    Uses minimum-image convention in fractional space, which is only
+    geometrically exact for orthogonal cells.
+    """
+    centres: np.ndarray
+    site_indices: list[int]
+
+    def nearest_site_index(self, frac_coords: np.ndarray) -> int:
+        """Return the site index nearest to the given fractional coordinates.
+
+        Uses minimum-image convention in fractional space.
+
+        Args:
+            frac_coords: Fractional coordinates to find the nearest site for.
+
+        Returns:
+            The site index of the nearest site.
+        """
+        diffs = self.centres - frac_coords
+        diffs -= np.round(diffs)
+        dists = np.linalg.norm(diffs, axis=1)
+        return self.site_indices[int(np.argmin(dists))]
 
 class SiteCollection(ABC):
     """Parent class for collections of sites.
