### 2026-02-28 623e4a0
Improve polyhedral site priority heuristic with distance ranking
Track the two most recent distinct sites per atom via Atom._recent_sites
(O(1) lookup replacing trajectory walking), catching both "stay put" and
"bounce back" patterns. Precompute distance-ranked site ordering from
reference centres, used as fallback when learned transitions miss.

Falls back to face-sharing neighbours then arbitrary order when reference
centres are unavailable.

Benchmarked on argyrodite Li6PS5Cl (1440 sites, 192 atoms, 420 frames):
~68% fewer containment checks, ~1.4x wall-clock speedup.
