+++ site_analysis/polyhedral_site_collection.py
@@ -9,27 +9,42 @@ specific functionality for polyhedral sites, including:
 - Maintaining a map of neighbouring polyhedral sites that share faces
 - Optimised atom assignment using priority-based site checking
 - Using observed transition patterns between sites for performance optimisation
+- Precomputed distance-ranked site ordering when reference centres are available
 
-The module also includes a utility function, construct_neighbouring_sites,
-which analyses a set of polyhedral sites to determine which ones are
-face-sharing neighbours (defined as sites sharing three or more vertices).
+The module also includes utility functions:
+- construct_neighbouring_sites: analyses polyhedral sites to determine which
+  ones are face-sharing neighbours (sharing three or more vertices).
+- _compute_distance_ranked_sites: precomputes distance-ranked site lists
+  from reference centres for efficient fallback ordering.
 
-For atom assignment, the collection implements an intelligent optimisation
+For atom assignment, the collection implements a priority-based optimisation
 that reduces average search complexity from O(N) to O(k):
 
-1. First check if an atom is still in its most recently assigned site
-2. Then check observed transition destinations from that site in decreasing 
-    frequency order
-3. Then check neighbouring sites that share faces with the most recent site 
-    (if these have not yet been checked)
-4. Finally check all remaining sites if not found in the priority categories
+When trajectory history exists:
+    1. Check the most recently visited site, then the previously visited
+       site (covers staying put and bouncing back)
+    2. Check learned transition destinations in frequency order
+    3. Check remaining sites by distance from anchor site (if reference
+       centres are available), otherwise face-sharing neighbours then
+       arbitrary order
 
-This approach leverages both spatial relationships (face-sharing neighbors)
-and learned behavior (observed transition patterns) to dramatically reduce
-the number of containment checks required.
+When no trajectory history exists:
+    - If reference centres are available: check the nearest site centre
+      first, then learned transitions, then distance-ranked outward
+    - Otherwise: check all sites in arbitrary order
+
+Note: distance ranking uses minimum-image convention in fractional space,
+which is only geometrically exact for orthogonal cells. For non-orthogonal
+cells the ranking is approximate, but correctness is unaffected since all
+sites are eventually checked.
+
+This approach leverages spatial relationships, learned transition patterns,
+and distance-based ordering to reduce the number of containment checks.
 """
 
-from .site_collection import SiteCollection
+from collections.abc import Generator
+
+from .site_collection import SiteCollection, _NearestSiteLookup
 from .polyhedral_site import PolyhedralSite
 from .atom import Atom
 from .site import Site
@@ -37,12 +52,13 @@ from .tools import x_pbc
 from pymatgen.core import Structure # type: ignore
 import numpy as np
 
+
 class PolyhedralSiteCollection(SiteCollection):
     """A collection of PolyhedralSite objects.
     
-    Extends the base SiteCollection class with specific functionality for 
-    polyhedral sites, including maintaining a map of neighboring polyhedral 
-    sites that share faces and implementing optimized atom assignment based 
+    Extends the base SiteCollection class with specific functionality for
+    polyhedral sites, including maintaining a map of neighbouring polyhedral
+    sites that share faces and implementing optimised atom assignment based
     on spatial relationships and learned transition patterns.
     
     Attributes:
@@ -67,6 +83,9 @@ class PolyhedralSiteCollection(SiteCollection):
         super(PolyhedralSiteCollection, self).__init__(sites)
         self.sites: list[PolyhedralSite]
         self._neighbouring_sites = construct_neighbouring_sites(self.sites)
+        self._distance_ranked_sites, self._nearest_site_lookup = (
+            _compute_distance_ranked_sites(self.sites)
+        )
 
     def analyse_structure(self,
             atoms: list[Atom],
[…]
