### 2026-02-28 bff073d
Extract PriorityAssignmentMixin and improve spherical site priority heuristic
Move priority-based site assignment logic into a shared
PriorityAssignmentMixin in site_collection.py, used by both
PolyhedralSiteCollection and SphericalSiteCollection.

The spherical collection now benefits from:
- Tracking two recent distinct sites (was one)
- Precomputed distance-ranked site ordering
- Nearest-site lookup for atoms with no history

Remove the now-redundant neighbouring_sites infrastructure from
SphericalSiteCollection (neighbour_cutoff parameter, lazy neighbour
calculation, lattice caching) since
[…]
