+++ site_analysis/polyhedral_site_collection.py
@@ -7,44 +7,18 @@ atoms to these sites based on their positions in a crystal structure.
 The PolyhedralSiteCollection extends the base SiteCollection class with
 specific functionality for polyhedral sites, including:
 - Maintaining a map of neighbouring polyhedral sites that share faces
-- Optimised atom assignment using priority-based site checking
-- Using observed transition patterns between sites for performance optimisation
-- Precomputed distance-ranked site ordering when reference centres are available
+- Optimised atom assignment via the PriorityAssignmentMixin
+- Precomputed distance-ranked site ordering when reference centres are
+  available
 
 The module also includes utility functions:
 - construct_neighbouring_sites: analyses polyhedral sites to determine which
   ones are face-sharing neighbours (sharing three or more vertices).
-- _compute_distance_ranked_sites: precomputes distance-ranked site lists
-  from reference centres for efficient fallback ordering.
-
-For atom assignment, the collection implements a priority-based optimisation
-that reduces average search complexity from O(N) to O(k):
-
-When trajectory history exists:
-    1. Check the most recently visited site, then the previously visited
-       site (covers staying put and bouncing back)
-    2. Check learned transition destinations in frequency order
-    3. Check remaining sites by distance from anchor site (if reference
-       centres are available), otherwise face-sharing neighbours then
-       arbitrary order
-
-When no trajectory history exists:
-    - If reference centres are available: check the nearest site centre
-      first, then learned transitions, then distance-ranked outward
-    - Otherwise: check all sites in arbitrary order
-
-Note: distance ranking uses minimum-image convention in fractional space,
-which is only geometrically exact for orthogonal cells. For non-orthogonal
-cells the ranking is approximate, but correctness is unaffected since all
-sites are eventually checked.
-
-This approach leverages spatial relationships, learned transition patterns,
-and distance-based ordering to reduce the number of containment checks.
+- _collect_reference_centres: extracts reference centres from polyhedral
+  sites for distance-ranked ordering.
 """
 
-from collections.abc import Generator
-
-from .site_collection import SiteCollection, _NearestSiteLookup
+from .site_collection import SiteCollection, PriorityAssignmentMixin
 from .polyhedral_site import PolyhedralSite
 from .atom import Atom
 from .site import Site
@@ -53,7 +27,7 @@ from pymatgen.core import Structure # type: ignore
 import numpy as np
 
 
-class PolyhedralSiteCollection(SiteCollection):
+class PolyhedralSiteCollection(PriorityAssignmentMixin, SiteCollection):
     """A collection of PolyhedralSite objects.
     
     Extends the base SiteCollection class with specific functionality for
@@ -80,12 +54,12 @@ class PolyhedralSiteCollection(SiteCollection):
         for s in sites:
             if not isinstance(s, PolyhedralSite):
                 raise TypeError
-        super(PolyhedralSiteCollection, self).__init__(sites)
+        super().__init__(sites)
         self.sites: list[PolyhedralSite]
         self._neighbouring_sites = construct_neighbouring_sites(self.sites)
-        self._distance_ranked_sites, self._nearest_site_lookup = (
-            _compute_distance_ranked_sites(self.sites)
-        )
+        centres, site_indices = _collect_reference_centres(self.sites)
+        if centres is not None:
+            self._init_priority_ranking(centres, site_indices)
 
     def analyse_structure(self,
             atoms: list[Atom],
@@ -119,71 +93,6 @@ class PolyhedralSiteCollection(SiteCollection):
                     self.update_occupation(site, atom)
                     break
     
-    def _get_priority_sites(self, atom: Atom) -> Generator[PolyhedralSite, None, None]:
-        """Generator that yields sites in priority order for optimised atom assignment.
-
-        The checking sequence depends on available information:
-
[…]

[…]
