### 2026-02-28 1537536
Update CHANGELOG for PriorityAssignmentMixin and spherical heuristic
