### 2026-02-28 0bb3827
Address PR review: guard empty sites, add PBC test, improve docstrings
- Guard _init_priority_ranking against empty centres array
- Add PBC test for distance ranking minimum-image convention
- Assert _distance_ranked_sites is set in spherical init test
- Clarify anchor term in _get_priority_sites docstring
- Specify when fallback to neighbouring_sites applies in mixin docstring
- Add error messages to bare TypeErrors in both constructors
