### 2026-02-28 176784e
Fix CHANGELOG: keep original polyhedral heuristic entry wording
