### 2026-02-28 589a22a
Standardise test indentation and add missing blank line
Convert test_spherical_site_collection.py from tabs to 4-space
indentation, matching the rest of the test suite. Fix over-indented
test_full_optimised_assignment_integration method body. Add missing
blank line between _NearestSiteLookup and PriorityAssignmentMixin.
