+++ tests/test_spherical_site_collection.py
@@ -10,401 +10,402 @@ from site_analysis.site import Site
 
 
 class SphericalSiteCollectionTestCase(unittest.TestCase):
-	
-	def setUp(self):
-		"""Set up test fixtures with real objects rather than complex mocks."""
-		# Reset Site ID counter
-		Site._newid = 0
-		
-		# Create a real lattice
-		self.lattice = Lattice.cubic(10.0)
-		
-		# Create real SphericalSite objects
-		self.site1 = SphericalSite(
-			frac_coords=np.array([0.1, 0.1, 0.1]),
-			rcut=1.5,
-			label="site1"
-		)
-		self.site2 = SphericalSite(
-			frac_coords=np.array([0.5, 0.5, 0.5]),
-			rcut=1.5,
-			label="site2"
-		)
-		
-		# Create collection with real sites
-		self.sites = [self.site1, self.site2]
-		self.collection = SphericalSiteCollection(sites=self.sites)
-		
-		# Create real atoms
-		self.atom1 = Atom(index=0)
-		self.atom2 = Atom(index=1)
-		self.atoms = [self.atom1, self.atom2]
-		
-		# Create real structure
-		self.structure = Structure(
-			lattice=self.lattice,
-			species=["Na", "Na"],
-			coords=[[0.1, 0.1, 0.1], [0.5, 0.5, 0.5]]
-		)
-	
-	def test_initialization(self):
-		"""Test that SphericalSiteCollection initializes correctly."""
-		self.assertEqual(self.collection.sites, self.sites)
-		self.assertIsNotNone(self.collection._distance_ranked_sites)
-		
-	def test_init_with_empty_sites_list(self):
-		"""Test that __init__ works with empty sites list."""
-		collection = SphericalSiteCollection([])
-		
-		self.assertEqual(collection.sites, [])
-		self.assertEqual(collection._site_lookup, {})
-		
-	def test_init_raises_type_error_with_non_spherical_sites(self):
-		"""Test that initialisation raises TypeError with non-SphericalSite objects."""
-		# Create a mix of site types
-		non_spherical_site = Mock()
-		mixed_sites = [self.site1, non_spherical_site]
-		
-		# Test initialisation with mixed site types
-		with self.assertRaises(TypeError):
-			SphericalSiteCollection(sites=mixed_sites)
-		
-	def test_analyse_structure(self):
-		"""Test that analyse_structure calls assign_coords and assign_site_occupations."""
-		# Patch the methods we want to verify
-		with patch.object(Atom, 'assign_coords') as mock_assign_coords, \
-			 patch.object(SphericalSiteCollection, 'assign_site_occupations') as mock_assign_occupations:
-			
-			# Call analyse_structure
-			self.collection.analyse_structure(self.atoms, self.structure)
-			
-			# Verify that assign_coords was called for each atom
-			self.assertEqual(mock_assign_coords.call_count, 2)
-			
-			# Verify that assign_site_occupations was called
-			mock_assign_occupations.assert_called_once_with(self.atoms, self.structure)
-	
-	def test_site_occupation_reset(self):
-		"""Test that site occupations are reset at the beginning of assign_site_occupations."""
-		# Add atoms to sites
-		self.site1.contains_atoms = [0]
-		self.site2.contains_atoms = [1]
-		
-		# Patch update_occupation to prevent it from running
-		with patch.object(self.collection, 'update_occupation'):
-			# Call the method
-			self.collection.assign_site_occupations([], self.structure)
-			
-			# Verify sites were reset
-			self.assertEqual(self.site1.contains_atoms, [])
-			self.assertEqual(self.site2.contains_atoms, [])
-	
-	def test_high_level_site_allocation(self):
-		"""Test the high-level site allocation logic without mocking fine details."""
-		# Set up atoms with coordinates that match the sites
[…]
