### 2026-02-28 4e6545b
Update CHANGELOG heading from Unreleased to 1.5.0
