### 2026-03-01 c11be6f
Consolidate CHANGELOG heuristic entries into single bullet
