### 2026-03-01 0f46f33
Extract PriorityAssignmentMixin and improve site assignment heuristic
Move priority-based site assignment logic into a shared
PriorityAssignmentMixin in site_collection.py, used by both
PolyhedralSiteCollection and SphericalSiteCollection.

The improved heuristic tracks two recent distinct sites (catches
bounce-back) and uses precomputed distance-ranked site ordering
instead of neighbour-cutoff-based ordering. ~68% fewer containment
checks and ~1.4x wall-clock speedup on argyrodite Li6PS5Cl
trajectories.

Remove the now-redundant neighbouring_sites infrastructure from
SphericalSiteCollection (
[…]
