+++ tests/test_spherical_site_collection.py
@@ -10,680 +10,402 @@ from site_analysis.site import Site
 
 
 class SphericalSiteCollectionTestCase(unittest.TestCase):
-	
-	def setUp(self):
-		"""Set up test fixtures with real objects rather than complex mocks."""
-		# Reset Site ID counter
-		Site._newid = 0
-		
-		# Create a real lattice
-		self.lattice = Lattice.cubic(10.0)
-		
-		# Create real SphericalSite objects
-		self.site1 = SphericalSite(
-			frac_coords=np.array([0.1, 0.1, 0.1]),
-			rcut=1.5,
-			label="site1"
-		)
-		self.site2 = SphericalSite(
-			frac_coords=np.array([0.5, 0.5, 0.5]),
-			rcut=1.5,
-			label="site2"
-		)
-		
-		# Create collection with real sites
-		self.sites = [self.site1, self.site2]
-		self.collection = SphericalSiteCollection(sites=self.sites)
-		
-		# Create real atoms
-		self.atom1 = Atom(index=0)
-		self.atom2 = Atom(index=1)
-		self.atoms = [self.atom1, self.atom2]
-		
-		# Create real structure
-		self.structure = Structure(
-			lattice=self.lattice,
-			species=["Na", "Na"],
-			coords=[[0.1, 0.1, 0.1], [0.5, 0.5, 0.5]]
-		)
-	
-	def test_initialization(self):
-		"""Test that SphericalSiteCollection initializes correctly."""
-		self.assertEqual(self.collection.sites, self.sites)
-		
-	def test_init_initializes_neighbour_attributes(self):
-		"""Test that __init__ initializes neighbour-related attributes to None."""
-		sites = [self.site1, self.site2]
-		collection = SphericalSiteCollection(sites)
-		
-		# Check that neighbour attributes are initialized to None
-		self.assertIsNone(collection._neighbouring_sites)
-		self.assertIsNone(collection._current_lattice)
-	
-	def test_init_with_empty_sites_list(self):
-		"""Test that __init__ works with empty sites list."""
-		collection = SphericalSiteCollection([])
-		
-		self.assertEqual(collection.sites, [])
-		self.assertIsNone(collection._neighbouring_sites)
-		self.assertIsNone(collection._current_lattice)
-		self.assertEqual(collection._site_lookup, {})
-		
-	def test_init_sets_default_neighbour_cutoff(self):
-		"""Test that __init__ sets default neighbour_cutoff to 10.0 Å."""
-		sites = [self.site1, self.site2]
-		collection = SphericalSiteCollection(sites)
-		
-		# Check that default cutoff is 10.0
-		self.assertEqual(collection._neighbour_cutoff, 10.0)
-	
-	def test_init_sets_custom_neighbour_cutoff(self):
-		"""Test that __init__ can set custom neighbour_cutoff."""
-		sites = [self.site1, self.site2]
-		collection = SphericalSiteCollection(sites, neighbour_cutoff=15.0)
-		
-		# Check that custom cutoff is set
-		self.assertEqual(collection._neighbour_cutoff, 15.0)
-		
-	def test_init_raises_type_error_with_non_spherical_sites(self):
-		"""Test that initialisation raises TypeError with non-SphericalSite objects."""
-		# Create a mix of site types
-		non_spherical_site = Mock()
-		mixed_sites = [self.site1, non_spherical_site]
-		
-		# Test initialisation with mixed site types
-		with self.assertRaises(TypeError):
-			SphericalSiteCollection(sites=mixed_sites)
-		
-	def test_get_current_lattice_returns_stored_lattice(self):
[…]
