### 2026-03-01 c7f5c99
Unify duplicated PBC correction logic into correct_pbc()
Extract a correct_pbc() utility in pbc_utils.py that encapsulates the
unwrapping decision (reference centre vs legacy spread-based). Both
PolyhedralSite and DynamicVoronoiSite now call this instead of
duplicating the if/else block.

Remove DynamicVoronoiSite._compute_corrected_coords entirely —
calculate_centre and the batch fallback in the collection both call
correct_pbc() directly, resolving the design smell where the method's
primary effect was a side effect.

PBC delegation tests consolidated in TestCorrectPbc; site-level tests
veri
[…]
