+++ tests/test_dynamic_voronoi_site_collection.py
@@ -5,8 +5,19 @@ from pymatgen.core import Structure, Lattice
 
 from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.dynamic_voronoi_site_collection import DynamicVoronoiSiteCollection
+from site_analysis.pbc_utils import correct_pbc
 from site_analysis.atom import Atom
 from site_analysis.site import Site
+
+
+def _reference_centre(frac_coords, lattice, reference_center=None):
+	"""Compute a reference site centre for test assertions.
+
+	Replicates the logic that DynamicVoronoiSite.calculate_centre uses
+	internally, for use in tests that compare batch vs per-site results.
+	"""
+	corrected, _ = correct_pbc(frac_coords, reference_center, lattice)
+	return np.mean(corrected, axis=0) % 1.0
 
 
 class DynamicVoronoiSiteCollectionTestCase(unittest.TestCase):
@@ -156,7 +167,7 @@ class BatchCentreCalculationTestCase(unittest.TestCase):
 		np.testing.assert_array_almost_equal(site_b.centre, site_d.centre)
 
 	def test_second_frame_uses_cached_batch_path(self):
-		"""After first frame, batch path should not call _compute_corrected_coords."""
+		"""After first frame, batch path should not call correct_pbc."""
 		lattice = Lattice.cubic(10.0)
 		coords1 = [[0.1, 0.1, 0.1], [0.2, 0.2, 0.2],
 				   [0.7, 0.7, 0.7], [0.8, 0.8, 0.8]]
@@ -173,12 +184,10 @@ class BatchCentreCalculationTestCase(unittest.TestCase):
 		collection._batch_calculate_centres(struct1.frac_coords, struct1.lattice)
 		self.assertTrue(collection._centre_groups[0].initialised)
 
-		# Second frame should use vectorised path (not per-site)
-		with patch.object(site1, '_compute_corrected_coords') as mock1, \
-			 patch.object(site2, '_compute_corrected_coords') as mock2:
+		# Second frame should use vectorised path (not per-site fallback)
+		with patch('site_analysis.dynamic_voronoi_site_collection.correct_pbc') as mock_pbc:
 			collection._batch_calculate_centres(struct2.frac_coords, struct2.lattice)
-			mock1.assert_not_called()
-			mock2.assert_not_called()
+			mock_pbc.assert_not_called()
 
 		# Centres should still be updated
 		self.assertIsNotNone(site1._centre_coords)
@@ -218,8 +227,8 @@ class BatchCentreCalculationTestCase(unittest.TestCase):
 			structure = Structure(lattice, ["Na"] * 4, coords.tolist())
 			frac = structure.frac_coords
 
-			site_a._compute_corrected_coords(frac[[0, 1]], structure.lattice)
-			site_b._compute_corrected_coords(frac[[2, 3]], structure.lattice)
+			site_a._centre_coords = _reference_centre(frac[[0, 1]], structure.lattice)
+			site_b._centre_coords = _reference_centre(frac[[2, 3]], structure.lattice)
 			collection._batch_calculate_centres(frac, structure.lattice)
 
 			np.testing.assert_array_almost_equal(site_a.centre, site_c.centre)
@@ -259,8 +268,8 @@ class BatchCentreCalculationTestCase(unittest.TestCase):
 			structure = Structure(lattice, ["Na"] * 4, coords.tolist())
 			frac = structure.frac_coords
 
-			ref_a._compute_corrected_coords(frac[[0, 1]], structure.lattice)
-			ref_b._compute_corrected_coords(frac[[2, 3]], structure.lattice)
[…]
