+++ tests/test_dynamic_voronoi_site.py
@@ -72,19 +72,20 @@ class DynamicVoronoiSiteTestCase(unittest.TestCase):
                   [0.1, 0.1, 0.2]]
         structure = Structure(lattice, ["Li", "Li", "Li", "Li"], coords)
         
-        with patch('site_analysis.dynamic_voronoi_site.apply_legacy_pbc_correction') as mock_pbc:
+        with patch('site_analysis.dynamic_voronoi_site.correct_pbc') as mock_pbc:
             # Mock returns coordinates that produce centre < 1.0
             corrected_coords = np.array([[0.1, 0.1, 0.1],
                                          [0.2, 0.1, 0.1],
                                          [0.1, 0.2, 0.1],
                                          [0.1, 0.1, 0.2]])
-            mock_pbc.return_value = corrected_coords
-            
+            mock_pbc.return_value = (corrected_coords,
+                                     np.zeros((4, 3), dtype=np.int64))
+
             site.calculate_centre(structure)
-            
+
             mock_pbc.assert_called_once()
             np.testing.assert_array_almost_equal(mock_pbc.call_args[0][0], coords)
-            
+
             # Expected centre: mean = [0.125, 0.125, 0.125], no wrapping needed
             expected_centre = np.mean(corrected_coords, axis=0)
             np.testing.assert_array_almost_equal(site._centre_coords, expected_centre)
@@ -101,19 +102,20 @@ class DynamicVoronoiSiteTestCase(unittest.TestCase):
                   [0.1, 0.1, 0.2]]
         structure = Structure(lattice, ["Li", "Li", "Li", "Li"], coords)
         
-        with patch('site_analysis.dynamic_voronoi_site.apply_legacy_pbc_correction') as mock_pbc:
+        with patch('site_analysis.dynamic_voronoi_site.correct_pbc') as mock_pbc:
             # Mock returns coordinates that produce centre > 1.0 to test wrapping
             corrected_coords = np.array([[1.1, 1.1, 1.1],
                                          [1.2, 1.1, 1.1],
                                          [1.1, 1.2, 1.1],
                                          [1.1, 1.1, 1.2]])
-            mock_pbc.return_value = corrected_coords
-            
+            mock_pbc.return_value = (corrected_coords,
+                                     np.ones((4, 3), dtype=np.int64))
+
             site.calculate_centre(structure)
-            
+
             mock_pbc.assert_called_once()
             np.testing.assert_array_almost_equal(mock_pbc.call_args[0][0], coords)
-            
+
             # Expected centre with wrapping: mean([1.1, 1.2, 1.1, 1.1]) = 1.125 -> 0.125
             centre_before_wrap = np.mean(corrected_coords, axis=0)
             expected_centre = centre_before_wrap % 1.0
@@ -334,82 +336,52 @@ class DynamicVoronoiSiteCentreSerialisationTestCase(unittest.TestCase):
         np.testing.assert_array_equal(site.reference_center, reference_center)
         self.assertEqual(site.reference_indices, reference_indices)
         
-    def test_calculate_centre_uses_legacy_when_no_reference_center(self):
-        """Test that legacy PBC correction is used when reference_center is None."""
+    def test_calculate_centre_passes_none_reference_center(self):
+        """Test that correct_pbc is called with reference_center=None when not set."""
         reference_indices = [0, 1, 2, 3]
-        site = DynamicVoronoiSite(reference_indices=reference_indices)  # No reference_center
-        
[…]

[…]
