### 2026-03-01 a1da134
Address review comments on PBC correction unification
- Fix missing trailing newlines in pbc_utils.py and test_pbc_utils.py
- Replace convoluted assertion with simple kwargs check
- Clarify lattice parameter docstrings in correct_pbc and
  _store_vertex_coords
- Add type hints and align parameter order on _reference_centre test
  helper to match correct_pbc(frac_coords, reference_center, lattice)
- Split int64 dtype test to cover both legacy and reference-centre paths
- Add empty array guard in correct_pbc for zero-size input
- Move test_reset_clears_centre into its own TestResetWithCalculated
[…]
