### 2026-03-01 29144e2
Dissolve containment.py and centralise optional dependency detection
Move HAS_NUMBA into a dedicated _compat.py module as the single source
of truth for optional dependency detection. Relocate update_pbc_shifts
(numpy + numba variants) to pbc_utils.py where the other PBC utilities
live. Absorb FaceTopologyCache and its numba helpers into
polyhedral_site.py, their sole consumer. Delete containment.py and
test_containment.py, with all tests moved to test_pbc_utils.py and
test_polyhedral_site.py respectively.
