+++ site_analysis/polyhedral_site.py
@@ -12,14 +12,181 @@ import itertools
 import warnings
 
 import numpy as np
-from scipy.spatial import Delaunay, QhullError # type: ignore
+from scipy.spatial import ConvexHull, Delaunay, QhullError # type: ignore
 from pymatgen.core import Lattice, Structure
 from site_analysis.site import Site
 from site_analysis.tools import x_pbc, species_string_from_site
 from site_analysis.atom import Atom
-from site_analysis.containment import HAS_NUMBA, FaceTopologyCache, update_pbc_shifts
-from site_analysis.pbc_utils import correct_pbc
+from site_analysis._compat import HAS_NUMBA
+from site_analysis.pbc_utils import correct_pbc, update_pbc_shifts
 from typing import Any
+
+
+if HAS_NUMBA:
+    import numba  # type: ignore
+
+    @numba.njit(cache=True)  # type: ignore[misc]
+    def _numba_update_faces(
+        vertex_coords: np.ndarray,
+        face_simplices: np.ndarray,
+    ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
+        """JIT-compiled face normal and centre sign computation.
+
+        Args:
+            vertex_coords: (N_vertices, 3) array of vertex positions.
+            face_simplices: (N_faces, 3) array of vertex indices per face.
+
+        Returns:
+            Tuple of (face_normals, face_ref_points, centre_signs).
+        """
+        n_faces = face_simplices.shape[0]
+        n_vertices = vertex_coords.shape[0]
+        face_normals = np.empty((n_faces, 3))
+        face_ref_points = np.empty((n_faces, 3))
+        centre_signs = np.empty(n_faces)
+
+        centre = np.zeros(3)
+        for i in range(n_vertices):
+            for k in range(3):
+                centre[k] += vertex_coords[i, k]
+        for k in range(3):
+            centre[k] /= n_vertices
+
+        for j in range(n_faces):
+            i0 = face_simplices[j, 0]
+            i1 = face_simplices[j, 1]
+            i2 = face_simplices[j, 2]
+
+            # Edge vectors from vertex 2
+            e0x = vertex_coords[i0, 0] - vertex_coords[i2, 0]
+            e0y = vertex_coords[i0, 1] - vertex_coords[i2, 1]
+            e0z = vertex_coords[i0, 2] - vertex_coords[i2, 2]
+            e1x = vertex_coords[i1, 0] - vertex_coords[i2, 0]
+            e1y = vertex_coords[i1, 1] - vertex_coords[i2, 1]
+            e1z = vertex_coords[i1, 2] - vertex_coords[i2, 2]
+
+            # Cross product
+            face_normals[j, 0] = e0y * e1z - e0z * e1y
+            face_normals[j, 1] = e0z * e1x - e0x * e1z
+            face_normals[j, 2] = e0x * e1y - e0y * e1x
+
+            # Reference point (first vertex of face)
+            for k in range(3):
+                face_ref_points[j, k] = vertex_coords[i0, k]
+
+            # Centre sign
+            dot = 0.0
+            for k in range(3):
+                dot += face_normals[j, k] * (centre[k] - face_ref_points[j, k])
+            centre_signs[j] = np.sign(dot)
+
+        return face_normals, face_ref_points, centre_signs
+
+    @numba.njit(cache=True)  # type: ignore[misc]
+    def _numba_sn_query(
+        x_pbc_points: np.ndarray,
+        face_normals: np.ndarray,
+        face_ref_points: np.ndarray,
+        centre_signs: np.ndarray,
+    ) -> bool:
+        """JIT-compiled containment check using surface normals.
+
+        For each PBC image point, checks whether the point lies on the
[…]
