+++ site_analysis/_compat.py
@@ -1,0 +1,7 @@
+"""Optional dependency detection."""
+
+try:
+    import numba  # type: ignore
+    HAS_NUMBA = True
+except ImportError:
+    HAS_NUMBA = False

[…]
