### 2026-03-01 5122fea
Fix dtype=int to dtype=np.int64 in unwrap empty-array guard
Aligns the empty-array early return in
unwrap_vertices_to_reference_center with the int64 dtype used
consistently elsewhere in pbc_utils.
