### 2026-03-01 1698405
Update changelog for PBC unification and containment dissolution
