### 2026-03-01 e53606a
Ensure _PERIODIC_SHIFTS has explicit int64 dtype
Fixes platform-dependent dtype in the image shifts returned by
unwrap_vertices_to_reference_center, guaranteeing the int64 contract
documented in correct_pbc on all platforms including Windows.
