### 2026-03-01 2716e37
Unify PBC correction logic and dissolve containment module
- Extract correct_pbc() in pbc_utils.py encapsulating the PBC unwrapping
  strategy decision (reference centre vs legacy spread-based), replacing
  duplicated if/else blocks in PolyhedralSite and DynamicVoronoiSite
- Remove DynamicVoronoiSite._compute_corrected_coords; calculate_centre
  and the batch fallback now call correct_pbc() directly
- Dissolve containment.py: move HAS_NUMBA to _compat.py, PBC shift
  updates to pbc_utils.py, and FaceTopologyCache into polyhedral_site.py
- Ensure _PERIODIC_SHIFTS uses explicit int64 dtype for c
[…]
