+++ site_analysis/dynamic_voronoi_site.py
@@ -11,7 +11,7 @@ from typing import Any
 from .site import Site
 from .atom import Atom
 from pymatgen.core import Lattice, Structure
-from site_analysis.pbc_utils import apply_legacy_pbc_correction, unwrap_vertices_to_reference_center
+from site_analysis.pbc_utils import correct_pbc
 
 class DynamicVoronoiSite(Site):
     """Site subclass corresponding to Voronoi cells with centres dynamically 
@@ -99,38 +99,8 @@ class DynamicVoronoiSite(Site):
             the centre as the mean of the reference atom positions.
         """
         ref_coords = np.array([structure[i].frac_coords for i in self.reference_indices])
-        self._compute_corrected_coords(ref_coords, structure.lattice)
-
-    def _compute_corrected_coords(self,
-            frac_coords: np.ndarray,
-            lattice: Lattice) -> np.ndarray:
-        """Apply PBC correction and compute the site centre.
-
-        Performs full PBC unwrapping using either the reference centre
-        method or the legacy spread-based method.
-
-        Args:
-            frac_coords: Raw fractional coordinates of the reference atoms,
-                shape ``(n_reference, 3)``.
-            lattice: Lattice for Cartesian distance calculations
-                (used only with reference centres).
-
-        Returns:
-            Integer image shifts applied, shape ``(n_reference, 3)``.
-
-        Note:
-            Also sets ``_centre_coords`` as a side-effect (mean of
-            PBC-corrected coordinates, wrapped to [0, 1)).
-        """
-        if self.reference_center is not None:
-            corrected, image_shifts = unwrap_vertices_to_reference_center(
-                frac_coords, self.reference_center, lattice,
-                return_image_shifts=True)
-        else:
-            corrected = apply_legacy_pbc_correction(frac_coords)
-            image_shifts = np.round(corrected - frac_coords).astype(np.int64)
+        corrected, _ = correct_pbc(ref_coords, self.reference_center, structure.lattice)
         self._centre_coords = np.mean(corrected, axis=0) % 1.0
-        return image_shifts
         
     @property
     def centre(self) -> np.ndarray:
