+++ site_analysis/pbc_utils.py
@@ -1,4 +1,8 @@
-"""Utilities for handling periodic boundary conditions."""
+"""Utilities for handling periodic boundary conditions.
+
+Includes PBC correction, reference-centre unwrapping, and incremental
+PBC shift updates (with optional numba acceleration).
+"""
 
 from __future__ import annotations
 
@@ -6,6 +10,8 @@ from typing import Literal, overload
 
 import numpy as np
 from pymatgen.core import Lattice
+
+from site_analysis._compat import HAS_NUMBA
 
 def apply_legacy_pbc_correction(frac_coords: np.ndarray) -> np.ndarray:
     """Apply the legacy spread-based periodic boundary condition handling.
@@ -35,9 +41,10 @@ def apply_legacy_pbc_correction(frac_coords: np.ndarray) -> np.ndarray:
     return corrected_coords
 
 # Generate all 27 possible shifts: [-1,0,1] for each dimension
-_PERIODIC_SHIFTS = np.array([[dx, dy, dz] for dx in [-1, 0, 1] 
-                                          for dy in [-1, 0, 1] 
-                                          for dz in [-1, 0, 1]])
+_PERIODIC_SHIFTS = np.array([[dx, dy, dz] for dx in [-1, 0, 1]
+                                          for dy in [-1, 0, 1]
+                                          for dz in [-1, 0, 1]],
+                                         dtype=np.int64)
                                         
 @overload
 def unwrap_vertices_to_reference_center(
@@ -78,7 +85,7 @@ def unwrap_vertices_to_reference_center(
     """
     if frac_coords.size == 0:
         if return_image_shifts:
-            return frac_coords, np.zeros((0, 3), dtype=int)
+            return frac_coords, np.zeros((0, 3), dtype=np.int64)
         return frac_coords
 
     n_vertices = len(frac_coords)
@@ -101,3 +108,128 @@ def unwrap_vertices_to_reference_center(
     if return_image_shifts:
         return result, image_shifts
     return np.asarray(result)  # no-op; satisfies mypy no-any-return
+
+
+def correct_pbc(
+    frac_coords: np.ndarray,
+    reference_center: np.ndarray | None,
+    lattice: Lattice,
+) -> tuple[np.ndarray, np.ndarray]:
+    """Apply PBC correction to fractional coordinates.
+
+    Selects the appropriate unwrapping strategy based on whether a
+    reference centre is provided. When a reference centre is given,
+    vertices are unwrapped to their closest periodic images relative
+    to that centre. Otherwise, the legacy spread-based correction is
+    applied.
+
+    Args:
+        frac_coords: Fractional coordinates, shape ``(n, 3)``.
+        reference_center: Reference centre for unwrapping, or ``None``
+            for legacy spread-based correction.
+        lattice: Lattice for Cartesian distance calculations.
+            Passed to the reference-centre unwrapping path; unused by
+            the legacy spread-based path.
+
+    Returns:
+        Tuple of ``(corrected_coords, image_shifts)`` where both have
+        shape ``(n, 3)`` and ``image_shifts`` has ``int64`` dtype.
+    """
+    if frac_coords.size == 0:
+        return frac_coords, np.zeros((0, 3), dtype=np.int64)
+    if reference_center is not None:
+        return unwrap_vertices_to_reference_center(
+            frac_coords, reference_center, lattice,
+            return_image_shifts=True)
+    corrected = apply_legacy_pbc_correction(frac_coords)
+    image_shifts = np.round(corrected - frac_coords).astype(np.int64)
+    return corrected, image_shifts
+
+
+def _numpy_update_pbc_shifts(
+    frac_coords: np.ndarray,
+    cached_raw_frac: np.ndarray,
[…]
