### 2026-03-01 a2e644c
Improve documentation readability, accuracy, and coverage
Fix broken content:
- Fix cross-references in pbc_handling.md (incorrect doc paths)
- Fix broken DOI link in publications.md (httpe:// typo)
- Fix clone URL in installation.md (underscore vs hyphen)
- Fix undefined variable in reference_workflow.md example
- Fix stale API call in quickstart notebook (.with_alignment_options
  renamed to .with_structure_alignment)
- Close unclosed figure directives in pbc_handling.md

Fix typos and spelling:
- Fix "quantiative" and "spatical" in what_is_site_analysis.md
- Fix "analyzing" to "analysing" i
[…]
