+++ docs/source/concepts/site_collections.md
@@ -2,159 +2,43 @@
 
 ## What is a Site Collection?
 
-A **site collection** is a fundamental abstraction in `site_analysis` that manages a group of related sites and handles the assignment of atoms to these sites based on their positions in a crystal structure. While individual sites define bounded volumes, site collections orchestrate how atoms are assigned to sites and maintain the overall spatial organization.
+A **site collection** manages a group of related [sites](sites.md) and handles the assignment of atoms to these sites based on their positions in a crystal structure. While individual sites define bounded volumes, site collections orchestrate how atoms are assigned to sites at each timestep.
 
-Site collections are essential for several reasons:
-- They ensure consistent handling of atom assignments across multiple sites
-- They implement different assignment strategies appropriate for each site type
-- They provide methods for analyzing structures and tracking occupations
-- They maintain relationships between sites (e.g., neighboring sites)
+Users rarely interact directly with site collection classes — the `TrajectoryBuilder` creates the appropriate collection automatically when you call methods like `.with_spherical_sites()` or `.with_polyhedral_sites()`. This page explains the assignment logic that happens behind the scenes, which is useful for understanding how your results are produced.
 
-## Core Site Collection Properties
+## Assignment Strategies
 
-All site collection types in `site_analysis` share common responsibilities:
+Each site collection type implements a different strategy for assigning atoms to sites. The choice of strategy is determined by the spatial properties of the corresponding site type.
 
-- **Site management**: Organizing and maintaining a group of related sites
-- **Atom assignment**: Implementing the logic to determine which atoms belong to which sites
-- **Structure analysis**: Processing structures to update site assignments
-- **Occupation tracking**: Maintaining records of which atoms occupy which sites
-- **Transition tracking**: Recording movements of atoms between sites
+### Priority-Based Assignment (Spherical and Polyhedral)
 
-Each specific site collection type provides specialised implementations of these responsibilities appropriate for its corresponding site type.
+Spherical and polyhedral sites can overlap or leave gaps between them. These collections check sites one at a time and assign the atom to the first containing site found. To avoid unnecessary containment checks, a priority heuristic determines the order in which sites are tested:
 
-## Site Collections and Spatial Discretisation
+1. **Recent history**: the atom's most recently occupied site(s) are checked first
+2. **Learned transitions**: sites that atoms have previously transitioned to from the anchor site, ordered by frequency
+3. **Distance ranking**: remaining sites ordered by distance from the anchor site centre
 
-Site collections determine how space is discretised and how ambiguities are resolved. This is particularly important for:
+If no trajectory history exists (e.g. the first timestep), the nearest site centre is used as the starting anchor instead.
 
-1. **Overlapping sites**: How to decide which site an atom belongs to when it could belong to multiple sites
-2. **Gaps between sites**: How to handle atoms that don't clearly belong to any site
-3. **Prioritization**: What rules to follow when assignment is ambiguous
+This ordering means atoms that remain in or near their current site are resolved in a single check, and common transitions are tested early. In practice, this eliminates the majority of containment checks compared to a naive sequential scan.
 
-Different site collection types implement different strategies for these scenarios, resulting in distinct behaviors that influence analysis results.
+If no site contains the atom, it is left unassigned (`None`).
 
-## Available Site Collection Types
+For polyhedral sites, the collection also maintains information about face-sharing neighbouring polyhedra, which can be used for identifying connected diffusion pathways.
 
[…]
