+++ docs/source/modules/reference_workflow.rst
@@ -1,0 +1,32 @@
+site\_analysis.reference\_workflow
+----------------------------------
+
+.. automodule:: site_analysis.reference_workflow
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: site_analysis.reference_workflow.reference_based_sites
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: site_analysis.reference_workflow.structure_aligner
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: site_analysis.reference_workflow.coord_finder
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: site_analysis.reference_workflow.index_mapper
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: site_analysis.reference_workflow.site_factory
+    :members:
+    :undoc-members:
+    :show-inheritance:

[…]
