### 2026-03-01 dfeb273
Fix citation year to 2026
