### 2026-03-01 e7d3464
Add CITATION.cff
