### 2026-03-01 304ea68
Replace deprecated tqdm_notebook with tqdm.auto
tqdm.auto auto-detects the environment (terminal, Jupyter notebook,
Jupyter lab), removing the need for callers to choose between
progress=True and progress='notebook'. The progress parameter is
now a simple boolean.
