### 2026-03-01 0b01c2e
Address review comments: type hints, import ordering, docstring
Add type hints to trajectory_from_structures signature (Sequence[Structure],
-> None). Group standard library imports together at the top. Capitalise
docstring parameter description.
