### 2026-03-01 ce94277
Add types-tqdm dev dependency and remove type: ignore
types-tqdm provides proper type stubs, consistent with the existing
types-PyYAML and scipy-stubs dev dependencies.
