+++ pyproject.toml
@@ -34,6 +34,7 @@ dev = [
     "pytest-cov",           # Pytest coverage plugin
     "mypy",                 # Type checking
     "types-PyYAML",         # Type stubs for PyYAML
+    "types-tqdm",           # Type stubs for tqdm
     "scipy-stubs",          # Type stubs for SciPy
     "sphinx",               # Documentation generation
     "nbsphinx",             # For notebook integration in docs
