### 2026-03-01 38ee120
Add PEP 561 py.typed marker
Signals to type checkers that the package ships inline type
annotations, consistent with the project's use of type hints
throughout.
