### 2026-03-01 818baf0
Bump version to 1.6.0 and update changelog
