### 2026-03-01 de4e898
Remove redundant mypy tqdm ignore now that types-tqdm is installed
