+++ mypy.ini
@@ -21,8 +21,5 @@ ignore_missing_imports = True
 [mypy-pymatgen.*]
 ignore_missing_imports = True
 
-[mypy-tqdm.*]
-ignore_missing_imports = True
-
 [mypy-scipy.*]
 ignore_missing_imports = True
