### 2026-03-01 112fa19
Remove redundant mypy ignores for pymatgen and scipy
pymatgen ships its own type information and scipy-stubs is already
a dev dependency, making both ignore_missing_imports entries
unnecessary.
