### 2026-03-01 5506323
Update changelog with mypy config cleanup
