### 2026-03-01 593177d
Add Python 3.14 to CI matrix and separate mypy into own job
Run tests across 3.10-3.14 (with allow-prereleases for 3.14). Move
static type checking into a dedicated job so it runs once rather
than repeating across the version matrix.
