### 2026-03-01 8e04263
Remove unnecessary allow-prereleases from CI
