+++ .github/workflows/build.yml
@@ -16,7 +16,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{matrix.python-version}}
-          allow-prereleases: true
           cache: 'pip'
 
       - name: Install Python dependencies
