### 2026-03-01 3f6d028
Update changelog with CI changes
