### 2026-03-01 54c8a2a
Modernise tqdm, add py.typed, and tidy CI and mypy config
- Replace deprecated tqdm_notebook with tqdm.auto for automatic
  environment detection; simplify progress parameter to bool
- Add PEP 561 py.typed marker for downstream type checking
- Add types-tqdm dev dependency; remove inline type: ignore
- Remove redundant mypy ignores for tqdm, pymatgen, and scipy
  (now covered by type stubs)
- Add Python 3.14 to CI test matrix; separate mypy into own job
