+++ site_analysis/trajectory.py
@@ -28,20 +28,24 @@ Note:
 """
 
 from collections import Counter
-from tqdm import tqdm, tqdm_notebook # type: ignore
-from .polyhedral_site_collection import PolyhedralSiteCollection
-from .polyhedral_site import PolyhedralSite
-from .voronoi_site import VoronoiSite
-from .voronoi_site_collection import VoronoiSiteCollection
-from .spherical_site import SphericalSite
-from .spherical_site_collection import SphericalSiteCollection
+from collections.abc import Iterable
+from typing import Sequence
+
+from tqdm.auto import tqdm
+
+from pymatgen.core import Structure
+
+from .atom import Atom
 from .dynamic_voronoi_site import DynamicVoronoiSite
 from .dynamic_voronoi_site_collection import DynamicVoronoiSiteCollection
-from .site_collection import SiteCollection
+from .polyhedral_site import PolyhedralSite
+from .polyhedral_site_collection import PolyhedralSiteCollection
 from .site import Site
-from .atom import Atom
-from typing import Sequence
-from pymatgen.core import Structure
+from .site_collection import SiteCollection
+from .spherical_site import SphericalSite
+from .spherical_site_collection import SphericalSiteCollection
+from .voronoi_site import VoronoiSite
+from .voronoi_site_collection import VoronoiSiteCollection
 
 class Trajectory:
     """Class for performing sites analysis on simulation trajectories."""
@@ -248,30 +252,24 @@ class Trajectory:
         """
         return self.sites_trajectory
 
-    def trajectory_from_structures(self, structures, progress=False):
+    def trajectory_from_structures(
+        self,
+        structures: Sequence[Structure],
+        progress: bool = False,
+    ) -> None:
         """Generate a trajectory from a list of structures.
-        
-        This method processes each structure in sequence, appending a timestep
-        for each one.
-        
+
         Args:
-            structures: list of pymatgen Structure objects to analyse.
-            progress: If False, no progress is shown. If True, a progress bar is displayed.
-                If 'notebook', a notebook-friendly progress bar is displayed.
-                
-        Notes:
-            This method uses tqdm for progress tracking when enabled.
+            structures: The structures to analyse, one per timestep.
+            progress: Show a progress bar. Automatically selects the
+                appropriate widget for terminal or notebook environments.
         """
-        generator = enumerate(structures, 1)
-        params = {'iterable': generator,
-                  'total': len(structures),
-                  'unit': ' steps',
-                  'desc': 'Analysing trajectory'}
+        generator: Iterable[tuple[int, Structure]] = enumerate(structures, 1)
         if progress:
-            if progress=='notebook':
-                generator = tqdm_notebook(**params)
-            else:
-                generator = tqdm(**params)
+            generator = tqdm(generator,
+                             total=len(structures),
+                             unit=' steps',
+                             desc='Analysing trajectory')
         for timestep, s in generator:
             self.append_timestep(s, t=timestep)
    
