+++ mypy.ini
@@ -17,12 +17,3 @@ ignore_missing_imports = True
 
 [mypy-monty.*]
 ignore_missing_imports = True
-
-[mypy-pymatgen.*]
-ignore_missing_imports = True
-
-[mypy-tqdm.*]
-ignore_missing_imports = True
-
-[mypy-scipy.*]
-ignore_missing_imports = True
