### 2026-03-16 #9606 @LIVazquezS
Dear Authors,

First of all, I thank you for your patience.

The presented code `site-analysis` presents a tool for the analysis of molecular dynamics trajectories of ionic transport in solid-state. The code looks very well established, with roots that can be traced back more than 20 years ago and a number of proved applications.

I have the following observations.

## General Checks

### Example Usage
The only aspect that I believe require further improvement is that related to examples of use. I tested the two tutorials in the documentation. The first one (_Comparing Site Definitions: Basic Ion Migration in a FCC Lattice_)  has a bug related to the library `figure_formatting`, that, I assume, is only for aesthetic settings as I could run the example after removing it.

The second tutorial (_Using the Builder Pattern with Complex Crystal Structures_) was not possible to finish it, as the trajectories are not available.

I suggest the authors to update this section and also to provide another example from one of the papers where `site-analysis` has been used.

## Overall

The code presented here is an established software for the calculation of properties related to ionic transport. Besides the need of further details on the tutorial section. The work is in great shape and could be accepted as it is.
