### 2026-03-17 #9606 @bjmorgan
> The only aspect that I believe require further improvement is that related to examples of use. I tested the two tutorials in the documentation. The first one (_Comparing Site Definitions: Basic Ion Migration in a FCC Lattice_) has a bug related to the library `figure_formatting`, that, I assume, is only for aesthetic settings as I could run the example after removing it.
>
> The second tutorial (_Using the Builder Pattern with Complex Crystal Structures_) was not possible to finish it, as the trajectories are not available.
>
> I suggest the authors to update this section and also to provide another example from one of the papers where `site-analysis` has been used.

@LIVazquezS thank you for the review.

I think issues 1 and 2 both point to the same issue: the current tutorials are written for rendering into the documentation, rather than as standalone tutorials that users can work through themselves. My suggestion is that I rewrite both of these as proper self-contained tutorials, with clear prerequisites and instructions that users can follow step-by-step.

Regarding the trajectory data for the second tutorial: the data files are included in the GitHub repository under docs/source/tutorials/data/. Could you let me know how you were trying to run the tutorial? Did you clone the repository, or were you copying code from the rendered documentation? This will help me ensure the rewritten tutorials are clear about how to obtain the necessary data.

On your suggestion to add another example: is there specific functionality or use case not covered by the existing two tutorials that you think would benefit from being presented in tutorial format in addition to the conventional documentation?
