### 2026-03-17 bjmorgan/site-analysis#43 @bjmorgan
Rewrite tutorials as self-contained, user-facing tutorials
## Summary

Rewrite the existing tutorials as proper self-contained tutorials that users can work through step-by-step.

The current tutorial notebooks (`simple_fcc_example.ipynb` and `argyrodite_site_definitions.ipynb`) are written as complete worked examples for rendering into the documentation, rather than as standalone tutorials. This causes two problems identified during JOSS review:

1. The first tutorial depends on `figure_formatting`, a custom library for aesthetics that is not a dependency of `site-analysis`. Users attempting to run the tutorial code hit an import error.
2. The second tutorial references trajectory data files via relative paths with no instructions on how to obtain them, making it impossible to follow along without cloning the repository.

## Proposed changes

- Rewrite both tutorials as markdown files with code blocks that users can copy and run
- Add clear prerequisites and learning objectives to each tutorial
- Add instructions for obtaining any required data files
- Move figure generation into scripts in `docs/scripts/` that also serve as a check that the tutorial code runs correctly
- Use only standard dependencies in tutorial code (`site-analysis`, `pymatgen`, `numpy`, `matplotlib`)

Ref: https://github.com/openjournals/joss-reviews/issues/9606#issuecomment-4069457678
