### 2026-03-17 #9606 @LIVazquezS
@bjmorgan Thanks for your prompt reply.

Regarding the first point, I agree. I think it would be easier for the users to open the tutorials in a separated jupyter notebook than copying and pasting the code.

For the trajectory data, I cloned the repository and create a local environment. Then, I extracted the code from the rendered documentation. It was not clear for me that the trajectories were in the `docs/source/tutorials/data` folder. I think a reference to the location of the files should be enough.

Finally, one aspect that I found very interesting is the possibility of calculating average residence times, as in this [paper](https://pubs.acs.org/doi/10.1021/acs.chemmater.5c02303). So, I think it would be very useful to provide a guidance on how to make such calculations. If it is not too much burden.

Once addressed those minimal points, I would be happy to recommend acceptance.
