### 2026-03-18 7cde131
Add Site.residence_times() method
Add a method to compute per-atom residence time run lengths from site
trajectory data. Includes a filter_length parameter for smoothing out
brief excursions before computing run lengths.

Closes #44
