### 2026-03-18 b253019
Raise ValueError for negative filter_length
