### 2026-03-18 2fe0198
Exclude truncated edge runs and remove edge gap filling
Runs touching the first or last timestep are now excluded by default
(include_edge_runs=False) since they underestimate the true residence
time. Filtering no longer fills gaps at the trajectory edges, as this
would bias towards longer occupation times.
