+++ tests/test_site.py
@@ -277,7 +277,7 @@ class SiteTestCase(unittest.TestCase):
         self.assertIn('average_occupation', summary)
         self.assertIsNone(summary['average_occupation'])
 
-    # --- residence_times: Phase 1 (core behaviour, no filtering) ---
+    # --- residence_times: core run extraction (include_edge_runs=True) ---
 
     def test_residence_times_empty_trajectory(self):
         """Test that residence_times returns empty tuple for empty trajectory."""
@@ -294,94 +294,138 @@ class SiteTestCase(unittest.TestCase):
         """Test single atom occupying site for all timesteps."""
         site = ConcreteSite()
         site.trajectory = [[1], [1], [1]]
-        self.assertEqual(site.residence_times(), (3,))
+        self.assertEqual(site.residence_times(include_edge_runs=True), (3,))
 
     def test_residence_times_gap_in_middle(self):
         """Test single atom with a gap produces two separate runs."""
         site = ConcreteSite()
         site.trajectory = [[1], [], [1]]
-        self.assertEqual(site.residence_times(), (1, 1))
+        self.assertEqual(site.residence_times(include_edge_runs=True), (1, 1))
 
     def test_residence_times_gap_at_start(self):
         """Test that leading empty timesteps are not counted."""
         site = ConcreteSite()
         site.trajectory = [[], [1], [1]]
-        self.assertEqual(site.residence_times(), (2,))
+        self.assertEqual(site.residence_times(include_edge_runs=True), (2,))
 
     def test_residence_times_gap_at_end(self):
         """Test that trailing empty timesteps are not counted."""
         site = ConcreteSite()
         site.trajectory = [[1], [1], []]
-        self.assertEqual(site.residence_times(), (2,))
+        self.assertEqual(site.residence_times(include_edge_runs=True), (2,))
 
     def test_residence_times_multiple_atoms_independent_runs(self):
         """Test multiple atoms produce independent run lengths."""
         site = ConcreteSite()
         site.trajectory = [[1], [1], [2], [2], [2]]
-        result = site.residence_times()
+        result = site.residence_times(include_edge_runs=True)
         self.assertCountEqual(result, (2, 3))
 
     def test_residence_times_multiple_atoms_same_timestep(self):
         """Test atoms co-occupying a site produce correct per-atom runs."""
         site = ConcreteSite()
         site.trajectory = [[1, 2], [1, 2], [1]]
-        result = site.residence_times()
+        result = site.residence_times(include_edge_runs=True)
         self.assertCountEqual(result, (3, 2))
 
     def test_residence_times_single_timestep(self):
         """Test single occupied timestep returns run of length 1."""
         site = ConcreteSite()
         site.trajectory = [[1]]
+        self.assertEqual(site.residence_times(include_edge_runs=True), (1,))
+
+    # --- residence_times: edge run exclusion (default) ---
+
+    def test_residence_times_excludes_run_at_trajectory_start(self):
+        """Test that a run touching the first timestep is excluded by default."""
+        site = ConcreteSite()
+        site.trajectory = [[1], [1], [], [1], [1], [1], []]
+        self.assertEqual(site.residence_times(), (3,))
+
+    def test_residence_times_excludes_run_at_trajectory_end(self):
+        """Test that a run touching the last timestep is excluded by default."""
+        site = ConcreteSite()
+        site.trajectory = [[], [1], [1], [1], [], [1], [1]]
+        self.assertEqual(site.residence_times(), (3,))
+
+    def test_residence_times_excludes_both_edge_runs(self):
+        """Test that runs at both edges are excluded, keeping interior runs."""
[…]
