+++ site_analysis/site.py
@@ -256,7 +256,9 @@ class Site(ABC):
         occupied_timesteps = sum(1 for timestep in self.trajectory if timestep)
         return occupied_timesteps / len(self.trajectory)
         
-    def residence_times(self, filter_length: int = 0) -> tuple[int, ...]:
+    def residence_times(self,
+                        filter_length: int = 0,
+                        include_edge_runs: bool = False) -> tuple[int, ...]:
         """Compute per-atom residence time run lengths from the site trajectory.
 
         For each atom that visits this site, builds a binary
@@ -264,25 +266,36 @@ class Site(ABC):
         the lengths of consecutive occupied runs. The result is a flat tuple
         of all run lengths from all atoms.
 
+        By default, runs that touch the first or last timestep are excluded
+        because they are truncated by the trajectory boundary and
+        underestimate the true residence time. Set ``include_edge_runs=True``
+        to include them.
+
         Args:
-            filter_length: Maximum gap length to fill before computing run
-                lengths. Gaps of ``filter_length`` or fewer consecutive
-                unoccupied frames are filled (treated as if the atom remained
-                in the site) provided the gap is flanked by occupied frames
-                from the same atom (or is at the trajectory edge with one
-                occupied neighbour). Default is 0 (no filtering).
+            filter_length: Maximum interior gap length to fill before
+                computing run lengths. Gaps of ``filter_length`` or fewer
+                consecutive unoccupied frames are filled (treated as if the
+                atom remained in the site) provided the gap is flanked by
+                occupied frames from the same atom on both sides. Gaps at the
+                trajectory edges are never filled. Default is 0 (no
+                filtering).
+            include_edge_runs: Whether to include runs that touch the first
+                or last timestep. Default is False (exclude truncated runs).
 
         Returns:
             A tuple of consecutive-occupation run lengths for all atoms
             that visit this site. Returns an empty tuple if the trajectory
             is empty or the site is never occupied.
 
+        Raises:
+            ValueError: If ``filter_length`` is negative.
+
         Examples:
-            >>> site.trajectory = [[1], [1], [1]]
+            >>> site.trajectory = [[], [1], [1], [1], []]
             >>> site.residence_times()
             (3,)
 
-            >>> site.trajectory = [[1], [], [1]]
+            >>> site.trajectory = [[], [1], [], [1], []]
             >>> site.residence_times()
             (1, 1)
             >>> site.residence_times(filter_length=1)
@@ -297,6 +310,7 @@ class Site(ABC):
             all_atoms.update(timestep)
         if not all_atoms:
             return ()
+        n_timesteps = len(self.trajectory)
         run_lengths: list[int] = []
         for atom in sorted(all_atoms):
             occupied = np.array([atom in ts for ts in self.trajectory], dtype=bool)
@@ -307,19 +321,26 @@ class Site(ABC):
             diffs = np.diff(padded.astype(int))
             starts = np.where(diffs == 1)[0]
             ends = np.where(diffs == -1)[0]
-            run_lengths.extend((ends - starts).tolist())
+            for s, e in zip(starts, ends):
+                if not include_edge_runs:
+                    if s == 0 or e == n_timesteps:
+                        continue
+                run_lengths.append(e - s)
         return tuple(run_lengths)
 
     @staticmethod
     def _filter_gaps(occupied: np.ndarray, filter_length: int) -> np.ndarray:
-        """Fill short gaps in an occupied/not-occupied boolean array.
+        """Fill short interior gaps in an occupied/not-occupied boolean array.
+
+        Only gaps flanked by occupied frames on both sides are filled.
+        Gaps at the trajectory edges are never filled.
 
         Args:
[…]

[…]
