### 2026-03-18 cdb4cef
Add test for filtering with edge gaps on both sides
