### 2026-03-18 6fc2463
Raise TypeError for non-integer filter_length
