### 2026-03-18 cb5a057
Improve residence_times() docstring examples and Raises section
Add contrasting example showing edge run exclusion behaviour, and
document TypeError in Raises section.
