### 2026-03-18 2acf796
Add guidance on choosing appropriate filter lengths
Document the intended use case for filtering (brief boundary-crossing
artefacts), the per-site independence caveat for rapid oscillation
patterns, and how heavy filtering dependence can signal poor site
definitions.
