### 2026-03-18 3b36094
Remove median and max from statistical analysis example
Residence times follow an exponential distribution in a hopping model,
so only the mean is a useful summary statistic.
