### 2026-03-18 31eaa6a
Bump version to 1.7.0 and update CHANGELOG
