+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [1.7.0] - 2026-03-18
+
+### Added
+
+- `Site.residence_times()` method for computing per-atom consecutive-occupation run lengths from site trajectory data.
+- `filter_length` parameter for smoothing brief boundary-crossing artefacts before computing run lengths.
+- `include_edge_runs` parameter to control whether runs truncated by trajectory boundaries are included (excluded by default to avoid biasing towards shorter times).
+- Residence time analysis guide in documentation.
+
 ## [1.6.0] - 2026-03-01
 
 ### Changed
