### 2026-03-18 766538f
Add Site.residence_times() method (#46)
Add Site.residence_times() method for computing per-atom consecutive-occupation
run lengths from site trajectory data (#44).

- Builds a per-atom boolean occupancy array and extracts consecutive run lengths
- filter_length parameter smooths brief boundary-crossing artefacts (interior gaps only)
- include_edge_runs parameter controls whether runs truncated by trajectory boundaries
  are included (excluded by default to avoid biasing towards shorter times)
- Input validation for negative and non-integer filter_length values
- Comprehensive test coverage (3
[…]
