+++ docs/source/index.md
@@ -52,6 +52,7 @@ guides/voronoi_sites
 guides/dynamic_voronoi_sites
 guides/polyhedral_sites
 guides/trajectories
+guides/residence_times
 
 ```
 ```{toctree}

[…]
