### 2026-03-18 b79db28
Add Trajectory.transition_counts() and transition_probabilities() methods
Add methods to aggregate per-site transition data into square
dict-of-dicts matrices. transition_counts() returns raw hop counts and
transition_probabilities() returns row-normalised probabilities.

Both methods accept a `by` parameter: 'site' (default) keys by site
index; 'label' aggregates across sites sharing a label, skipping
unlabelled sites. The output is always square with explicit zeros for
missing transitions.

Closes #45
