+++ tests/test_trajectory.py
@@ -1,4 +1,5 @@
 import unittest
+from collections import Counter
 import numpy as np
 from pymatgen.core import Structure, Lattice
 
@@ -534,6 +535,199 @@ class TrajectoryFunctionalityTestCase(unittest.TestCase):
         finally:
             # Clean up
             os.unlink(temp_filename)
+
+
+class TransitionCountsBySiteTestCase(unittest.TestCase):
+    """Tests for Trajectory.transition_counts(by='site')."""
+
+    def setUp(self):
+        Site._newid = 0
+        self.site0 = SphericalSite(frac_coords=np.array([0.0, 0.0, 0.0]), rcut=0.3, label="A")
+        self.site1 = SphericalSite(frac_coords=np.array([0.5, 0.5, 0.5]), rcut=0.3, label="B")
+        self.site2 = SphericalSite(frac_coords=np.array([0.5, 0.0, 0.5]), rcut=0.3)
+        self.sites = [self.site0, self.site1, self.site2]
+        self.atoms = [Atom(index=0)]
+        self.trajectory = Trajectory(sites=self.sites, atoms=self.atoms)
+
+    def test_all_sites_have_transitions(self):
+        """Test square dict with transitions on every site."""
+        self.site0.transitions = Counter({1: 3, 2: 1})
+        self.site1.transitions = Counter({0: 2})
+        self.site2.transitions = Counter({0: 1, 1: 4})
+        result = self.trajectory.transition_counts(by='site')
+        self.assertEqual(result, {
+            0: {0: 0, 1: 3, 2: 1},
+            1: {0: 2, 1: 0, 2: 0},
+            2: {0: 1, 1: 4, 2: 0},
+        })
+
+    def test_site_with_no_outgoing_transitions(self):
+        """Test that a site with no transitions appears as a row of zeros."""
+        self.site0.transitions = Counter({1: 5})
+        # site1 and site2 have no transitions
+        result = self.trajectory.transition_counts(by='site')
+        self.assertEqual(result[1], {0: 0, 1: 0, 2: 0})
+        self.assertEqual(result[2], {0: 0, 1: 0, 2: 0})
+
+    def test_asymmetric_transitions(self):
+        """Test that A->B and B->A are independent entries."""
+        self.site0.transitions = Counter({1: 3})
+        # site1 has no transition back to site0
+        result = self.trajectory.transition_counts(by='site')
+        self.assertEqual(result[0][1], 3)
+        self.assertEqual(result[1][0], 0)
+
+    def test_no_transitions_at_all(self):
+        """Test square dict of all zeros when no transitions recorded."""
+        result = self.trajectory.transition_counts(by='site')
+        self.assertEqual(result, {
+            0: {0: 0, 1: 0, 2: 0},
+            1: {0: 0, 1: 0, 2: 0},
+            2: {0: 0, 1: 0, 2: 0},
+        })
+
+    def test_single_site(self):
+        """Test single-site trajectory produces 1x1 zero matrix."""
+        Site._newid = 0
+        site = SphericalSite(frac_coords=np.array([0.0, 0.0, 0.0]), rcut=0.3)
+        trajectory = Trajectory(sites=[site], atoms=[Atom(index=0)])
+        result = trajectory.transition_counts(by='site')
+        self.assertEqual(result, {0: {0: 0}})
+
+
+class TransitionCountsByLabelTestCase(unittest.TestCase):
+    """Tests for Trajectory.transition_counts(by='label')."""
+
+    def setUp(self):
+        Site._newid = 0
+        # Two sites labelled "A", one labelled "B", one unlabelled
+        self.site0 = SphericalSite(frac_coords=np.array([0.0, 0.0, 0.0]), rcut=0.3, label="A")
[…]
