### 2026-03-19 10af7fa
Add Literal type hint, invalid-index guard, and additional tests
- Use Literal['site', 'label'] for the `by` parameter on
  transition_counts() and transition_probabilities() for better
  static analysis and IDE support.
- Guard against unknown destination indices in the by='site' path,
  emitting a warning and skipping rather than raising a KeyError.
- Add test for self-transitions in by='site' mode.
- Add test for zero-transition labels in by='label' probabilities.
