### 2026-03-19 00c1558
Add @overload signatures and clarify self-transition test
- Add @overload signatures for transition_counts() and
  transition_probabilities() so callers get precise return types
  depending on the by='site' or by='label' argument.
- Rename and expand docstring for self-transition test to clarify
  that self-transitions are not expected in normal trajectory data.
