### 2026-03-19 67c2f67
Introduce TransitionTable result type
Replace dict-of-dicts and tuple returns with a frozen TransitionTable
dataclass that provides multiple access patterns:

- .matrix for the raw numpy array
- .loc[from_key, to_key] for key-based lookup
- .to_dict() for square dict-of-dicts

transition_counts() and transition_probabilities() now return
TransitionTable directly, with an optional keys parameter for custom
row/column ordering. This removes the need for separate _matrix
methods, @overload signatures, and union return types.
