+++ site_analysis/trajectory.py
@@ -30,11 +30,14 @@ Note:
 import warnings
 from collections import Counter
 from collections.abc import Iterable
-from typing import Literal, Sequence, overload
+from typing import Literal, Sequence
 
+import numpy as np
 from tqdm.auto import tqdm
 
 from pymatgen.core import Structure
+
+from site_analysis.transition_table import TransitionTable
 
 from .atom import Atom
 from .dynamic_voronoi_site import DynamicVoronoiSite
@@ -163,32 +166,36 @@ class Trajectory:
         """
         return [s.label for s in self.sites]
 
-    @overload
-    def transition_counts(self, by: Literal['site'] = 'site') -> dict[int, dict[int, int]]: ...
-    @overload
-    def transition_counts(self, by: Literal['label']) -> dict[str, dict[str, int]]: ...
-
-    def transition_counts(self, by: Literal['site', 'label'] = 'site') -> dict[int, dict[int, int]] | dict[str, dict[str, int]]:
-        """Return transition counts as a square dict-of-dicts.
+    def transition_counts(
+        self,
+        by: Literal['site', 'label'] = 'site',
+        *,
+        keys: Sequence[int] | Sequence[str] | None = None,
+    ) -> TransitionTable:
+        """Return transition counts as a :class:`TransitionTable`.
 
         Args:
             by: Aggregation key. ``'site'`` (default) keys by site index;
                 ``'label'`` aggregates by site label (unlabelled sites are
                 skipped).
+            keys: Optional key ordering for rows and columns. If ``None``,
+                keys are sorted.
 
         Returns:
-            A square dict-of-dicts ``{from_key: {to_key: count}}``.
+            A :class:`TransitionTable` of integer counts.
 
         Raises:
-            ValueError: If ``by`` is not ``'site'`` or ``'label'``.
+            ValueError: If ``by`` is not ``'site'`` or ``'label'``, or if
+                *keys* contains a key not present in the data.
         """
         if by == 'site':
-            indices = [s.index for s in self.sites]
-            index_set = set(indices)
-            result_site: dict[int, dict[int, int]] = {
-                i: {j: 0 for j in indices} for i in indices
-            }
+            site_keys: list[int] = sorted(s.index for s in self.sites)
+            index_set = set(site_keys)
+            idx_lookup = {k: i for i, k in enumerate(site_keys)}
+            n = len(site_keys)
+            matrix = np.zeros((n, n), dtype=int)
             for site in self.sites:
+                row = idx_lookup[site.index]
                 for dest, count in site.transitions.items():
                     if dest not in index_set:
                         warnings.warn(
@@ -197,16 +204,16 @@ class Trajectory:
                             stacklevel=2,
                         )
                         continue
-                    result_site[site.index][dest] = count
-            return result_site
-        if by == 'label':
+                    matrix[row, idx_lookup[dest]] = count
+            result_keys: list[int] | list[str] = site_keys
+        elif by == 'label':
             index_to_label = {
                 s.index: s.label for s in self.sites if s.label is not None
             }
-            labels = sorted(set(index_to_label.values()))
-            result_label: dict[str, dict[str, int]] = {
-                la: {lb: 0 for lb in labels} for la in labels
-            }
+            label_keys: list[str] = sorted(set(index_to_label.values()))
+            label_lookup = {k: i for i, k in enumerate(label_keys)}
+            n = len(label_keys)
+            matrix = np.zeros((n, n), dtype=int)
             for site in self.sites:
                 src_label = index_to_label.get(site.index)
                 if src_label is None:
@@ -215,17 +222,21 @@ class Trajectory:
                     dest_label = index_to_label.get(dest)
[…]
