+++ tests/test_trajectory.py
@@ -550,12 +550,12 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         self.trajectory = Trajectory(sites=self.sites, atoms=self.atoms)
 
     def test_all_sites_have_transitions(self):
-        """Test square dict with transitions on every site."""
+        """Test correct counts with transitions on every site."""
         self.site0.transitions = Counter({1: 3, 2: 1})
         self.site1.transitions = Counter({0: 2})
         self.site2.transitions = Counter({0: 1, 1: 4})
         result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result, {
+        self.assertEqual(result.to_dict(), {
             0: {0: 0, 1: 3, 2: 1},
             1: {0: 2, 1: 0, 2: 0},
             2: {0: 1, 1: 4, 2: 0},
@@ -566,25 +566,21 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         self.site0.transitions = Counter({1: 5})
         # site1 and site2 have no transitions
         result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result[1], {0: 0, 1: 0, 2: 0})
-        self.assertEqual(result[2], {0: 0, 1: 0, 2: 0})
+        np.testing.assert_array_equal(result.matrix[1], [0, 0, 0])
+        np.testing.assert_array_equal(result.matrix[2], [0, 0, 0])
 
     def test_asymmetric_transitions(self):
         """Test that A->B and B->A are independent entries."""
         self.site0.transitions = Counter({1: 3})
         # site1 has no transition back to site0
         result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result[0][1], 3)
-        self.assertEqual(result[1][0], 0)
+        self.assertEqual(result.loc[0, 1], 3)
+        self.assertEqual(result.loc[1, 0], 0)
 
     def test_no_transitions_at_all(self):
-        """Test square dict of all zeros when no transitions recorded."""
+        """Test all zeros when no transitions recorded."""
         result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result, {
-            0: {0: 0, 1: 0, 2: 0},
-            1: {0: 0, 1: 0, 2: 0},
-            2: {0: 0, 1: 0, 2: 0},
-        })
+        np.testing.assert_array_equal(result.matrix, np.zeros((3, 3)))
 
     def test_self_transitions_are_preserved_if_present(self):
         """Test that pre-populated self-transition keys are preserved.
@@ -596,16 +592,16 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         """
         self.site0.transitions = Counter({0: 2, 1: 3})
         result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result[0][0], 2)
-        self.assertEqual(result[0][1], 3)
+        self.assertEqual(result.loc[0, 0], 2)
+        self.assertEqual(result.loc[0, 1], 3)
 
     def test_unknown_destination_index_warns_and_is_skipped(self):
         """Test that a transition to a non-existent site index emits a warning."""
         self.site0.transitions = Counter({1: 3, 99: 5})
         with self.assertWarns(UserWarning):
             result = self.trajectory.transition_counts(by='site')
-        self.assertEqual(result[0][1], 3)
-        self.assertNotIn(99, result[0])
+        self.assertEqual(result.loc[0, 1], 3)
+        self.assertNotIn(99, result.keys)
 
     def test_single_site(self):
         """Test single-site trajectory produces 1x1 zero matrix."""
[…]

[…]
