### 2026-03-19 fb1568c
Harden TransitionTable and address review feedback
- Rename data field to matrix; drop redundant .matrix property
- Make keys a tuple and matrix read-only for true immutability
- Add __eq__ (using np.array_equal) and __hash__
- Move reorder() onto TransitionTable; require exact key permutation
- Unknown destination site index now raises ValueError (was warning)
- Transitions to/from unlabelled sites now emit a warning with count
- Add NaN post-condition assertion in transition_probabilities()
- Tighten type hints on _LocAccessor and to_dict()
- Use relative import for transition_table
- Export
[…]
