+++ site_analysis/transition_table.py
@@ -8,6 +8,7 @@ access patterns.
 from __future__ import annotations
 
 from dataclasses import dataclass, field
+from typing import Sequence
 
 import numpy as np
 
@@ -21,7 +22,7 @@ class _LocAccessor:
     def __init__(self, table: TransitionTable) -> None:
         self._table = table
 
-    def __getitem__(self, key: tuple) -> int | float:
+    def __getitem__(self, key: tuple[int | str, int | str]) -> int | float:
         from_key, to_key = key
         index = self._table._key_to_index
         try:
@@ -32,7 +33,7 @@ class _LocAccessor:
             j = index[to_key]
         except KeyError:
             raise KeyError(to_key) from None
-        value: int | float = self._table.data[i, j].item()
+        value: int | float = self._table.matrix[i, j].item()
         return value
 
 
@@ -49,29 +50,29 @@ class TransitionTable:
 
     Args:
         keys: Row and column labels (site indices or site labels).
-        data: A square 2-D numpy array of transition values.
+        matrix: A square 2-D numpy array of transition values.
 
     Raises:
-        ValueError: If *data* is not 2-D and square, if
-            ``len(keys) != data.shape[0]``, or if *keys* contains
+        ValueError: If *matrix* is not 2-D and square, if
+            ``len(keys) != matrix.shape[0]``, or if *keys* contains
             duplicates.
     """
 
-    keys: list[int] | list[str]
-    data: np.ndarray
+    keys: tuple[int, ...] | tuple[str, ...] | tuple[()]
+    matrix: np.ndarray
     _key_to_index: dict[int | str, int] = field(
         init=False, repr=False, compare=False,
     )
 
     def __post_init__(self) -> None:
-        if self.data.ndim != 2 or self.data.shape[0] != self.data.shape[1]:
+        if self.matrix.ndim != 2 or self.matrix.shape[0] != self.matrix.shape[1]:
             raise ValueError(
-                f"data must be a square 2-D array, got shape {self.data.shape}"
+                f"matrix must be a square 2-D array, got shape {self.matrix.shape}"
             )
-        if len(self.keys) != self.data.shape[0]:
+        if len(self.keys) != self.matrix.shape[0]:
             raise ValueError(
-                f"len(keys) ({len(self.keys)}) != data dimension "
-                f"({self.data.shape[0]})"
+                f"len(keys) ({len(self.keys)}) != matrix dimension "
+                f"({self.matrix.shape[0]})"
             )
         if len(set(self.keys)) != len(self.keys):
             raise ValueError("keys must not contain duplicates")
@@ -80,11 +81,7 @@ class TransitionTable:
             self, '_key_to_index',
             {k: i for i, k in enumerate(self.keys)},
         )
-
-    @property
-    def matrix(self) -> np.ndarray:
-        """The transition data as a 2-D numpy array."""
-        return self.data
+        self.matrix.flags.writeable = False
 
     @property
     def loc(self) -> _LocAccessor:
@@ -96,7 +93,7 @@ class TransitionTable:
         """
         return _LocAccessor(self)
 
-    def to_dict(self) -> dict:
+    def to_dict(self) -> dict[int | str, dict[int | str, int | float]]:
         """Convert to a square dict-of-dicts.
 
         Returns:
@@ -104,14 +101,51 @@ class TransitionTable:
         """
         return {
             self.keys[i]: {
-                self.keys[j]: self.data[i, j].item()
+                self.keys[j]: self.matrix[i, j].item()
                 for j in range(len(self.keys))
             }
             for i in range(len(self.keys))
         }
 
+    def reorder(self, keys: Sequence[int] | Sequence[str]) -> TransitionTable:
+        """Return a new table with rows and columns reordered.
+
[…]
