+++ site_analysis/__init__.py
@@ -19,3 +19,4 @@ from site_analysis.builders import (
     create_trajectory_with_polyhedral_sites,
     create_trajectory_with_dynamic_voronoi_sites
 )
+from site_analysis.transition_table import TransitionTable
