### 2026-03-19 94142f2
Address remaining review items
- Replace assert with RuntimeError for NaN check (not stripped by -O)
- Defensive copy of input matrix in TransitionTable.__post_init__
- Add type: ignore[arg-type] on intentionally invalid by='invalid' tests
