### 2026-03-19 61b1425
Convert TransitionTable from dataclass to plain class
The dataclass was generating only __init__ while we overrode __eq__,
__hash__, __repr__, and fought frozen=True with object.__setattr__
calls. A plain class with __slots__ and a __setattr__ guard is more
honest about the immutability contract and simpler to read.
