+++ site_analysis/transition_table.py
@@ -1,13 +1,12 @@
 """Transition table for storing labelled transition data.
 
-Provides the :class:`TransitionTable` dataclass, which stores transition
+Provides the :class:`TransitionTable` class, which stores transition
 counts or probabilities as a labelled square matrix with convenient
 access patterns.
 """
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
 from typing import Sequence
 
 import numpy as np
@@ -37,14 +36,13 @@ class _LocAccessor:
         return value
 
 
-@dataclass(frozen=True)
 class TransitionTable:
     """A labelled square matrix of transition data.
 
     Stores transition counts or probabilities with named keys for
     rows and columns. Provides multiple access patterns:
 
-    - ``.matrix`` — the raw :class:`numpy.ndarray`
+    - ``.matrix`` — the raw (read-only) :class:`numpy.ndarray`
     - ``.loc[from_key, to_key]`` — key-based lookup
     - ``.to_dict()`` — square dict-of-dicts
 
@@ -58,32 +56,39 @@ class TransitionTable:
             duplicates.
     """
 
-    keys: tuple[int, ...] | tuple[str, ...] | tuple[()]
-    matrix: np.ndarray
-    _key_to_index: dict[int | str, int] = field(
-        init=False, repr=False, compare=False,
-    )
+    __slots__ = ('_keys', '_matrix', '_key_to_index', '_loc')
 
-    def __post_init__(self) -> None:
-        # Defensive copy so we don't mutate the caller's array
-        object.__setattr__(self, 'matrix', np.array(self.matrix, copy=True))
-        if self.matrix.ndim != 2 or self.matrix.shape[0] != self.matrix.shape[1]:
+    def __init__(
+        self,
+        keys: tuple[int, ...] | tuple[str, ...] | tuple[()],
+        matrix: np.ndarray,
+    ) -> None:
+        self._matrix = np.array(matrix, copy=True)
+        if self._matrix.ndim != 2 or self._matrix.shape[0] != self._matrix.shape[1]:
             raise ValueError(
-                f"matrix must be a square 2-D array, got shape {self.matrix.shape}"
+                f"matrix must be a square 2-D array, got shape {self._matrix.shape}"
             )
-        if len(self.keys) != self.matrix.shape[0]:
+        if len(keys) != self._matrix.shape[0]:
             raise ValueError(
-                f"len(keys) ({len(self.keys)}) != matrix dimension "
-                f"({self.matrix.shape[0]})"
+                f"len(keys) ({len(keys)}) != matrix dimension "
+                f"({self._matrix.shape[0]})"
             )
-        if len(set(self.keys)) != len(self.keys):
+        if len(set(keys)) != len(keys):
             raise ValueError("keys must not contain duplicates")
-        # frozen=True requires object.__setattr__ for init
-        object.__setattr__(
-            self, '_key_to_index',
-            {k: i for i, k in enumerate(self.keys)},
-        )
-        self.matrix.flags.writeable = False
+        self._keys = keys
+        self._key_to_index = {k: i for i, k in enumerate(keys)}
+        self._loc = _LocAccessor(self)
+        self._matrix.flags.writeable = False
+
+    @property
+    def keys(self) -> tuple[int, ...] | tuple[str, ...] | tuple[()]:
+        """Row and column labels."""
+        return self._keys
+
+    @property
+    def matrix(self) -> np.ndarray:
+        """The transition data as a read-only 2-D numpy array."""
+        return self._matrix
 
     @property
     def loc(self) -> _LocAccessor:
@@ -93,7 +98,7 @@ class TransitionTable:
 
             table.loc["A", "B"]  # value for A -> B
         """
-        return _LocAccessor(self)
+        return self._loc
 
[…]
