### 2026-03-19 4b9d0fc
Polish TransitionTable and NaN diagnostics
- Use explicit _frozen sentinel for immutability guard instead of
  order-dependent hasattr check.
- Catch duplicate keys in reorder() with a length check before the
  set comparison, giving a clearer error message.
- Drop redundant tuple[()] from key type union.
- Include offending keys and row sums in NaN RuntimeError message.
