### 2026-03-19 1e07692
Replace .loc[] accessor with .get() method
Remove _LocAccessor helper class and cached _loc field. Key-based
lookup is now table.get(from_key, to_key) with @overload signatures
enforcing matched key types (int, int or str, str).
