### 2026-03-19 0c2f498
Validate unknown dest in by='label' path and count source drops
- In by='label', validate destination indices against all site indices
  and raise ValueError for truly unknown indices (not just unlabelled).
- Count transitions from unlabelled source sites in the dropped total.
- Remove redundant np.array() copy in reorder().
- Update docstring to reflect .get() API (was .loc[]).
