### 2026-03-19 7fd1979
Validate unknown dest indices for unlabelled sources and fix errstate
- In by='label', validate destination indices against all site indices
  even when the source site is unlabelled, before skipping.
- Add divide='ignore' to np.errstate alongside invalid='ignore' to
  suppress the divide-by-zero warning from np.where on zero-sum rows.
