### 2026-03-19 9ccc069
Avoid divide-by-zero in transition_probabilities
Only divide rows with nonzero sums instead of dividing all rows and
masking with np.where. Removes the need for np.errstate and the NaN
post-condition check.
