### 2026-03-19 7bf4c1f
Drop __hash__ from TransitionTable
TransitionTable has no use case as a dict key or set member, and the
mutable numpy array makes hash invariants fragile. Replace __hash__
with explicit None to make the type unhashable.
