### 2026-03-19 79d327a
Sort missing/extra keys in reorder() error messages
Ensures deterministic error messages regardless of set ordering.
