### 2026-03-19 c7eeefe
Add transition_counts() and transition_probabilities() methods
Add methods to aggregate per-site transition data into a TransitionTable,
a labelled square matrix with convenient access patterns (.matrix,
.get(), .to_dict(), .reorder()).

- transition_counts(by='site'|'label') returns raw hop counts
- transition_probabilities(by='site'|'label') returns row-normalised
  probabilities (rows with no outgoing transitions remain as zeros)
- by='label' aggregates across sites sharing a label; unlabelled sites
  are skipped with a warning if transitions are dropped
- Optional keys parameter for custom
[…]
