+++ site_analysis/trajectory.py
@@ -27,13 +27,17 @@ Note:
     ...              .build())
 """
 
+import warnings
 from collections import Counter
 from collections.abc import Iterable
-from typing import Sequence
+from typing import Literal, Sequence
 
+import numpy as np
 from tqdm.auto import tqdm
 
 from pymatgen.core import Structure
+
+from .transition_table import TransitionTable
 
 from .atom import Atom
 from .dynamic_voronoi_site import DynamicVoronoiSite
@@ -161,7 +165,126 @@ class Trajectory:
             A list of site labels (or None for sites without labels).
         """
         return [s.label for s in self.sites]
-   
+
+    def transition_counts(
+        self,
+        by: Literal['site', 'label'] = 'site',
+        *,
+        keys: Sequence[int] | Sequence[str] | None = None,
+    ) -> TransitionTable:
+        """Return transition counts as a :class:`TransitionTable`.
+
+        Args:
+            by: Aggregation key. ``'site'`` (default) keys by site index;
+                ``'label'`` aggregates by site label (unlabelled sites are
+                skipped, with a warning if any transitions are dropped).
+            keys: Optional key ordering for rows and columns. If ``None``,
+                keys are sorted. Must be a permutation of the default keys.
+
+        Returns:
+            A :class:`TransitionTable` of integer counts.
+
+        Raises:
+            ValueError: If ``by`` is not ``'site'`` or ``'label'``, if
+                *keys* does not match the default key set, or if a site
+                has a transition to an unknown site index.
+        """
+        if by == 'site':
+            site_keys = tuple(sorted(s.index for s in self.sites))
+            index_set = set(site_keys)
+            idx_lookup = {k: i for i, k in enumerate(site_keys)}
+            n = len(site_keys)
+            matrix = np.zeros((n, n), dtype=int)
+            for site in self.sites:
+                row = idx_lookup[site.index]
+                for dest, count in site.transitions.items():
+                    if dest not in index_set:
+                        raise ValueError(
+                            f"Site {site.index} has a transition to unknown "
+                            f"site index {dest}."
+                        )
+                    matrix[row, idx_lookup[dest]] = count
+            result_keys: tuple[int, ...] | tuple[str, ...] = site_keys
+        elif by == 'label':
+            all_site_indices = {s.index for s in self.sites}
+            index_to_label = {
+                s.index: s.label for s in self.sites if s.label is not None
+            }
+            label_keys = tuple(sorted(set(index_to_label.values())))
+            label_lookup = {k: i for i, k in enumerate(label_keys)}
+            n = len(label_keys)
+            matrix = np.zeros((n, n), dtype=int)
+            dropped = 0
+            for site in self.sites:
+                src_label = index_to_label.get(site.index)
+                if src_label is None:
+                    for dest in site.transitions:
+                        if dest not in all_site_indices:
+                            raise ValueError(
+                                f"Site {site.index} has a transition to unknown "
+                                f"site index {dest}."
+                            )
+                    dropped += sum(site.transitions.values())
+                    continue
+                for dest, count in site.transitions.items():
+                    if dest not in all_site_indices:
+                        raise ValueError(
+                            f"Site {site.index} has a transition to unknown "
+                            f"site index {dest}."
+                        )
+                    dest_label = index_to_label.get(dest)
+                    if dest_label is None:
+                        dropped += count
+                        continue
+                    matrix[label_lookup[src_label], label_lookup[dest_label]] += count
+            if dropped > 0:
+                warnings.warn(
+                    f"{dropped} transition(s) involving unlabelled sites were "
[…]

[…]
