### 2026-03-19 c55d660
Add transition counts and probabilities to changelog
