### 2026-03-19 5601b6a
Move transition table changelog entry under v1.7.0
