+++ CHANGELOG.md
@@ -4,20 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Unreleased]
-
-### Added
-
-- `Trajectory.transition_counts()` and `Trajectory.transition_probabilities()` methods for aggregating per-site transition data, with `by='site'` or `by='label'` aggregation.
-- `TransitionTable` class: a labelled square matrix with `.matrix`, `.get()`, `.to_dict()`, and `.reorder()` access patterns.
-- `TransitionTable` exported from top-level `site_analysis` package.
-
 ## [1.7.0] - 2026-03-18
 
 ### Added
 
 - `Site.residence_times()` method for computing per-atom consecutive-occupation run lengths from site trajectory data.
 - `filter_length` parameter for smoothing brief boundary-crossing artefacts before computing run lengths.
+- `Trajectory.transition_counts()` and `Trajectory.transition_probabilities()` methods for aggregating per-site transition data, with `by='site'` or `by='label'` aggregation.
+- `TransitionTable` class: a labelled square matrix with `.matrix`, `.get()`, `.to_dict()`, and `.reorder()` access patterns.
+- `TransitionTable` exported from top-level `site_analysis` package.
 - `include_edge_runs` parameter to control whether runs truncated by trajectory boundaries are included (excluded by default to avoid biasing towards shorter times).
 - Residence time analysis guide in documentation.
 
