### 2026-03-19 2376743
Fix mypy error in reorder() sorted() calls
Use key=str to satisfy type checker when sorting sets of int | str keys.
