### 2026-03-19 ad34462
Use type: ignore instead of key=str for sorted() in reorder()
key=str would sort integers lexicographically (1, 10, 2). Since keys
are homogeneous at runtime, sorted() works correctly; the type checker
just cannot prove it from the union type.
