### 2026-03-19 09b4442
Make TransitionTable generic over key type
- TransitionTable is now Generic[TableKey] where TableKey is
  constrained to int or str, eliminating union types and type: ignore
  comments throughout the class.
- Add @overload signatures to transition_counts() and
  transition_probabilities() so callers get precise return types.
- Extract _normalise_counts() and _validate_destination() helpers to
  reduce duplication in trajectory.py.
