### 2026-03-19 243c910
Split transition methods into _by_site and _by_label variants
Replace transition_counts(by=...) and transition_probabilities(by=...)
with four explicit methods:

- transition_counts_by_site()
- transition_counts_by_label()
- transition_probabilities_by_site()
- transition_probabilities_by_label()

Each method has a precise return type (TransitionTable[int] or
TransitionTable[str]) with no overloads, casts, or unions needed.
Remove the now-unnecessary TransitionValidationTestCase (no by
parameter to validate).
