+++ site_analysis/trajectory.py
@@ -30,14 +30,14 @@ Note:
 import warnings
 from collections import Counter
 from collections.abc import Iterable
-from typing import Literal, Sequence, cast, overload
+from typing import Sequence
 
 import numpy as np
 from tqdm.auto import tqdm
 
 from pymatgen.core import Structure
 
-from .transition_table import TableKey, TransitionTable
+from .transition_table import TableKey, TransitionTable  # noqa: F401 (TableKey used in _normalise_counts)
 
 from .atom import Atom
 from .dynamic_voronoi_site import DynamicVoronoiSite
@@ -176,7 +176,8 @@ class Trajectory:
         probs[nonzero] = count_data[nonzero] / row_sums[nonzero, np.newaxis]
         return TransitionTable(keys=counts.keys, matrix=probs)
 
-    def _validate_destination(self, site_index: int, dest: int, valid_indices: set[int]) -> None:
+    @staticmethod
+    def _validate_destination(site_index: int, dest: int, valid_indices: set[int]) -> None:
         """Raise ValueError if dest is not in valid_indices."""
         if dest not in valid_indices:
             raise ValueError(
@@ -184,125 +185,140 @@ class Trajectory:
                 f"site index {dest}."
             )
 
-    @overload
-    def transition_counts(self, by: Literal['site'] = 'site', *, keys: Sequence[int] | None = None) -> TransitionTable[int]: ...
-    @overload
-    def transition_counts(self, by: Literal['label'] = ..., *, keys: Sequence[str] | None = None) -> TransitionTable[str]: ...
-
-    def transition_counts(
+    def transition_counts_by_site(
         self,
-        by: Literal['site', 'label'] = 'site',
         *,
-        keys: Sequence[int] | Sequence[str] | None = None,
-    ) -> TransitionTable[int] | TransitionTable[str]:
-        """Return transition counts as a :class:`TransitionTable`.
+        keys: Sequence[int] | None = None,
+    ) -> TransitionTable[int]:
+        """Return per-site transition counts as a :class:`TransitionTable`.
 
         Args:
-            by: Aggregation key. ``'site'`` (default) keys by site index;
-                ``'label'`` aggregates by site label (unlabelled sites are
-                skipped, with a warning if any transitions are dropped).
             keys: Optional key ordering for rows and columns. If ``None``,
                 keys are sorted. Must be a permutation of the default keys.
 
         Returns:
-            A :class:`TransitionTable` of integer counts.
+            A :class:`TransitionTable` of integer counts keyed by site index.
 
         Raises:
-            ValueError: If ``by`` is not ``'site'`` or ``'label'``, if
-                *keys* does not match the default key set, or if a site
-                has a transition to an unknown site index.
+            ValueError: If *keys* does not match the default key set, or
+                if a site has a transition to an unknown site index.
         """
-        if by == 'site':
-            site_keys = tuple(sorted(s.index for s in self.sites))
-            index_set = set(site_keys)
-            idx_lookup = {k: i for i, k in enumerate(site_keys)}
-            n = len(site_keys)
-            matrix = np.zeros((n, n), dtype=int)
-            for site in self.sites:
-                row = idx_lookup[site.index]
-                for dest, count in site.transitions.items():
-                    self._validate_destination(site.index, dest, index_set)
-                    matrix[row, idx_lookup[dest]] = count
-            site_table = TransitionTable(keys=site_keys, matrix=matrix)
-            if keys is not None:
-                return site_table.reorder(cast(Sequence[int], keys))
-            return site_table
-        elif by == 'label':
-            all_site_indices = {s.index for s in self.sites}
-            index_to_label = {
-                s.index: s.label for s in self.sites if s.label is not None
[…]
