### 2026-03-19 88f345f
Added missing type hint
