### 2026-03-19 abed75f
Fix changelog method names and add missing test coverage
- Update changelog to reference the actual method names
  (transition_counts_by_site, transition_counts_by_label, etc.)
- Add test for unlabelled site transitioning to an invalid index,
  covering the _validate_destination call at trajectory.py:254-255
