### 2026-03-19 2aff20a
Fix test class docstring to use full method name
