### 2026-03-19 bd4733d
Update changelog, fix test docstring, and add missing test coverage
- Update CHANGELOG entry under v1.7.0 to reference the current method
  names (transition_counts_by_site, transition_counts_by_label, etc.)
- Fix test docstring that truncated method names
- Add test for unlabelled site transitioning to an invalid destination
  index, covering the _validate_destination code path in the by_label
  flow
- Remove unnecessary noqa comment on import
