+++ tests/test_trajectory.py
@@ -538,7 +538,7 @@ class TrajectoryFunctionalityTestCase(unittest.TestCase):
 
 
 class TransitionCountsBySiteTestCase(unittest.TestCase):
-    """Tests for Trajectory.transition_counts(by='site')."""
+    """Tests for Trajectory.transition_counts_by_site()."""
 
     def setUp(self):
         Site._newid = 0
@@ -554,7 +554,7 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         self.site0.transitions = Counter({1: 3, 2: 1})
         self.site1.transitions = Counter({0: 2})
         self.site2.transitions = Counter({0: 1, 1: 4})
-        result = self.trajectory.transition_counts(by='site')
+        result = self.trajectory.transition_counts_by_site()
         self.assertEqual(result.to_dict(), {
             0: {0: 0, 1: 3, 2: 1},
             1: {0: 2, 1: 0, 2: 0},
@@ -565,7 +565,7 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         """Test that a site with no transitions appears as a row of zeros."""
         self.site0.transitions = Counter({1: 5})
         # site1 and site2 have no transitions
-        result = self.trajectory.transition_counts(by='site')
+        result = self.trajectory.transition_counts_by_site()
         np.testing.assert_array_equal(result.matrix[1], [0, 0, 0])
         np.testing.assert_array_equal(result.matrix[2], [0, 0, 0])
 
@@ -573,13 +573,13 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         """Test that A->B and B->A are independent entries."""
         self.site0.transitions = Counter({1: 3})
         # site1 has no transition back to site0
-        result = self.trajectory.transition_counts(by='site')
+        result = self.trajectory.transition_counts_by_site()
         self.assertEqual(result.get(0, 1), 3)
         self.assertEqual(result.get(1, 0), 0)
 
     def test_no_transitions_at_all(self):
         """Test all zeros when no transitions recorded."""
-        result = self.trajectory.transition_counts(by='site')
+        result = self.trajectory.transition_counts_by_site()
         np.testing.assert_array_equal(result.matrix, np.zeros((3, 3)))
 
     def test_self_transitions_are_preserved_if_present(self):
@@ -587,11 +587,11 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
 
         Note: SiteCollection.update_occupation() skips self-transitions,
         so these should not appear in normal trajectory data. This test
-        verifies that transition_counts() faithfully reports whatever is
+        verifies that transition_counts_by_site() faithfully reports whatever is
         present in site.transitions without filtering.
         """
         self.site0.transitions = Counter({0: 2, 1: 3})
-        result = self.trajectory.transition_counts(by='site')
+        result = self.trajectory.transition_counts_by_site()
         self.assertEqual(result.get(0, 0), 2)
         self.assertEqual(result.get(0, 1), 3)
 
@@ -599,19 +599,19 @@ class TransitionCountsBySiteTestCase(unittest.TestCase):
         """Test that a transition to a non-existent site index raises ValueError."""
         self.site0.transitions = Counter({1: 3, 99: 5})
         with self.assertRaises(ValueError):
-            self.trajectory.transition_counts(by='site')
+            self.trajectory.transition_counts_by_site()
 
     def test_single_site(self):
         """Test single-site trajectory produces 1x1 zero matrix."""
         Site._newid = 0
         site = SphericalSite(frac_coords=np.array([0.0, 0.0, 0.0]), rcut=0.3)
[…]
