### 2026-03-19 7965d2c
Add filter() method to TransitionTable
Allows extracting a sub-table for a subset of keys without
re-normalising. Preserves the order of the provided keys.
