### 2026-03-19 e1f6b83
Replace sorted() with list comprehensions in error messages
Avoids TypeError on mixed key types when building error messages
in reorder() and filter().
