### 2026-03-19 6317bcc
Clarify filter() docstring on key ordering
