### 2026-03-19 dafdf1b
Cache set(new_keys) in reorder() to avoid repeated computation
