### 2026-03-19 5f805ea
Add TransitionTable documentation and update tutorials to use filter()
- Add API docs page for transition_table module
- Add transition tables guide covering usage, filtering, reordering,
  and the distinction between counts and probabilities
- Update tutorial and notebook generator to use filter() instead of
  manual np.ix_ slicing
- Simplify print_transition_table helper to use table.keys directly
