+++ docs/scripts/generate_notebooks.py
@@ -1,0 +1,966 @@
+"""Generate downloadable Jupyter notebooks for the tutorials.
+
+This script creates .ipynb files that mirror the code in the
+markdown tutorials, so users can download and run them directly.
+"""
+
+import json
+import uuid
+from pathlib import Path
+
+TUTORIALS_DIR = Path(__file__).resolve().parent.parent / "source" / "tutorials"
+NOTEBOOKS_DIR = Path(__file__).resolve().parent.parent.parent / "tutorials"
+
+
+def new_notebook():
+    return {
+        "nbformat": 4,
+        "nbformat_minor": 5,
+        "metadata": {
+            "kernelspec": {
+                "display_name": "Python 3",
+                "language": "python",
+                "name": "python3",
+            },
+            "language_info": {"name": "python", "version": "3.11.0"},
+        },
+        "cells": [],
+    }
+
+
+def md_cell(source):
+    return {"cell_type": "markdown", "id": uuid.uuid4().hex[:8], "metadata": {}, "source": source.splitlines(True)}
+
+
+def code_cell(source):
+    return {
+        "cell_type": "code",
+        "id": uuid.uuid4().hex[:8],
+        "metadata": {},
+        "source": source.splitlines(True),
+        "outputs": [],
+        "execution_count": None,
+    }
+
+
+def generate_comparing_site_definitions():
+    nb = new_notebook()
+    cells = nb["cells"]
+
+    cells.append(md_cell(
+        "# Comparing Site Definitions: Ion Migration in an FCC Lattice\n"
+        "\n"
+        "This tutorial introduces the core concepts of `site-analysis` by working through a simple "
+        "example: tracking a lithium ion as it migrates between interstitial sites in an FCC lattice. "
+        "You will define sites using three different approaches \u2014 spherical, Voronoi, and polyhedral "
+        "\u2014 and compare how each one assigns the mobile ion during a migration event, highlighting "
+        "the tradeoffs between methods.\n"
+        "\n"
+        "## Prerequisites\n"
+        "\n"
+        "This tutorial requires the following packages:\n"
+        "\n"
+        "- `site-analysis`\n"
+        "- `matplotlib`\n"
+        "\n"
+        "All data in this tutorial is generated synthetically, so no external data files are needed.\n"
+        "\n"
+        "## Overview\n"
+        "\n"
+        "We will create a 3x3x3 FCC oxygen lattice and a synthetic trajectory of a single lithium ion "
+        "migrating from an octahedral site, through a tetrahedral site, to another octahedral site. "
+        "We then analyse this trajectory using three different site definitions and compare the results."
+    ))
+
+    cells.append(code_cell(
+        "%config InlineBackend.figure_format = 'retina'"
+    ))
+
+    cells.append(md_cell(
+        "## Creating the FCC lattice\n"
+        "\n"
+        "First, create the host framework \u2014 a 3x3x3 supercell of an FCC oxygen lattice:"
+    ))
+
+    cells.append(code_cell(
+        "import numpy as np\n"
+        "import matplotlib.pyplot as plt\n"
+        "from scipy.interpolate import make_interp_spline\n"
+        "from pymatgen.core import Lattice, Structure, DummySpecies\n"
+        "from site_analysis import TrajectoryBuilder\n"
+        "\n"
+        "a = 4.0  # lattice parameter\n"
+        "lattice = Lattice.cubic(a)\n"
+        "supercell_expansion = [3, 3, 3]\n"
+        "fcc_structure = Structure.from_spacegroup(\n"
+        "    sg='Fm-3m',\n"
+        "    lattice=lattice,\n"
+        "    species=['O'],\n"
[…]
