+++ docs/source/modules/transition_table.rst
@@ -1,0 +1,7 @@
+site\_analysis.transition\_table
+---------------------------------
+
+.. automodule:: site_analysis.transition_table
+    :members:
+    :undoc-members:
+    :show-inheritance:
