### 2026-03-19 e553d0f
Add filter() method to TransitionTable
Add TransitionTable.filter(keys) for extracting a sub-table containing
only the specified keys (rows and columns) without re-normalising.

- Validates against unknown and duplicate keys
- Result follows the order given in keys
- Replace manual np.ix_ slicing in tutorials with filter()
- Add TransitionTable API docs page and transition tables guide
- Add TransitionTable to docs toctree and guides index
