+++ docs/source/modules.rst
@@ -21,5 +21,6 @@ Submodules
    modules/spherical_site_collection
    modules/tools
    modules/trajectory
+   modules/transition_table
    modules/voronoi_site
    modules/voronoi_site_collection
