### 2026-03-20 ba92bbf
Use exact type lookup for site collection mapping
Replace isinstance loop with direct dict lookup for mapping site types
to collection classes. Update init tests to use real site objects
instead of mocks with patched collection classes.
