### 2026-03-20 40f9e6a
Address PR review comments
Narrow except KeyError scope to the dict lookup only, so KeyErrors from
collection constructors are not caught. Fix missing self. on assertEqual
in test___len___returns_number_of_timesteps.
