### 2026-03-20 7379d58
Simplify site collection type lookup in Trajectory
Replace isinstance loop with direct dict lookup for mapping site types
to collection classes. Replace mocked sites in init tests with real
site objects.
