+++ site_analysis/trajectory.py
@@ -91,12 +91,12 @@ class Trajectory:
         }
         
         # Find the appropriate site collection class
-        for registered_type, collection_class in site_collection_map.items():
-            if isinstance(sites[0], registered_type):
-                self.site_collection = collection_class(sites)
-                break
-        else:  # This executes if no break occurs in the for loop
-            raise TypeError(f"Site type {type(sites[0])} not recognised for Trajectory initialisation")
+        site_type = type(sites[0])
+        try:
+            collection_class = site_collection_map[site_type]
+        except KeyError:
+            raise TypeError(f"Site type {site_type} not recognised for Trajectory initialisation")
+        self.site_collection = collection_class(sites)
         
         self.sites = sites
         self.atoms = atoms
