+++ tests/test_trajectory.py
@@ -9,6 +9,10 @@ from site_analysis.spherical_site import SphericalSite
 from site_analysis.voronoi_site import VoronoiSite
 from site_analysis.dynamic_voronoi_site import DynamicVoronoiSite
 from site_analysis.site_collection import SiteCollection
+from site_analysis.polyhedral_site_collection import PolyhedralSiteCollection
+from site_analysis.voronoi_site_collection import VoronoiSiteCollection
+from site_analysis.spherical_site_collection import SphericalSiteCollection
+from site_analysis.dynamic_voronoi_site_collection import DynamicVoronoiSiteCollection
 from site_analysis.atom import Atom
 from site_analysis.site import Site
 from unittest.mock import Mock, patch, PropertyMock
@@ -21,69 +25,48 @@ import json
 class TrajectoryInitializationTestCase(unittest.TestCase):
     """Tests for Trajectory initialization with different site types."""
 
+    def setUp(self):
+        Site._newid = 0
+
     def test_initialisation_with_polyhedral_sites(self):
-        atoms = [Mock(spec=Atom, index=2)]
-        sites = [Mock(spec=PolyhedralSite, index=1)]
-        with patch('site_analysis.trajectory.PolyhedralSiteCollection') as mock_PolyhedralSiteCollection:
-            trajectory = Trajectory(atoms=atoms, sites=sites)
-        self.assertEqual(trajectory.atoms, atoms)
-        self.assertEqual(trajectory.sites, sites)
-        mock_PolyhedralSiteCollection.assert_called_with(sites)
-    
+        sites = [PolyhedralSite(vertex_indices=[0, 1, 2, 3])]
+        trajectory = Trajectory(atoms=[Atom(index=0)], sites=sites)
+        self.assertIsInstance(trajectory.site_collection, PolyhedralSiteCollection)
+
     def test_initialisation_with_voronoi_sites(self):
-        atoms = [Mock(spec=Atom, index=2)]
-        sites = [Mock(spec=VoronoiSite, index=1)]
-        with patch('site_analysis.trajectory.VoronoiSiteCollection') as mock_VoronoiSiteCollection:
-            trajectory = Trajectory(atoms=atoms, sites=sites)
-        self.assertEqual(trajectory.atoms, atoms)
-        self.assertEqual(trajectory.sites, sites)
-        mock_VoronoiSiteCollection.assert_called_with(sites)
-    
+        sites = [VoronoiSite(frac_coords=np.array([0.5, 0.5, 0.5]))]
+        trajectory = Trajectory(atoms=[Atom(index=0)], sites=sites)
+        self.assertIsInstance(trajectory.site_collection, VoronoiSiteCollection)
+
     def test_initialisation_with_spherical_sites(self):
-        atoms = [Mock(spec=Atom, index=2)]
-        sites = [Mock(spec=SphericalSite, index=1)]
-        with patch('site_analysis.trajectory.SphericalSiteCollection') as mock_SphericalSiteCollection:
-            trajectory = Trajectory(atoms=atoms, sites=sites)
-        self.assertEqual(trajectory.atoms, atoms)
-        self.assertEqual(trajectory.sites, sites)
-        mock_SphericalSiteCollection.assert_called_with(sites)
-    
+        sites = [SphericalSite(frac_coords=np.array([0.5, 0.5, 0.5]), rcut=1.0)]
+        trajectory = Trajectory(atoms=[Atom(index=0)], sites=sites)
+        self.assertIsInstance(trajectory.site_collection, SphericalSiteCollection)
+
     def test_initialisation_with_dynamic_voronoi_sites(self):
-        atoms = [Mock(spec=Atom, index=2)]
[…]
