### 2026-03-20 f5f88a8
Add frac_to_cart coordinate conversion function
