### 2026-03-20 97a73bc
Add mic_distance single-pair minimum-image distance
