+++ tests/test_distances.py
@@ -34,3 +34,60 @@ class TestFracToCart(unittest.TestCase):
         expected = lattice.get_cartesian_coords(frac)
         result = frac_to_cart(frac, lattice.matrix)
         np.testing.assert_allclose(result, expected, atol=1e-12)
+
+
+class TestMicDistance(unittest.TestCase):
+    """Tests for single-pair minimum-image distance."""
+
+    def test_same_point_returns_zero(self):
+        """Distance between a point and itself is zero."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac = np.array([0.5, 0.5, 0.5])
+        result = mic_distance(frac, frac, lattice.matrix)
+        self.assertAlmostEqual(result, 0.0, places=12)
+
+    def test_cubic_no_pbc(self):
+        """Distance within cell matches pymatgen for cubic lattice."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac1 = np.array([0.1, 0.2, 0.3])
+        frac2 = np.array([0.4, 0.5, 0.6])
+        expected = lattice.get_distance_and_image(frac1, frac2)[0]
+        result = mic_distance(frac1, frac2, lattice.matrix)
+        self.assertAlmostEqual(result, expected, places=10)
+
+    def test_cubic_across_boundary(self):
+        """Distance across periodic boundary is shorter than direct path."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac1 = np.array([0.05, 0.5, 0.5])
+        frac2 = np.array([0.95, 0.5, 0.5])
+        expected = lattice.get_distance_and_image(frac1, frac2)[0]
+        result = mic_distance(frac1, frac2, lattice.matrix)
+        self.assertAlmostEqual(result, expected, places=10)
+        # Verify it chose the short path (1.0 A) not direct (9.0 A)
+        self.assertAlmostEqual(result, 1.0, places=10)
+
+    def test_triclinic(self):
+        """Distance matches pymatgen for a triclinic lattice."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.from_parameters(5.0, 6.0, 7.0, 80, 70, 60)
+        frac1 = np.array([0.1, 0.9, 0.5])
+        frac2 = np.array([0.9, 0.1, 0.5])
+        expected = lattice.get_distance_and_image(frac1, frac2)[0]
+        result = mic_distance(frac1, frac2, lattice.matrix)
+        self.assertAlmostEqual(result, expected, places=10)
+
+    def test_matches_pymatgen_random_points(self):
+        """Distance matches pymatgen for many random point pairs."""
+        from site_analysis.distances import mic_distance
+        rng = np.random.default_rng(42)
+        lattice = Lattice.from_parameters(5.0, 6.0, 7.0, 80, 70, 60)
+        for _ in range(100):
+            frac1 = rng.random(3)
+            frac2 = rng.random(3)
+            expected = lattice.get_distance_and_image(frac1, frac2)[0]
+            result = mic_distance(frac1, frac2, lattice.matrix)
+            self.assertAlmostEqual(result, float(expected), places=10,
+                msg=f"Mismatch for {frac1} -> {frac2}")
