### 2026-03-20 17ccce3
Add all_mic_distances batch distance matrix function
