### 2026-03-20 e049aa8
Add numba-accelerated mic_distance with automatic dispatch
