### 2026-03-20 2f39c26
Add benchmark comparing distance primitives to pymatgen
