### 2026-03-20 6d6d6e1
Optimise all_mic_distances to avoid large intermediate allocation
Loop over 27 shifts with a running minimum instead of allocating
an (N, M, 27, 3) array. Uses einsum for squared norms. Reduces
200x200 batch time from ~35ms to ~20ms.
