### 2026-03-20 c02a6a8
Add numba-accelerated batch distance matrix
_all_mic_distances_numba uses parallel prange over outer loop.
At 200x200, numba batch is ~2.7x faster than pymatgen's Cython.
Dispatch in all_mic_distances selects numba when available.
