### 2026-03-20 ecfcf56
Clean up distance module from review feedback
- Make _SHIFTS_27 read-only to prevent accidental mutation
- Use scalar variables in _mic_distance_numba (consistent with batch)
- Pre-allocate work buffers in numpy all_mic_distances fallback
- Move numba batch warm-up before benchmark loop
- Add comment explaining inlined numpy benchmark
