### 2026-03-20 e0bb17c
Address review feedback and fix mypy errors
- Fix docstrings: "in Angstroms" -> "in the same units as the lattice
  matrix" (unit-agnostic)
- Document non-finite inputs as undefined behaviour
- Replace inlined benchmark with patched mic_distance call
- Add frac_to_cart convention note in docstring
- Add tests: symmetry, coords outside [0,1), diagonal zeros, empty
  second array
- Fix all mypy no-any-return errors with float() casts and np.asarray()
