### 2026-03-20 c0d3534
Address remaining review items in distance module
- Broaden module docstring to mention coordinate conversion
- Add lattice convention note to private numba docstrings
- Use np.zeros instead of np.empty for empty-input early return
