### 2026-03-20 45a29f9
Reduce fractional displacement before 27-image search
The 27-image search only checks shifts in {-1,0,1}^3, so coordinates
differing by more than one unit cell would return incorrect distances.
Reduce by the nearest integer first so the search always covers the
true minimum image.

Also adds regression tests with coordinates many cells apart, and
moves numba warm-up before mic_distance timing in the benchmark.
