### 2026-03-20 a644f3b
Add minimum-image distance primitives (#51)
New `site_analysis.distances` module providing pymatgen-free distance
calculations operating on numpy arrays and a lattice matrix:

- `mic_distance`: single-pair minimum-image distance (27-image search)
- `all_mic_distances`: batch (N,M) distance matrix
- `frac_to_cart`: fractional to Cartesian coordinate conversion

Optional numba acceleration for both single-pair and batch distances,
following the existing `HAS_NUMBA` pattern. Fractional displacements
are reduced by the nearest integer before the 27-image search, so
coordinates outside [0,1) are ha
[…]
