+++ tests/test_distances.py
@@ -1,0 +1,276 @@
+import unittest
+import numpy as np
+from pymatgen.core import Lattice
+from site_analysis._compat import HAS_NUMBA
+
+
+class TestFracToCart(unittest.TestCase):
+    """Tests for fractional to Cartesian coordinate conversion."""
+
+    def test_single_point_cubic(self):
+        """Conversion matches pymatgen for a cubic lattice."""
+        from site_analysis.distances import frac_to_cart
+        lattice = Lattice.cubic(10.0)
+        frac = np.array([0.1, 0.2, 0.3])
+        expected = lattice.get_cartesian_coords(frac)
+        result = frac_to_cart(frac, lattice.matrix)
+        np.testing.assert_allclose(result, expected, atol=1e-12)
+
+    def test_single_point_triclinic(self):
+        """Conversion matches pymatgen for a triclinic lattice."""
+        from site_analysis.distances import frac_to_cart
+        lattice = Lattice.from_parameters(5.0, 6.0, 7.0, 80, 70, 60)
+        frac = np.array([0.3, 0.5, 0.7])
+        expected = lattice.get_cartesian_coords(frac)
+        result = frac_to_cart(frac, lattice.matrix)
+        np.testing.assert_allclose(result, expected, atol=1e-12)
+
+    def test_batch_points(self):
+        """Conversion works for an (N, 3) array of points."""
+        from site_analysis.distances import frac_to_cart
+        lattice = Lattice.from_parameters(5.0, 6.0, 7.0, 80, 70, 60)
+        frac = np.array([[0.1, 0.2, 0.3],
+                         [0.4, 0.5, 0.6],
+                         [0.7, 0.8, 0.9]])
+        expected = lattice.get_cartesian_coords(frac)
+        result = frac_to_cart(frac, lattice.matrix)
+        np.testing.assert_allclose(result, expected, atol=1e-12)
+
+
+class TestMicDistance(unittest.TestCase):
+    """Tests for single-pair minimum-image distance."""
+
+    def test_same_point_returns_zero(self):
+        """Distance between a point and itself is zero."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac = np.array([0.5, 0.5, 0.5])
+        result = mic_distance(frac, frac, lattice.matrix)
+        self.assertAlmostEqual(result, 0.0, places=12)
+
+    def test_cubic_no_pbc(self):
+        """Distance within cell matches pymatgen for cubic lattice."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac1 = np.array([0.1, 0.2, 0.3])
+        frac2 = np.array([0.4, 0.5, 0.6])
+        expected = lattice.get_distance_and_image(frac1, frac2)[0]
+        result = mic_distance(frac1, frac2, lattice.matrix)
+        self.assertAlmostEqual(result, expected, places=10)
+
+    def test_cubic_across_boundary(self):
+        """Distance across periodic boundary is shorter than direct path."""
+        from site_analysis.distances import mic_distance
+        lattice = Lattice.cubic(10.0)
+        frac1 = np.array([0.05, 0.5, 0.5])
+        frac2 = np.array([0.95, 0.5, 0.5])
+        expected = lattice.get_distance_and_image(frac1, frac2)[0]
+        result = mic_distance(frac1, frac2, lattice.matrix)
+        self.assertAlmostEqual(result, expected, places=10)
+        # Verify it chose the short path (1.0 A) not direct (9.0 A)
+        self.assertAlmostEqual(result, 1.0, places=10)
+
+    def test_triclinic(self):
[…]
