### 2026-03-20 6884090
Change Atom.assign_coords to accept coordinate array
Takes a full (N, 3) fractional coordinate array instead of a pymatgen
Structure. Extracts coordinates for this atom using self.index.

Part of #53
