### 2026-03-20 eb5a202
Update atoms_from_structure to pass coordinate array
Extract structure.frac_coords at the boundary and pass to
assign_coords().

Part of #53
