### 2026-03-20 f6ca476
Change pbc_utils to accept lattice matrix instead of Lattice
Replace pymatgen.Lattice parameter with a numpy lattice matrix in
correct_pbc() and unwrap_vertices_to_reference_center(). Use
frac_to_cart() from the distances module instead of
Lattice.get_cartesian_coords().

Part of #52
