+++ tests/test_pbc_utils.py
@@ -268,7 +268,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.5, 0.5, 0.5])
 		
-		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice)
+		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice.matrix)
 		
 		# Should remain unchanged since all vertices are already close
 		np.testing.assert_array_almost_equal(result, vertex_coords, decimal=7)
@@ -286,7 +286,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.05, 0.5, 0.5])
 		
-		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice)
+		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice.matrix)
 		
 		# After unwrapping and shifting to ensure non-negative coordinates
 		expected = np.array([
@@ -312,7 +312,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		reference_center = np.array([0.25, 0.25, 0.25])
 		expected_result_centre = np.array([1.25, 1.25, 1.25])
 		
-		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice)
+		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice.matrix)
 		
 		# The centre of the unwrapped vertices should be close to the reference centre
 		result_centre = np.mean(result, axis=0)
@@ -328,7 +328,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.5, 0.5, 0.5])
 		
-		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice)
+		result = unwrap_vertices_to_reference_center(vertex_coords, reference_center, self.lattice.matrix)
 		
 		# Should remain unchanged since all vertices are already close
 		np.testing.assert_array_almost_equal(result, vertex_coords, decimal=7)
@@ -343,7 +343,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.5, 0.5, 0.5])
 		result = unwrap_vertices_to_reference_center(
-			vertex_coords, reference_center, self.lattice, return_image_shifts=True)
+			vertex_coords, reference_center, self.lattice.matrix, return_image_shifts=True)
 		self.assertIsInstance(result, tuple)
 		self.assertEqual(len(result), 2)
 		coords, shifts = result
@@ -360,7 +360,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.5, 0.5, 0.5])
 		result = unwrap_vertices_to_reference_center(
-			vertex_coords, reference_center, self.lattice, return_image_shifts=False)
+			vertex_coords, reference_center, self.lattice.matrix, return_image_shifts=False)
 		self.assertIsInstance(result, np.ndarray)
 		self.assertEqual(result.shape, (2, 3))
 
@@ -372,7 +372,7 @@ class TestReferenceBasedUnwrapping(unittest.TestCase):
 		])
 		reference_center = np.array([0.0, 0.5, 0.5])
 		coords, shifts = unwrap_vertices_to_reference_center(
[…]
