### 2026-03-20 cbf0d33
Extract structure.frac_coords at the boundary in all collections
Each analyse_structure() now extracts the coordinate array once and
passes it to assign_coords(). Removes redundant extraction in
polyhedral and dynamic voronoi collections.

Closes #53
