+++ tests/test_dynamic_voronoi_site_collection.py
@@ -93,8 +93,8 @@ class DynamicVoronoiSiteCollectionTestCase(unittest.TestCase):
 
 		atom1 = Atom(index=4)
 		atom2 = Atom(index=5)
-		atom1.assign_coords(structure)
-		atom2.assign_coords(structure)
+		atom1.assign_coords(structure.frac_coords)
+		atom2.assign_coords(structure.frac_coords)
 
 		collection.assign_site_occupations([atom1, atom2], structure)
 
