+++ tests/benchmark_containment.py
@@ -77,7 +77,7 @@ def build_argyrodite_trajectory(structures):
 def _eager_analyse_structure(self, atoms, structure):
     """Original eager analysis: assign all vertex coords upfront."""
     for a in atoms:
-        a.assign_coords(structure)
+        a.assign_coords(structure.frac_coords)
     for s in self.sites:
         s.assign_vertex_coords(structure)
     self.assign_site_occupations(atoms, structure)
