### 2026-03-20 77cc7f9
Update pbc_utils callers to pass lattice matrix
Pass lattice.matrix (numpy array) instead of Lattice objects to
correct_pbc() and related functions. The pymatgen boundary remains
at analyse_structure() / assign_vertex_coords() where
structure.lattice.matrix is extracted.

Closes #52
