### 2026-03-20 6c71ed4
Rename _pending_lattice to _pending_lattice_matrix
The attribute now stores a numpy lattice matrix rather than a pymatgen
Lattice object, so rename it to match the parameter naming used
throughout the rest of the refactoring. Also update tests to pass
lattice matrices instead of Lattice objects to internal methods.
