+++ tests/test_polyhedral_site.py
@@ -65,10 +65,10 @@ class PolyhedralSiteTestCase(unittest.TestCase):
         """reset() clears pending fractional coords and lattice."""
         site = self.site
         site._pending_frac_coords = np.zeros((4, 3))
-        site._pending_lattice = Lattice.cubic(10.0)
+        site._pending_lattice_matrix = Lattice.cubic(10.0).matrix
         site.reset()
         self.assertIsNone(site._pending_frac_coords)
-        self.assertIsNone(site._pending_lattice)
+        self.assertIsNone(site._pending_lattice_matrix)
 
     def test_reset_clears_pbc_shift_caches(self):
         """reset() clears PBC image shift caches."""
@@ -435,10 +435,10 @@ class TestNotifyStructureChanged(unittest.TestCase):
                              [0.1, 0.2, 0.3],
                              [0.5, 0.5, 0.5],
                              [0.4, 0.4, 0.4]])
-        lattice = Lattice.cubic(10.0)
-        self.site.notify_structure_changed(all_frac, lattice)
+        lattice_matrix = Lattice.cubic(10.0).matrix
+        self.site.notify_structure_changed(all_frac, lattice_matrix)
         np.testing.assert_array_equal(self.site._pending_frac_coords, all_frac)
-        self.assertIs(self.site._pending_lattice, lattice)
+        self.assertIs(self.site._pending_lattice_matrix, lattice_matrix)
 
     def test_contains_point_triggers_assign_from_pending(self):
         """contains_point flushes pending data before testing."""
@@ -447,13 +447,13 @@ class TestNotifyStructureChanged(unittest.TestCase):
                              [0.4, 0.6, 0.6],
                              [0.6, 0.6, 0.4],
                              [0.6, 0.4, 0.6]])
-        lattice = Lattice.cubic(10.0)
+        lattice_matrix = Lattice.cubic(10.0).matrix
         site = PolyhedralSite(vertex_indices=[1, 2, 3, 4])
-        site.notify_structure_changed(all_frac, lattice)
+        site.notify_structure_changed(all_frac, lattice_matrix)
         # After contains_point, pending data should be cleared
         site.contains_point(np.array([0.5, 0.5, 0.5]))
         self.assertIsNone(site._pending_frac_coords)
-        self.assertIsNone(site._pending_lattice)
+        self.assertIsNone(site._pending_lattice_matrix)
         self.assertIsNotNone(site.vertex_coords)
 
     def test_assign_from_pending_uses_vertex_indices(self):
@@ -462,9 +462,9 @@ class TestNotifyStructureChanged(unittest.TestCase):
                              [0.1, 0.2, 0.3],
                              [0.5, 0.5, 0.5],
                              [0.4, 0.4, 0.4]])
-        lattice = Lattice.cubic(10.0)
+        lattice_matrix = Lattice.cubic(10.0).matrix
         with patch.object(self.site, '_store_vertex_coords') as mock_store:
-            self.site._assign_from_pending(all_frac, lattice)
+            self.site._assign_from_pending(all_frac, lattice_matrix)
             called_frac = mock_store.call_args[0][0]
             expected = all_frac[[1, 3]]
             np.testing.assert_array_equal(called_frac, expected)
@@ -475,7 +475,7 @@ class TestStoreVertexCoords(unittest.TestCase):
 
     def setUp(self):
         Site._newid = 0
-        self.lattice = Lattice.cubic(10.0)
+        self.lattice_matrix = Lattice.cubic(10.0).matrix
 
     def test_first_call_computes_full_shifts(self):
         """First call with no cached shifts uses full computation."""
[…]
