### 2026-03-20 67234da
Fix missed .matrix calls and stale type annotation in tests
Update _reference_centre helper signature from lattice: Lattice to
lattice_matrix: np.ndarray. Fix three call sites that passed
struct.lattice instead of struct.lattice.matrix.
