### 2026-03-20 94ca44f
Fix stale docstring referencing "lattice" instead of "lattice matrix"
