### 2026-03-20 df477dd
Cache structure.frac_coords before atom loops
Avoid repeated property access inside per-atom loops in
atoms_from_structure, benchmarks, and test helpers.
