### 2026-03-20 828c6c0
Replace pymatgen Lattice with numpy lattice matrix in pbc_utils (#52)
Replace pymatgen.Lattice parameter with a numpy lattice matrix in
correct_pbc() and unwrap_vertices_to_reference_center(). Use
frac_to_cart() from the distances module instead of
Lattice.get_cartesian_coords().

Propagate the type change to callers in polyhedral_site,
dynamic_voronoi_site, and their collections. The pymatgen boundary
remains at analyse_structure() / assign_vertex_coords() where
structure.lattice.matrix is extracted.

Closes #52
