### 2026-03-20 7eb284b
Change Atom.assign_coords to accept coordinate array (#53)
Change Atom.assign_coords() to accept a fractional coordinate array
instead of a pymatgen Structure. Extract structure.frac_coords at
the boundary in each site collection's analyse_structure() method.

This removes the last direct pymatgen dependency from the Atom class
and ensures the coordinate array is extracted once per frame rather
than once per atom.

Closes #53
