### 2026-03-20 1297e1d
Replace lattice.get_all_distances with all_mic_distances in VoronoiSiteCollection
Use the numpy distance function from site_analysis.distances instead
of pymatgen Lattice.get_all_distances(). assign_site_occupations now
takes a lattice matrix instead of a Structure.

Part of #54
