### 2026-03-20 b926e88
Replace lattice.get_all_distances with all_mic_distances in DynamicVoronoiSiteCollection
Same change as VoronoiSiteCollection: assign_site_occupations now
takes a lattice matrix and uses all_mic_distances() from the
distances module.

Closes #54
