### 2026-03-20 faba6b5
Replace Lattice.get_distance_and_image with mic_distance in SphericalSite
contains_point() and contains_atom() now take a lattice_matrix
(numpy array) instead of a pymatgen Lattice object. Uses
mic_distance() from the distances module. Remove isinstance type
checks that enforced the Lattice type.

Part of #55
