### 2026-03-20 df7bf2d
Pass lattice matrix through SphericalSiteCollection to sites
assign_site_occupations() takes a lattice matrix instead of a
Structure. analyse_structure() extracts structure.lattice.matrix at
the boundary.

Closes #55
