### 2026-03-20 757358c
Fix British English in SphericalSite docstrings
Replace 'center' with 'centre' and 'centered' with 'centred' in
module and class docstrings.
