### 2026-03-20 031c0f9
Fix Trajectory.assign_site_occupations to extract lattice matrix
Trajectory.assign_site_occupations() now extracts structure.lattice.matrix
and passes it to the site collection, matching the updated signatures.
Also fixes British English in voronoi_site_collection module docstring and
the benchmark file's direct assign_site_occupations call.
