### 2026-03-20 2594ae0
Remove vacuous test_non_cubic_lattice test
