### 2026-03-21 2c0df14
Update SphericalSite class docstring to reflect lattice matrix usage
