### 2026-03-21 d847d50
Revert "Fix Trajectory.assign_site_occupations to extract lattice matrix"
This reverts commit feec70390df798ffae4ccd5d36568c52da72ca4b.
