+++ tests/test_trajectory.py
@@ -307,17 +307,19 @@ class TrajectoryFunctionalityTestCase(unittest.TestCase):
         self.assertEqual(labels[1], "site2")
         
     def test_assign_site_occupations(self):
-        """Test that assign_site_occupations extracts lattice matrix and delegates."""
+        """Test that assign_site_occupations delegates to site_collection."""
+        # Setup
         structure = Mock()
-        lattice_matrix = Mock()
-        structure.lattice.matrix = lattice_matrix
-
+        
+        # Mock the site_collection
         self.trajectory.site_collection = Mock()
-
+        
+        # Call the method
         self.trajectory.assign_site_occupations(structure)
-
+        
+        # Check delegation
         self.trajectory.site_collection.assign_site_occupations.assert_called_once_with(
-            self.atoms, lattice_matrix
+            self.atoms, structure
         )
     
     def test_trajectory_from_structures_with_progress(self):
