### 2026-03-21 3813ac6
Revert "Fix Trajectory.assign_site_occupations to extract lattice matrix"
This reverts commit 031c0f9cd62cd9fbdf47c50575ae2fb0229bb734.
