### 2026-03-21 a4e85f1
Fix British English in voronoi_site_collection docstrings
Also fix benchmark to pass lattice_matrix to assign_site_occupations.
