### 2026-03-21 0a32487
Fix remaining American English in voronoi_site_collection docstring
