+++ tests/test_polyhedral_site_collection.py
@@ -108,8 +108,11 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
             # Verify each site was notified of the new structure
             self.assertEqual(mock_notify.call_count, 3)
 
-            # Verify assign_site_occupations was called with atoms and structure
-            mock_assign.assert_called_once_with(self.atoms, self.structure)
+            # Verify assign_site_occupations was called with atoms and lattice_matrix
+            mock_assign.assert_called_once()
+            args = mock_assign.call_args[0]
+            self.assertIs(args[0], self.atoms)
+            np.testing.assert_array_equal(args[1], self.structure.lattice.matrix)
     
     def test_assign_site_occupations_atom_in_site(self):
         """Test assign_site_occupations when atoms are already in sites."""
@@ -126,7 +129,7 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
             mock_contains_atom.return_value = True
             
             # Call method
-            self.collection.assign_site_occupations(self.atoms, self.structure)
+            self.collection.assign_site_occupations(self.atoms, self.lattice.matrix)
             
             # Verify site occupations were reset
             self.assertEqual(self.site1.contains_atoms, [])
@@ -154,7 +157,7 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
             mock_contains_atom.return_value = False
             
             # Call method
-            self.collection.assign_site_occupations(self.atoms, self.structure)
+            self.collection.assign_site_occupations(self.atoms, self.lattice.matrix)
             
             # Verify site occupations were reset
             self.assertEqual(self.site1.contains_atoms, [])
@@ -184,7 +187,7 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
             mock_contains_atom.side_effect = lambda atom, **kwargs: atom is self.atom1 and mock_contains_atom.mock_calls[0][1][0] is atom
             
             # Call method
-            self.collection.assign_site_occupations(self.atoms, self.structure)
+            self.collection.assign_site_occupations(self.atoms, self.lattice.matrix)
             
             # Verify site occupations were reset
             self.assertEqual(self.site1.contains_atoms, [])
@@ -217,7 +220,7 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
         )
         
         # Call the method with empty atom list
-        collection.assign_site_occupations([], structure)
+        collection.assign_site_occupations([], lattice.matrix)
         
         # Verify that contains_atoms was reset for both sites
         self.assertEqual(site1.contains_atoms, [])
@@ -311,7 +314,7 @@ class PolyhedralSiteCollectionTestCase(unittest.TestCase):
             patch.object(collection, 'site_by_index') as mock_site_by_index:
             mock_site_by_index.return_value = mock_site
 
-            collection.assign_site_occupations([mock_atom], mock_structure)
+            collection.assign_site_occupations([mock_atom], np.eye(3) * 10.0)
 
             mock_site_by_index.assert_called_with(5)
             mock_update.assert_called_with(mock_site, mock_atom)
@@ -501,7 +504,7 @@ class TestAssignSiteOccupationsInteraction(unittest.TestCase):
     def test_calls_generator_for_each_atom(self):
         """Test that _get_priority_sites is called once per atom."""
         with patch.object(self.collection, '_get_priority_sites', return_value=[]):
-            self.collection.assign_site_occupations(self.atoms, self.structure)
+            self.collection.assign_site_occupations(self.atoms, self.lattice.matrix)
             self.collection._get_priority_sites.assert_called_once_with(self.atom)
     
     def test_calls_generator_for_multiple_atoms(self):
[…]
