### 2026-03-21 00e0af4
Unify assign_site_occupations signature to accept lattice_matrix
Update the abstract SiteCollection.assign_site_occupations() and all
implementations to accept lattice_matrix (numpy array) instead of
Structure. Trajectory.assign_site_occupations() extracts
structure.lattice.matrix at the boundary before delegating.

PolyhedralSiteCollection accepts lattice_matrix for interface
consistency (not currently used for containment checks).

Partially addresses #56
