### 2026-03-21 c90febe
Replace lattice.get_all_distances with all_mic_distances in Voronoi collections (#64)
VoronoiSiteCollection and DynamicVoronoiSiteCollection now use
all_mic_distances() from the distances module instead of
lattice.get_all_distances(). assign_site_occupations() takes a
lattice_matrix (numpy array) instead of a Structure object.

Unifies the assign_site_occupations interface across all
SiteCollection implementations (including PolyhedralSiteCollection
and Trajectory) to accept lattice_matrix consistently.

Closes #54, partially addresses #56
