### 2026-03-21 ee3da8b
Replace Lattice.get_distance_and_image with mic_distance in SphericalSite (#65)
SphericalSite.contains_point() and contains_atom() now use
mic_distance() from the distances module instead of
Lattice.get_distance_and_image(). Both methods take a lattice_matrix
(numpy array) instead of a pymatgen Lattice object.

SphericalSiteCollection.assign_site_occupations() takes lattice_matrix
instead of Structure. Removes isinstance type checks that enforced
the Lattice type.

Closes #55
