+++ tests/test_spherical_site.py
@@ -79,28 +79,22 @@ class SphericalSiteTestCase(unittest.TestCase):
             ]
         )
 
-    def test_contains_atom_raises_valueerror_if_lattice_is_not_passed(self):
+    def test_contains_atom_raises_valueerror_if_lattice_matrix_is_not_passed(self):
         mock_atom = Mock(spec=Atom)
         with self.assertRaises(ValueError):
             self.spherical_site.contains_atom(atom=mock_atom)
-
-    def test_contains_atom_raises_typeerror_if_lattice_is_not_Lattice(self):
-        mock_atom = Mock(spec=Atom)
-        with self.assertRaises(TypeError):
-            self.spherical_site.contains_atom(atom=mock_atom,
-                    lattice="bar")
 
     def test_contains_atom(self):
         with patch('site_analysis.spherical_site.SphericalSite.contains_point') as mock_contains_point:
             mock_contains_point.return_value = "called"
             mock_atom = Mock(spec=Atom)
             mock_atom.frac_coords = np.array([2.0, 3.0, 4.0])
-            mock_lattice = Mock(spec=Lattice)
-            returned = self.spherical_site.contains_atom(atom=mock_atom, lattice=mock_lattice)
+            mock_lattice_matrix = np.eye(3) * 10.0
+            returned = self.spherical_site.contains_atom(atom=mock_atom, lattice_matrix=mock_lattice_matrix)
             self.assertEqual(returned, "called")
             mock_contains_point.assert_called_with(
                     x=mock_atom.frac_coords,
-                    lattice=mock_lattice)
+                    lattice_matrix=mock_lattice_matrix)
 
     def test_centre_property(self):
         """Test the centre property."""
@@ -117,14 +111,14 @@ class SphericalSiteTestCase(unittest.TestCase):
         # Point inside the sphere
         point_inside = np.array([0.55, 0.55, 0.55])
         self.assertTrue(
-            self.test_site.contains_point(point_inside, lattice=self.lattice),
+            self.test_site.contains_point(point_inside, lattice_matrix=self.lattice.matrix),
             "Point inside sphere should be contained"
         )
         
         # Point outside the sphere
         point_outside = np.array([0.8, 0.8, 0.8])
         self.assertFalse(
-            self.test_site.contains_point(point_outside, lattice=self.lattice),
+            self.test_site.contains_point(point_outside, lattice_matrix=self.lattice.matrix),
             "Point outside sphere should not be contained"
         )
         
@@ -132,51 +126,45 @@ class SphericalSiteTestCase(unittest.TestCase):
         point_on_boundary = np.array([0.5, 0.5, 0.7])
         # Due to floating point precision, this might be either True or False
         # We just need to make sure it doesn't raise an error
-        result = self.test_site.contains_point(point_on_boundary, lattice=self.lattice)
+        result = self.test_site.contains_point(point_on_boundary, lattice_matrix=self.lattice.matrix)
         self.assertIsInstance(result, bool)
         
         # Test a point very close to the boundary (just inside)
         point_just_inside = np.array([0.5, 0.5, 0.699])
         self.assertTrue(
-            self.test_site.contains_point(point_just_inside, lattice=self.lattice),
+            self.test_site.contains_point(point_just_inside, lattice_matrix=self.lattice.matrix),
             "Point just inside sphere should be contained"
         )
         
         # Test a point very close to the boundary (just outside)
         point_just_outside = np.array([0.5, 0.5, 0.701])
         self.assertFalse(
-            self.test_site.contains_point(point_just_outside, lattice=self.lattice),
+            self.test_site.contains_point(point_just_outside, lattice_matrix=self.lattice.matrix),
             "Point just outside sphere should not be contained"
         )
 
-    def test_contains_point_without_lattice(self):
[…]
