### 2026-03-21 72f4d33
Convert DynamicVoronoiSite.calculate_centre to accept arrays
calculate_centre() now takes (all_frac_coords, lattice_matrix) instead
of a Structure. Uses array indexing consistent with the batch path in
DynamicVoronoiSiteCollection._batch_calculate_centres().

Part of #56
