+++ tests/test_pbc_utils.py
@@ -160,7 +160,7 @@ class TestCurrentPBCBehaviorRegression(unittest.TestCase):
 		structure = Structure(self.lattice, species, coords)
 		
 		site = DynamicVoronoiSite(reference_indices=[0, 1, 2, 3])
-		site.calculate_centre(structure)
+		site.calculate_centre(structure.frac_coords, structure.lattice.matrix)
 		
 		# Should calculate centre correctly (this works correctly)
 		expected_centre = np.mean(coords, axis=0)
@@ -201,7 +201,7 @@ class TestCurrentPBCBehaviorRegression(unittest.TestCase):
 		structure = Structure(self.lattice, species, coords)
 		
 		site = DynamicVoronoiSite(reference_indices=[0, 1, 2, 3])
-		site.calculate_centre(structure)
+		site.calculate_centre(structure.frac_coords, structure.lattice.matrix)
 		
 		# DynamicVoronoiSite applies correction AND wraps centre back to [0,1)
 		corrected_coords = np.array([
@@ -244,7 +244,7 @@ class TestCurrentPBCBehaviorRegression(unittest.TestCase):
 		structure = Structure(self.lattice, species, coords)
 		
 		site = DynamicVoronoiSite(reference_indices=[0, 1, 2, 3])
-		site.calculate_centre(structure)
+		site.calculate_centre(structure.frac_coords, structure.lattice.matrix)
 		
 		# Should calculate centre correctly (no correction needed)
 		expected_centre = np.mean(coords, axis=0) % 1.0
