+++ site_analysis/dynamic_voronoi_site.py
@@ -10,7 +10,6 @@ import numpy as np
 from typing import Any
 from .site import Site
 from .atom import Atom
-from pymatgen.core import Structure
 from site_analysis.pbc_utils import correct_pbc
 
 class DynamicVoronoiSite(Site):
@@ -87,19 +86,19 @@ class DynamicVoronoiSite(Site):
         super(DynamicVoronoiSite, self).reset()
         self._centre_coords = None
         
-    def calculate_centre(self, structure: Structure) -> None:
-        """Calculate the centre of this site based on the positions of reference atoms.
+    def calculate_centre(self,
+            all_frac_coords: np.ndarray,
+            lattice_matrix: np.ndarray) -> None:
+        """Calculate the centre of this site from reference atom positions.
 
         Args:
-            structure: The pymatgen Structure used to assign
-                fractional coordinates to the reference atoms.
-
-        Notes:
-            This method handles periodic boundary conditions and calculates
-            the centre as the mean of the reference atom positions.
+            all_frac_coords: Full fractional coordinate array from the
+                structure, shape ``(n_atoms, 3)``.
+            lattice_matrix: (3, 3) lattice matrix where rows are lattice
+                vectors.
         """
-        ref_coords = np.array([structure[i].frac_coords for i in self.reference_indices])
-        corrected, _ = correct_pbc(ref_coords, self.reference_center, structure.lattice.matrix)
+        ref_coords = all_frac_coords[self.reference_indices]
+        corrected, _ = correct_pbc(ref_coords, self.reference_center, lattice_matrix)
         self._centre_coords = np.mean(corrected, axis=0) % 1.0
         
     @property
