### 2026-03-21 f8225a1
Convert PolyhedralSite.assign_vertex_coords and get_vertex_species to arrays
assign_vertex_coords() now takes (all_frac_coords, lattice_matrix)
instead of a Structure. get_vertex_species() takes a list of species
strings instead of a Structure. SiteFactory extracts arrays eagerly
in __init__ and passes them through.

Part of #56
