### 2026-03-21 855ded2
Remove stale type: ignore on scipy import
scipy-stubs is installed and provides types for scipy.spatial.
