### 2026-03-21 e93c43b
Remove structure parameter from PolyhedralSite.contains_point
The structure parameter was a legacy convenience that triggered
assign_vertex_coords internally. The normal path via
notify_structure_changed / _assign_from_pending is now the only way
to set vertex coordinates before calling contains_point.

Part of #56
