### 2026-03-21 1decbd6
Convert sites_contain_points to accept arrays
sites_contain_points() now takes (points, all_frac_coords,
lattice_matrix) instead of (points, structure). Removes the
isinstance(structure, Structure) type check. Updates base class
signature in SiteCollection.

Part of #56
