### 2026-03-21 a1e1000
Remove unused pymatgen imports from polyhedral and site collection modules
polyhedral_site.py no longer imports Structure or species_string_from_site.
site_collection.py no longer imports Structure.
Removes stale type: ignore comments on pymatgen imports.

Closes #56
