### 2026-03-21 802d384
Clean up stale imports and docstrings from decoupling refactor
Remove unused Structure import from site.py. Update stale module
docstring in dynamic_voronoi_site_collection.py and stale
analyse_structure docstrings in dynamic_voronoi_site_collection.py
and spherical_site_collection.py.
