### 2026-03-21 def018a
Address PR review feedback
Add length validation to sites_contain_points to prevent silent
truncation via zip. Add test verifying notify_structure_changed is
called with correct arrays. Add test for length mismatch error.

Restore lost docstring notes: Delaunay reset in assign_vertex_coords,
PBC handling in calculate_centre. Add docstring to
PolyhedralSiteCollection.analyse_structure. Fix trailing newline in
site_factory.py.
