### 2026-03-21 b802475
Reject legacy structure kwarg in PolyhedralSite.contains_point
Raises TypeError with a clear message if callers pass the removed
structure keyword argument, instead of silently ignoring it via
**kwargs.
