### 2026-03-21 36dd052
Use concrete signatures for contains_point and contains_atom
Remove *args/**kwargs from Site.contains_point and
Site.contains_atom base class signatures. Each subclass now
declares its own keyword-only parameters explicitly, so invalid
arguments raise TypeError immediately rather than being silently
swallowed.

Use typing.cast for type narrowing where _get_priority_sites
yields the base Site type but the collection knows the concrete
type.
