### 2026-03-21 19f60ec
Make PriorityAssignmentMixin generic over site type
PriorityAssignmentMixin[SiteT] now yields the concrete site type
from _get_priority_sites, so collections get back SphericalSite or
PolyhedralSite directly without needing cast() calls.
