+++ site_analysis/spherical_site_collection.py
@@ -17,15 +17,13 @@ persistence can be useful for tracking atoms through small oscillations
 without generating spurious site transitions.
 """
 
-from typing import cast
-
 import numpy as np
 from pymatgen.core import Structure
 from site_analysis.atom import Atom
 from site_analysis.spherical_site import SphericalSite
 from site_analysis.site_collection import SiteCollection, PriorityAssignmentMixin
 
-class SphericalSiteCollection(PriorityAssignmentMixin, SiteCollection):
+class SphericalSiteCollection(PriorityAssignmentMixin[SphericalSite], SiteCollection):
 
 
     def __init__(self,
@@ -87,8 +85,7 @@ class SphericalSiteCollection(PriorityAssignmentMixin, SiteCollection):
             atom.in_site = None
 
             # Check sites in priority order until found
-            for site_ in self._get_priority_sites(atom):
-                site = cast(SphericalSite, site_)
+            for site in self._get_priority_sites(atom):
                 if site.contains_atom(atom, lattice_matrix=lattice_matrix):
                     self.update_occupation(site, atom)
                     break
