### 2026-03-21 1a9fd4a
Remove Structure acceptance from polyhedral and dynamic voronoi site methods (#66)
Convert remaining site methods to accept numpy arrays instead of
pymatgen Structure objects:

- DynamicVoronoiSite.calculate_centre() takes (all_frac_coords,
  lattice_matrix)
- PolyhedralSite.assign_vertex_coords() takes (all_frac_coords,
  lattice_matrix)
- PolyhedralSite.get_vertex_species() takes a species list
- PolyhedralSite.contains_point() no longer accepts structure
- PolyhedralSiteCollection.sites_contain_points() takes arrays
- SiteFactory extracts arrays eagerly in __init__

Also removes *args/**kwa
[…]
