### 2026-03-21 c6e35bf
Decouple calculate_species_distances from pymatgen Structure
calculate_species_distances() now accepts arrays (frac_coords,
lattice_matrix, species lists) instead of Structure objects. Add
indices_for_species() helper replacing structure.indices_from_symbol().
StructureAligner updated to extract arrays at the boundary.

Part of #57
