+++ tests/test_tools.py
@@ -594,171 +594,205 @@ class GetCoordinationIndicesTestCase(unittest.TestCase):
         
     def test_calculate_species_distances_identical_structures(self):
         """Test distance calculation between identical structures."""
-        # Create a simple structure
-        lattice = Lattice.cubic(5.0)
         species = ["Na", "Cl", "Na", "Cl"]
-        coords = [
+        coords = np.array([
             [0.1, 0.1, 0.1],
             [0.6, 0.6, 0.6],
             [0.3, 0.3, 0.3],
             [0.8, 0.8, 0.8]
-        ]
-        structure1 = Structure(lattice, species, coords)
-        structure2 = Structure(lattice, species, coords)
-        
+        ])
+        lattice_matrix = np.eye(3) * 5.0
+
         # Calculate distances
-        species_distances, all_distances = calculate_species_distances(structure1, structure2)
-        
+        species_distances, all_distances = calculate_species_distances(
+            coords, coords, lattice_matrix, species, species)
+
         # Check results
         self.assertEqual(len(species_distances), 2)  # Na and Cl
         self.assertIn("Na", species_distances)
         self.assertIn("Cl", species_distances)
-        
+
         # Distances should be zero for identical structures
         self.assertEqual(len(species_distances["Na"]), 2)  # Two Na atoms
         self.assertEqual(len(species_distances["Cl"]), 2)  # Two Cl atoms
         for dist in all_distances:
             self.assertAlmostEqual(dist, 0.0, places=5)
-    
+
     def test_calculate_species_distances_translated_structures(self):
         """Test distance calculation between translated structures."""
-        # Create a simple structure
-        lattice = Lattice.cubic(5.0)
         species = ["Na", "Cl", "Na", "Cl"]
-        coords1 = [
+        coords1 = np.array([
             [0.1, 0.1, 0.1],
             [0.6, 0.6, 0.6],
             [0.3, 0.3, 0.3],
             [0.8, 0.8, 0.8]
-        ]
-        structure1 = Structure(lattice, species, coords1)
-        
+        ])
+
         # Create a translated version (shift by [0.1, 0.1, 0.1])
-        coords2 = [
+        coords2 = np.array([
             [0.2, 0.2, 0.2],  # +0.1 in each direction
             [0.7, 0.7, 0.7],
             [0.4, 0.4, 0.4],
             [0.9, 0.9, 0.9]
-        ]
-        structure2 = Structure(lattice, species, coords2)
-        
+        ])
+        lattice_matrix = np.eye(3) * 5.0
+
         # Calculate distances
-        species_distances, all_distances = calculate_species_distances(structure1, structure2)
-        
-        # Expected distance for a [0.1, 0.1, 0.1] shift in a 5.0 Å unit cell
+        species_distances, all_distances = calculate_species_distances(
+            coords1, coords2, lattice_matrix, species, species)
+
+        # Expected distance for a [0.1, 0.1, 0.1] shift in a 5.0 A unit cell
         expected_dist = 5.0 * np.sqrt(3 * (0.1**2))
-        
+
         # Check results
         for sp, distances in species_distances.items():
             for dist in distances:
                 self.assertAlmostEqual(dist, expected_dist, places=5)
-    
+
     def test_calculate_species_distances_different_compositions(self):
         """Test distance calculation between structures with different compositions."""
-        # Create structures with different compositions
-        lattice = Lattice.cubic(5.0)
-        
+        lattice_matrix = np.eye(3) * 5.0
+
         # Structure 1: Na2Cl2
         species1 = ["Na", "Na", "Cl", "Cl"]
-        coords1 = [
+        coords1 = np.array([
[…]
