### 2026-03-21 e39fead
Decouple site_index_mapping from pymatgen Structure
site_index_mapping() now accepts arrays instead of Structure objects.
Uses all_mic_distances instead of lattice.get_all_distances.

Part of #57
