### 2026-03-21 ea24487
Replace structure.get_neighbors with all_mic_distances in get_coordination_indices
get_coordination_indices() now accepts arrays (frac_coords, lattice_matrix,
species) instead of a Structure. Computes the full centre-to-coordination
distance matrix and filters by cutoff, replacing the per-site
get_neighbors() call. CoordinationEnvironmentFinder extracts arrays in
__init__.

Tests updated to pass arrays directly and to account for minimum-image
semantics (each atom pair yields one distance, no periodic-image
duplication).

Part of #57
