### 2026-03-21 5657714
Decouple IndexMapper from pymatgen Structure
map_coordinating_atoms() accepts arrays instead of Structure objects.
Uses all_mic_distances instead of lattice.get_all_distances.
ReferenceBasedSites._map_environments updated to pass arrays.

Part of #57
