+++ tests/test_reference_workflow/test_reference_based_sites.py
@@ -70,12 +70,22 @@ class TestReferenceBasedSites(unittest.TestCase):
         """Set up a mock StructureAligner."""
         mock = Mock()
         aligned_reference = Mock(spec=Structure)
+        # Give aligned reference the array attributes needed by _map_environments
+        aligned_reference.frac_coords = np.array([
+            [0.05, 0.05, 0.05],
+            [0.55, 0.55, 0.55],
+            [0.55, 0.05, 0.05],
+            [0.05, 0.55, 0.55],
+        ])
+        aligned_lattice = Mock()
+        aligned_lattice.matrix = np.diag([5.0, 5.0, 5.0])
+        aligned_reference.lattice = aligned_lattice
         translation_vector = np.array([0.1, 0.1, 0.1])
         metrics = {'rmsd': 0.1, 'max_dist': 0.2, 'mean_dist': 0.15}
-        
+
         # Mock align method to return the aligned reference structure
         mock.align.return_value = (aligned_reference, translation_vector, metrics)
-        
+
         return mock
     
     def _setup_mock_coord_finder(self):
@@ -205,12 +215,13 @@ class TestReferenceBasedSites(unittest.TestCase):
             )
             
             # Check that map_coordinating_atoms was called with correct parameters
-            self.mock_index_mapper.map_coordinating_atoms.assert_called_with(
-                self.reference,
-                self.target,  # Should use target directly since no alignment
-                self.ref_environments,
-                target_species='Cl'
-            )
+            args, kwargs = self.mock_index_mapper.map_coordinating_atoms.call_args
+            np.testing.assert_array_equal(kwargs['ref_frac_coords'], self.reference.frac_coords)
+            np.testing.assert_array_equal(kwargs['target_frac_coords'], self.target.frac_coords)
+            np.testing.assert_array_equal(kwargs['lattice_matrix'], self.reference.lattice.matrix)
+            self.assertEqual(kwargs['ref_coordinating'], self.ref_environments)
+            self.assertEqual(kwargs['target_species'], [s.species_string for s in self.target])
+            self.assertEqual(kwargs['species_filter'], 'Cl')
             
             # Check returned environments
             self.assertEqual(mapped, self.mapped_environments)
@@ -232,16 +243,14 @@ class TestReferenceBasedSites(unittest.TestCase):
             
             # Check map_coordinating_atoms call - use individual argument checks
             args, kwargs = self.mock_index_mapper.map_coordinating_atoms.call_args
-            
-            # First arg should be the aligned reference
-            self.assertIs(args[0], aligned_reference)
-            
-            # Second arg should be the target
-            self.assertIs(args[1], self.target)
-            
-            # Verify other arguments
-            self.assertEqual(args[2], self.ref_environments)
-            self.assertEqual(kwargs['target_species'], 'Cl')
+
+            # Should use aligned reference coords and lattice
+            np.testing.assert_array_equal(kwargs['ref_frac_coords'], aligned_reference.frac_coords)
+            np.testing.assert_array_equal(kwargs['target_frac_coords'], self.target.frac_coords)
+            np.testing.assert_array_equal(kwargs['lattice_matrix'], aligned_reference.lattice.matrix)
+            self.assertEqual(kwargs['ref_coordinating'], self.ref_environments)
+            self.assertEqual(kwargs['target_species'], [s.species_string for s in self.target])
+            self.assertEqual(kwargs['species_filter'], 'Cl')
             
             # Check returned environments
             self.assertEqual(mapped, self.mapped_environments)
@@ -277,14 +286,15 @@ class TestReferenceBasedSites(unittest.TestCase):
                 # Check that _calculate_reference_centers_from_indices was called
[…]
