+++ site_analysis/reference_workflow/index_mapper.py
@@ -20,68 +20,77 @@ of site definitions between different structures or timesteps in a simulation.
 
 import numpy as np
 
-from pymatgen.core import Structure
+from site_analysis.distances import all_mic_distances
 
 
 class IndexMapper:
     """Maps coordinating atom indices between reference and target crystal structures.
-    
+
     Used to translate coordination environments defined in an ideal reference
     structure to corresponding atoms in a target structure, handling permutations
     and structural distortions via distance-based matching.
-    
+
     The mapper verifies 1:1 correspondence between reference and target atoms,
     ensuring each coordinating position maps to exactly one target atom. If
     this constraint is violated, a ValueError is raised.
     """
-    
+
     def map_coordinating_atoms(
         self,
-        reference: Structure,
-        target: Structure,
+        ref_frac_coords: np.ndarray,
+        target_frac_coords: np.ndarray,
+        lattice_matrix: np.ndarray,
         ref_coordinating: list[list[int]],
-        target_species: str | list[str] | None = None
+        target_species: list[str] | None = None,
+        species_filter: str | list[str] | None = None,
     ) -> list[list[int]]:
         """Map coordinating atom indices from reference to target structure.
-        
+
         Args:
-            reference: Reference structure containing ideal coordination environments.
-            target: target structure with potentially distorted or permuted atoms.
+            ref_frac_coords: Fractional coordinates of the reference structure,
+                shape (N, 3).
+            target_frac_coords: Fractional coordinates of the target structure,
+                shape (M, 3).
+            lattice_matrix: Lattice matrix (3x3) for distance calculations.
             ref_coordinating: List of coordinating atom index lists from reference.
                 Each sublist contains indices of atoms that define a site.
-            target_species: Optional filter for target atom species to map to.
-                If specified, only maps to atoms of these species in the target structure.
-                
+            target_species: Full species list for all atoms in the target structure.
+                Required when using species_filter.
+            species_filter: Optional filter for which species to map to.
+                If specified, only maps to atoms of these species in the target
+                structure.
+
         Returns:
             List of coordinating atom index lists mapped to the target structure.
             Maintains the same structure as input but with updated indices.
-            
+
         Raises:
             ValueError: If 1:1 mapping cannot be achieved (e.g., missing atoms,
                 ambiguous distances, or insufficient target atoms in target structure).
         """
         # Extract all unique coordinating atoms from reference structure
         unique_indices = self._extract_unique_coordinating_atoms(ref_coordinating)
-        
+
         # Create mapping from reference to target structure
         index_mapping = self._find_closest_atom_mapping(
-            reference, target, unique_indices, target_species
+            ref_frac_coords, target_frac_coords, lattice_matrix,
+            unique_indices, target_species, species_filter,
         )
-        
+
         # Map the coordination lists using the established mapping
         mapped_coordinating = self._apply_mapping(ref_coordinating, index_mapping)
-        
+
         return mapped_coordinating
-    
+
     def _extract_unique_coordinating_atoms(
         self,
         ref_coordinating: list[list[int]]
     ) -> list[int]:
         """Extract unique coordinating atom indices from coordination lists.
-        
+
         Args:
             ref_coordinating: List of coordinating atom index lists.
-            
+
         Returns:
             Sorted list of unique coordinating atom indices.
         """
@@ -89,54 +98,66 @@ class IndexMapper:
         for coord_list in ref_coordinating:
             unique_indices.update(coord_list)
         return sorted(list(unique_indices))
-    
+
     def _find_closest_atom_mapping(
         self,
-        reference: Structure,
-        target: Structure,
+        ref_frac_coords: np.ndarray,
[…]
