### 2026-03-21 4ddb972
Eliminate temp-Structure creation in StructureAligner optimisation loop
_create_objective_function now works with pure numpy arrays.
_validate_structures accepts species lists instead of Structure
composition. The optimisation loop no longer calls structure.copy()
or performs per-site assignment on every iteration.

Part of #57
