+++ site_analysis/reference_workflow/structure_aligner.py
@@ -40,74 +40,76 @@ class StructureAligner:
         algorithm: str = 'Nelder-Mead',
         minimizer_options: dict[str, Any] | None = None) -> tuple[Structure, np.ndarray, dict[str, float]]:
         """Align reference structure to target structure via translation."""
+        # Extract arrays from Structure at the public boundary
+        ref_frac_coords = reference.frac_coords
+        target_frac_coords = target.frac_coords
+        lattice_matrix = reference.lattice.matrix
+        ref_species = [s.species_string for s in reference]
+        target_species_list = [s.species_string for s in target]
+
         # Validate structures and get species to use
-        valid_species = self._validate_structures(reference, target, species)
-        
+        valid_species = self._validate_structures(
+            ref_species, target_species_list, species)
+
         # Create objective function
-        objective_function = self._create_objective_function(reference, target, valid_species, metric)
-        
-        # Run the appropriate optimizer using the dispatcher
-        translation_vector = self._run_minimizer(algorithm, objective_function, tolerance, minimizer_options)
-        
+        objective_function = self._create_objective_function(
+            ref_frac_coords, target_frac_coords, lattice_matrix,
+            ref_species, target_species_list, valid_species, metric)
+
+        # Run the appropriate optimiser using the dispatcher
+        translation_vector = self._run_minimizer(
+            algorithm, objective_function, tolerance, minimizer_options)
+
         # Apply the translation to get the aligned structure
         aligned_structure = self._apply_translation(reference, translation_vector)
-        
-        # Calculate final metrics
-        aligned_species = [s.species_string for s in aligned_structure]
-        target_species = [s.species_string for s in target]
+
+        # Calculate final metrics using arrays
+        aligned_coords = (ref_frac_coords + translation_vector) % 1.0
         species_distances, all_distances = calculate_species_distances(
-            aligned_structure.frac_coords, target.frac_coords,
-            aligned_structure.lattice.matrix,
-            aligned_species, target_species, species=valid_species)
-        
+            aligned_coords, target_frac_coords, lattice_matrix,
+            ref_species, target_species_list, species=valid_species)
+
         metrics = {
             'rmsd': float(np.sqrt(np.mean(np.array(all_distances)**2))) if all_distances else float('inf'),
             'max_dist': float(np.max(all_distances)) if all_distances else float('inf'),
             'mean_dist': float(np.mean(all_distances)) if all_distances else float('inf'),
         }
-        
+
         return aligned_structure, translation_vector, metrics
         
-    def _create_objective_function(self, 
-        reference: Structure,
-        target: Structure,
+    def _create_objective_function(self,
+        ref_frac_coords: np.ndarray,
+        target_frac_coords: np.ndarray,
+        lattice_matrix: np.ndarray,
+        ref_species: list[str],
+        target_species: list[str],
         valid_species: list[str],
         metric: str) -> Callable[[np.ndarray], float]:
-        """Create the objective function for optimization.
-        
-        Args:
-            reference: Reference structure
-            target: Target structure
-            valid_species: List of species to include in alignment
-            metric: Metric to optimize ('rmsd' or 'max_dist')
-            
-        Returns:
-            function: The objective function that takes a translation vector and
-                    returns the distance metric value
-        """
-        # Extract arrays that don't change per iteration
-        ref_frac_coords = reference.frac_coords
-        ref_species = [s.species_string for s in reference]
-        target_frac_coords = target.frac_coords
-        target_species = [s.species_string for s in target]
-        lattice_matrix = reference.lattice.matrix
+        """Create the objective function for optimisation.
 
+        Args:
+            ref_frac_coords: Fractional coordinates of the reference structure.
[…]
