+++ tests/test_reference_workflow/test_structure_aligner.py
@@ -345,184 +345,100 @@ class TestStructureAligner(unittest.TestCase):
         
     def test_validate_structures(self):
         """Test the _validate_structures method with various scenarios."""
-        # Create test structures with different compositions
-        lattice = Lattice.cubic(5.0)
-        
-        # Structure A: Na2Cl2
-        species_a = ["Na", "Na", "Cl", "Cl"]
-        coords_a = [
-            [0.1, 0.1, 0.1],
-            [0.3, 0.3, 0.3],
-            [0.5, 0.5, 0.5],
-            [0.7, 0.7, 0.7]
-        ]
-        structure_a = Structure(lattice, species_a, coords_a)
-        
-        # Structure B: Na2Cl2 with different atom ordering
-        species_b = ["Cl", "Na", "Cl", "Na"]
-        coords_b = [
-            [0.5, 0.5, 0.5],
-            [0.3, 0.3, 0.3],
-            [0.7, 0.7, 0.7],
-            [0.1, 0.1, 0.1]
-        ]
-        structure_b = Structure(lattice, species_b, coords_b)
-        
-        # Structure C: Na3Cl2 (different Na count)
-        species_c = ["Na", "Na", "Na", "Cl", "Cl"]
-        coords_c = [
-            [0.1, 0.1, 0.1],
-            [0.3, 0.3, 0.3],
-            [0.4, 0.4, 0.4],
-            [0.5, 0.5, 0.5],
-            [0.7, 0.7, 0.7]
-        ]
-        structure_c = Structure(lattice, species_c, coords_c)
-        
-        # Structure D: Na2Cl2F1 (extra species)
-        species_d = ["Na", "Na", "Cl", "Cl", "F"]
-        coords_d = [
-            [0.1, 0.1, 0.1],
-            [0.3, 0.3, 0.3],
-            [0.5, 0.5, 0.5],
-            [0.7, 0.7, 0.7],
-            [0.9, 0.9, 0.9]
-        ]
-        structure_d = Structure(lattice, species_d, coords_d)
-        
-        # Create StructureAligner
+        # Species lists for different compositions
+        species_a = ["Na", "Na", "Cl", "Cl"]       # Na2Cl2
+        species_b = ["Cl", "Na", "Cl", "Na"]       # Na2Cl2 (different ordering)
+        species_c = ["Na", "Na", "Na", "Cl", "Cl"] # Na3Cl2
+        species_d = ["Na", "Na", "Cl", "Cl", "F"]  # Na2Cl2F1
+
         aligner = StructureAligner()
-        
-        # Case 1: Both structures have identical composition, species=None
-        species_list = aligner._validate_structures(structure_a, structure_b, None)
-        self.assertCountEqual(species_list, ["Na", "Cl"])
-        
-        # Case 2: Both structures have different composition, species=None
+
+        # Case 1: Both have identical composition, species=None
+        species_list = aligner._validate_structures(species_a, species_b, None)
+        self.assertCountEqual(species_list, ["Cl", "Na"])
+
+        # Case 2: Different composition, species=None
         with self.assertRaises(ValueError) as context:
-            aligner._validate_structures(structure_a, structure_c, None)
+            aligner._validate_structures(species_a, species_c, None)
         self.assertIn("different compositions", str(context.exception))
-        
+
         with self.assertRaises(ValueError) as context:
-            aligner._validate_structures(structure_a, structure_d, None)
+            aligner._validate_structures(species_a, species_d, None)
         self.assertIn("different compositions", str(context.exception))
-        
-        # Case 3: Species is explicitly specified and exists in both structures
-        species_list = aligner._validate_structures(structure_a, structure_b, ["Na"])
+
+        # Case 3: Species explicitly specified and exists in both
[…]
