### 2026-03-21 1254d97
Extract arrays eagerly in ReferenceBasedSites constructor
ReferenceBasedSites now stores frac_coords, lattice_matrix, and
species lists from both structures at construction time. Internal
methods use these stored arrays instead of accessing Structure
objects directly.

Part of #57
