+++ tests/test_reference_workflow/test_reference_based_sites.py
@@ -127,29 +127,45 @@ class TestReferenceBasedSites(unittest.TestCase):
             translation_vector = np.array([0.1, 0.1, 0.1])
             metrics = {'rmsd': 0.1, 'max_dist': 0.2, 'mean_dist': 0.15}
             self.mock_structure_aligner.align.return_value = (aligned_reference, translation_vector, metrics)
-            
+
             # Initialize with alignment
             rbs = ReferenceBasedSites(self.reference, self.target, align=True)
-            
-            # Check attributes
+
+            # Check public attributes
             self.assertEqual(rbs.reference_structure, self.reference)
             self.assertEqual(rbs.target_structure, self.target)
             self.assertEqual(rbs.aligned_structure, aligned_reference)
             np.testing.assert_array_equal(rbs.translation_vector, translation_vector)
             self.assertEqual(rbs.alignment_metrics, metrics)
-        
+
+            # Check eagerly extracted arrays
+            np.testing.assert_array_equal(rbs._ref_frac_coords, self.reference.frac_coords)
+            np.testing.assert_array_equal(rbs._ref_lattice_matrix, self.reference.lattice.matrix)
+            self.assertEqual(rbs._ref_species, [s.species_string for s in self.reference])
+            np.testing.assert_array_equal(rbs._target_frac_coords, self.target.frac_coords)
+            np.testing.assert_array_equal(rbs._target_lattice_matrix, self.target.lattice.matrix)
+            self.assertEqual(rbs._target_species, [s.species_string for s in self.target])
+            np.testing.assert_array_equal(rbs._aligned_frac_coords, aligned_reference.frac_coords)
+
     def test_init_without_alignment(self):
         """Test initialization without structure alignment."""
         with self.structure_aligner_patch:
             # Initialize without alignment
             rbs = ReferenceBasedSites(self.reference, self.target, align=False)
-            
-            # Check attributes
+
+            # Check public attributes
             self.assertEqual(rbs.reference_structure, self.reference)
             self.assertEqual(rbs.target_structure, self.target)
             self.assertIsNone(rbs.aligned_structure)
             self.assertIsNone(rbs.translation_vector)
             self.assertIsNone(rbs.alignment_metrics)
+
+            # Check eagerly extracted arrays
+            np.testing.assert_array_equal(rbs._ref_frac_coords, self.reference.frac_coords)
+            np.testing.assert_array_equal(rbs._ref_lattice_matrix, self.reference.lattice.matrix)
+            self.assertEqual(rbs._ref_species, [s.species_string for s in self.reference])
+            np.testing.assert_array_equal(rbs._target_frac_coords, self.target.frac_coords)
+            self.assertIsNone(rbs._aligned_frac_coords)
     
     def test_align_structures(self):
         """Test the _align_structures method."""
@@ -207,22 +223,22 @@ class TestReferenceBasedSites(unittest.TestCase):
         """Test the _map_environments method."""
         with self.index_mapper_patch:
             rbs = ReferenceBasedSites(self.reference, self.target, align=False)
-            
+
             # Map environments
             mapped = rbs._map_environments(
                 ref_environments=self.ref_environments,
                 target_species='Cl'
             )
-            
-            # Check that map_coordinating_atoms was called with correct parameters
+
+            # Check that map_coordinating_atoms was called with stored arrays
             args, kwargs = self.mock_index_mapper.map_coordinating_atoms.call_args
-            np.testing.assert_array_equal(kwargs['ref_frac_coords'], self.reference.frac_coords)
-            np.testing.assert_array_equal(kwargs['target_frac_coords'], self.target.frac_coords)
[…]
