### 2026-03-21 59e4b3b
Clean up dead code and redundant copies from /simplify review
Remove dead _translate_coords method and duplicate scipy.optimize
import from StructureAligner. Remove unnecessary .copy() on arrays
already returned as copies by pymatgen. Extract _effective_ref_coords
property to avoid repeating the aligned/reference conditional.
