### 2026-03-21 cc49e86
Address PR review feedback
Fix critical issues:
- Fix wrong deprecation target: get_coordinating_indices ->
  get_coordination_indices
- Fix incorrect aligned_structure docstring: reference is translated
  to match target, None when align=False

Fix important issues:
- Add species/frac_coords length validation at public boundaries
  for get_coordination_indices, calculate_species_distances,
  site_index_mapping, and IndexMapper.map_coordinating_atoms
- Narrow except Exception to except ValueError in
  ReferenceBasedSites to avoid masking programming errors
- Add trailing newlines to tools.py a
[…]
