### 2026-03-21 90b8cd7
Address Copilot review comments
Fix _validate_structures docstring (not always sorted). Update
error message in get_coordination_indices to not reference Structure.
Handle empty ref_indices in IndexMapper._find_closest_atom_mapping.
Add validation for empty species2_filter match in site_index_mapping.
