### 2026-03-22 ca164bb
Fix species filter tests to actually pass strings
The _as_string test variants were passing lists, not exercising
the string normalisation code path.
