### 2026-03-22 8cb80d1
Decouple reference workflow from pymatgen internals (#67)
Remove pymatgen Structure/Lattice usage from all reference workflow
internals. Structure objects are now only accepted at public API
boundaries and arrays are extracted eagerly.

- calculate_species_distances() and site_index_mapping() accept
  arrays and species lists instead of Structure objects
- get_coordination_indices() uses all_mic_distances instead of
  structure.get_neighbors()
- IndexMapper.map_coordinating_atoms() accepts arrays
- StructureAligner optimisation loop no longer creates temp-Structure
  objects per iteration
- Co
[…]
