+++ site_analysis/reference_workflow/coord_finder.py
@@ -13,8 +13,8 @@ enabling the creation of corresponding sites.
 Key features include:
 - Finding atoms with exactly the specified coordination environment
 - Supporting both single and multiple coordinating species
-- Filtering by center atom species and coordinating atom species
-- Customizing coordination requirements per center atom
+- Filtering by centre atom species and coordinating atom species
+- Customising coordination requirements per centre atom
 
 This module is a key component of the reference-based workflow, providing the
 initial identification of coordination environments that will be used to define
@@ -27,78 +27,80 @@ from site_analysis.tools import get_coordination_indices
 
 class CoordinationEnvironmentFinder:
     """Finds coordination environments in a structure."""
-    
+
     def __init__(self, structure: Structure):
-        """Initialize a CoordinationEnvironmentFinder for a specific structure.
-        
-        Creates a new finder that will analyze the given structure to identify
-        coordination environments.
-        
+        """Initialise with a structure. Extracts arrays for internal use.
+
         Args:
-            structure: The pymatgen Structure to analyze
+            structure: The pymatgen Structure to analyse.
         """
-        self.structure = structure
+        self._frac_coords = structure.frac_coords
+        self._lattice_matrix = structure.lattice.matrix
+        self._species = [site.species_string for site in structure]
         self._atom_indices = self._index_atoms_by_species()
-    
+
     def _index_atoms_by_species(self) -> dict[str, list[int]]:
         """Create a mapping of species to atom indices.
-        
-        Indexes all atoms in the structure by their species string,
-        creating a dictionary where keys are species strings and
-        values are lists of atom indices having that species.
-        
+
+        Indexes all atoms by their species string, creating a dictionary
+        where keys are species strings and values are lists of atom indices
+        having that species.
+
         Returns:
-            A dictionary mapping species strings to lists of atom indices
+            A dictionary mapping species strings to lists of atom indices.
         """
         indices: dict[str, list[int]] = {}
-        for i, site in enumerate(self.structure):
-            species = site.species_string
-            if species not in indices:
-                indices[species] = []
-            indices[species].append(i)
+        for i, sp in enumerate(self._species):
+            if sp not in indices:
+                indices[sp] = []
+            indices[sp].append(i)
         return indices
-    
-    def find_environments(self, 
-                          center_species: str, 
+
+    def find_environments(self,
+                          center_species: str,
                           coordination_species: str | list[str],
-                          n_coord: int, 
+                          n_coord: int,
                           cutoff: float) -> dict[int, list[int]]:
         """Find coordination environments in the structure.
-        
-        Locates atoms of center_species and finds environments where these atoms
-        are coordinated by exactly n_coord atoms of coordination_species within
-        the specified cutoff distance.
-        
+
+        Locates atoms of center_species and finds environments where these
+        atoms are coordinated by exactly n_coord atoms of
+        coordination_species within the specified cutoff distance.
+
         Args:
-            center_species: Species at the center of the coordination environment
-            coordination_species: Species of the coordinating atoms (can be a string
-                or a list of strings)
-            n_coord: Number of coordinating atoms required for each environment
-            cutoff: Maximum distance (in Ångströms) for atoms to be considered coordinating
-            
+            center_species: Species at the centre of the coordination
+                environment.
+            coordination_species: Species of the coordinating atoms (can
+                be a string or a list of strings).
+            n_coord: Number of coordinating atoms required for each
+                environment.
+            cutoff: Maximum distance (in Angstroms) for atoms to be
+                considered coordinating.
+
         Returns:
-            A dictionary mapping center atom indices to lists of coordinating atom indices
-            
+            A dictionary mapping centre atom indices to lists of
+            coordinating atom indices.
+
         Raises:
[…]
