+++ tests/test_tools.py
@@ -73,168 +73,140 @@ class ToolsTestCase(unittest.TestCase):
  
     def test_site_index_mapping_one(self):
         a = 6.19399
-        lattice = Lattice.from_parameters(a, a, a, 90, 90, 90)
+        lattice_matrix = np.eye(3) * a
         coords1 = np.array([[0.0, 0.0, 0.0],
                             [0.5, 0.5, 0.5]])
         coords2 = np.array([[0.1, 0.1, 0.1],
                             [0.4, 0.6, 0.4]])
-        sites1 = [PeriodicSite(species='Na', coords=c, lattice=lattice) for c in coords1]
-        sites2 = [PeriodicSite(species='Na', coords=c, lattice=lattice) for c in coords2]
-        structure1 = Structure.from_sites(sites1)
-        structure2 = Structure.from_sites(sites2)
-        mapping = site_index_mapping(structure1, structure2)
+        species = ['Na', 'Na']
+        mapping = site_index_mapping(coords1, coords2, lattice_matrix, species, species)
         np.testing.assert_array_equal(mapping, np.array([0, 1]))
 
     def test_site_index_mapping_two(self):
         a = 6.19399
-        lattice = Lattice.from_parameters(a, a, a, 90, 90, 90)
+        lattice_matrix = np.eye(3) * a
         coords1 = np.array([[0.0, 0.0, 0.0],
                             [0.5, 0.5, 0.5]])
         coords2 = np.array([[0.4, 0.6, 0.4],
                             [0.1, 0.1, 0.1]])
-        sites1 = [PeriodicSite(species='Na', coords=c, lattice=lattice) for c in coords1]
-        sites2 = [PeriodicSite(species='Na', coords=c, lattice=lattice) for c in coords2]
-        structure1 = Structure.from_sites(sites1)
-        structure2 = Structure.from_sites(sites2)
-        mapping = site_index_mapping(structure1, structure2)
+        species = ['Na', 'Na']
+        mapping = site_index_mapping(coords1, coords2, lattice_matrix, species, species)
         np.testing.assert_array_equal(mapping, np.array([1, 0]))
-        mapping = site_index_mapping(structure1, structure2)
+        mapping = site_index_mapping(coords1, coords2, lattice_matrix, species, species)
         np.testing.assert_array_equal(mapping, np.array([1, 0]))
-        
-    def test_site_index_mapping_with_species_1_as_string(self):
+
+    def test_site_index_mapping_with_species1_filter_as_string(self):
         a = 6.19399
-        lattice = Lattice.from_parameters(a, a, a, 90, 90, 90)
+        lattice_matrix = np.eye(3) * a
         coords1 = np.array([[0.0, 0.0, 0.0],
                             [0.5, 0.5, 0.5]])
         coords2 = np.array([[0.4, 0.6, 0.4],
                             [0.1, 0.1, 0.1]])
         species1 = ['Na', 'Cl']
-        sites1 = [PeriodicSite(species=s, coords=c, lattice=lattice) for s, c in zip(species1, coords1)]
-        sites2 = [PeriodicSite(species='Na', coords=c, lattice=lattice) for c in coords2]
-        structure1 = Structure.from_sites(sites1)
-        structure2 = Structure.from_sites(sites2)
-        mapping = site_index_mapping(structure1, structure2, species1='Na')
+        species2 = ['Na', 'Na']
+        mapping = site_index_mapping(coords1, coords2, lattice_matrix, species1, species2, species1_filter='Na')
         np.testing.assert_array_equal(mapping, np.array([1]))
-        
[…]

[…]
