### 2026-03-22 09f7060
Fix type 2 site Wyckoff position in argyrodite reference structure
The Mg coordinate [0.23, 0.92, 0.09] sat on the 96i general
position in F-43m, generating pairs of atoms 0.14 A apart inside
the same S tetrahedron. This produced 384 duplicate type 2 sites
with identical vertex/reference atoms.

Replace with [0.77, 0.585, 0.585] which sits on the 48h special
position, giving the correct 384 unique type 2 sites with no
duplicates.

Closes #68
