+++ docs/scripts/generate_notebooks.py
@@ -441,7 +441,7 @@ def generate_argyrodite_site_analysis():
         "coords = np.array(\n"
         "    [[0.5,     0.5,     0.5],     # P (type 0) - PS4 tetrahedra\n"
         "     [0.9,     0.9,     0.6],     # type 1 (Li in reference)\n"
-        "     [0.23,    0.92,    0.09],    # type 2 (Mg in reference)\n"
+        "     [0.77,    0.585,   0.585],   # type 2 (Mg in reference, 48h)\n"
         "     [0.25,    0.25,    0.25],    # type 3 (Na in reference)\n"
         "     [0.15,    0.15,    0.15],    # type 4 (Be in reference)\n"
         "     [0.0,     0.183,   0.183],   # type 5 (K in reference)\n"
@@ -710,7 +710,7 @@ def generate_residence_times_and_transitions():
         "coords = np.array(\n"
         "    [[0.5,     0.5,     0.5],     # P (type 0) - PS4 tetrahedra\n"
         "     [0.9,     0.9,     0.6],     # type 1 (Li in reference)\n"
-        "     [0.23,    0.92,    0.09],    # type 2 (Mg in reference)\n"
+        "     [0.77,    0.585,   0.585],   # type 2 (Mg in reference, 48h)\n"
         "     [0.25,    0.25,    0.25],    # type 3 (Na in reference)\n"
         "     [0.15,    0.15,    0.15],    # type 4 (Be in reference)\n"
         "     [0.0,     0.183,   0.183],   # type 5 (K in reference)\n"
