### 2026-03-22 9b6233c
Add builder validation for close atoms and duplicate sites
Pre-build: check reference structure for same-species atom pairs
closer than min_atom_distance (default 0.5 A). Raises ValueError
with instructions to disable via .with_min_atom_distance(0).

Post-build: check for duplicate PolyhedralSite (by vertex_indices)
or DynamicVoronoiSite (by reference_indices). Raises ValueError.
Does not apply to VoronoiSite or SphericalSite.

Part of #68
