### 2026-03-22 2df2d0c
Address /simplify and PR review feedback
Use indices_for_species from tools.py instead of inline list
comprehension. Remove assert in favour of caller guard. Reject
negative min_atom_distance values. Add explicit VoronoiSite and
SphericalSite handling in _validate_unique_sites. Update build()
docstring with new ValueError cases. Remove unit assumption from
distance threshold (uses lattice units, not necessarily angstroms).
