### 2026-03-22 be25e12
Address Copilot review and document builder validation
- Fix with_min_atom_distance docstring to accurately describe when the
  check runs (whenever a reference structure is set, not only for
  polyhedral/dynamic Voronoi workflows)
- Rename `label` to `index_attr` in _validate_unique_sites to avoid
  confusion with Site.label
- Add with_min_atom_distance to the builder guide method reference
- Add Reference Structure Validation section to the builder guide
  covering close-atom and duplicate-site checks
