### 2026-03-22 c2de344
Fix duplicate sites from incorrect Wyckoff position and add builder validation (#69)
Fix the type 2 (Mg) site coordinate in the argyrodite reference
structure from [0.23, 0.92, 0.09] (96i general position) to
[0.77, 0.585, 0.585] (48h special position). The general position
generated pairs of atoms 0.14 A apart inside the same S tetrahedron,
producing 384 duplicate type 2 sites.

Add two validation checks to TrajectoryBuilder.build():

Pre-build: reject reference structures with same-species atom pairs
closer than min_atom_distance (default 0.5, configurable via
.with_min_atom_distance()). Err
[…]
