+++ tests/benchmark_all_site_types.py
@@ -1,0 +1,150 @@
+"""Benchmark per-frame site assignment for all four site types.
+
+Isolates the assign_site_occupations hot path by pre-building the
+trajectory and then timing only the per-frame analyse_structure calls
+(which include coord assignment + site assignment).
+
+Uses the Li6PS5Cl argyrodite MD trajectory.
+
+Usage:
+    python tests/benchmark_all_site_types.py [--repeats N]
+"""
+
+import argparse
+import sys
+import time
+from pathlib import Path
+
+import numpy as np
+from pymatgen.core import Lattice, Structure
+from pymatgen.io.vasp import Xdatcar
+
+from site_analysis import TrajectoryBuilder
+from site_analysis.site import Site
+
+DATA_DIR = Path(__file__).resolve().parent.parent / "docs" / "source"
+ARGYRODITE_XDATCAR = DATA_DIR / "tutorials" / "data" / "Li6PS5Cl_0p_XDATCAR.gz"
+
+
+def build_reference_structure():
+    lattice = Lattice.cubic(a=10.155)
+    coords = np.array([
+        [0.5, 0.5, 0.5],
+        [0.9, 0.9, 0.6],
+        [0.77, 0.585, 0.585],
+        [0.25, 0.25, 0.25],
+        [0.15, 0.15, 0.15],
+        [0.0, 0.183, 0.183],
+        [0.0, 0.0, 0.0],
+        [0.75, 0.25, 0.25],
+        [0.11824, 0.11824, 0.38176],
+    ])
+    return Structure.from_spacegroup(
+        sg="F-43m", lattice=lattice,
+        species=["P", "Li", "Mg", "Na", "Be", "K", "S", "S", "S"],
+        coords=coords,
+    ) * [2, 2, 2]
+
+
+def build_trajectory(site_type, structures, reference_structure):
+    Site._newid = 0
+    builder = (
+        TrajectoryBuilder()
+        .with_structure(structures[0])
+        .with_reference_structure(reference_structure)
+        .with_mobile_species("Li")
+        .with_structure_alignment(align_species="P")
+        .with_site_mapping(mapping_species=["S", "Cl"])
+    )
+
+    if site_type == "polyhedral":
+        for sp, label in [("Li", "1"), ("Mg", "2"), ("Na", "3"),
+                          ("Be", "4"), ("K", "5")]:
+            builder.with_polyhedral_sites(
+                centre_species=sp, vertex_species="S",
+                cutoff=3.0, n_vertices=4, label=label)
+
+    elif site_type == "voronoi":
+        li_idx = [i for i, s in enumerate(reference_structure)
+                  if s.species_string == "Li"]
+        centres = [reference_structure[i].frac_coords.tolist() for i in li_idx]
+        builder.with_voronoi_sites(centres=centres)
+
+    elif site_type == "dynamic_voronoi":
+        for sp, label in [("Li", "1"), ("Mg", "2"), ("Na", "3"),
+                          ("Be", "4"), ("K", "5")]:
+            builder.with_dynamic_voronoi_sites(
+                centre_species=sp, reference_species="S",
+                cutoff=3.0, n_reference=4, label=label)
+
+    elif site_type == "spherical":
+        centres = []
+        for sp in ["Li", "Mg", "Na", "Be", "K"]:
+            idx = [i for i, s in enumerate(reference_structure)
+                   if s.species_string == sp]
+            centres.extend([reference_structure[i].frac_coords.tolist()
+                           for i in idx])
+        builder.with_spherical_sites(centres=centres, radii=1.5)
+
+    return builder.build()
+
+
+def benchmark_site_type(site_type, structures, reference_structure, repeats):
[…]
