+++ tests/validate_decoupling.py
@@ -1,0 +1,309 @@
+"""Validation and benchmarking for pymatgen decoupling (#59).
+
+Runs the full Li6PS5Cl argyrodite analysis with all four site types
+(polyhedral, Voronoi, dynamic Voronoi, spherical) using the high-level
+TrajectoryBuilder API. Captures per-frame site assignments and timing.
+
+This script uses only the public TrajectoryBuilder API, which is
+unchanged between the pre-decoupling baseline and the current code.
+Run on both commits and compare the JSON output files for correctness.
+
+Usage:
+    # On current code:
+    python tests/validate_decoupling.py --output results_after.json
+
+    # On baseline (git checkout 7379d58):
+    python tests/validate_decoupling.py --output results_before.json
+
+    # Compare:
+    python tests/validate_decoupling.py --compare results_before.json results_after.json
+"""
+
+import argparse
+import json
+import sys
+import time
+from pathlib import Path
+
+import numpy as np
+from pymatgen.core import Lattice, Structure
+from pymatgen.io.vasp import Xdatcar
+
+from site_analysis import TrajectoryBuilder
+from site_analysis.site import Site
+
+DATA_DIR = Path(__file__).resolve().parent.parent / "docs" / "source"
+ARGYRODITE_XDATCAR = DATA_DIR / "tutorials" / "data" / "Li6PS5Cl_0p_XDATCAR.gz"
+
+# Reference structure for Li6PS5Cl argyrodite (2x2x2 supercell)
+LATTICE = Lattice.cubic(a=10.155)
+REF_COORDS = np.array([
+    [0.5, 0.5, 0.5],
+    [0.9, 0.9, 0.6],
+    [0.77, 0.585, 0.585],
+    [0.25, 0.25, 0.25],
+    [0.15, 0.15, 0.15],
+    [0.0, 0.183, 0.183],
+    [0.0, 0.0, 0.0],
+    [0.75, 0.25, 0.25],
+    [0.11824, 0.11824, 0.38176],
+])
+REF_SPECIES = ["P", "Li", "Mg", "Na", "Be", "K", "S", "S", "S"]
+
+
+def build_reference_structure():
+    return Structure.from_spacegroup(
+        sg="F-43m", lattice=LATTICE, species=REF_SPECIES, coords=REF_COORDS,
+    ) * [2, 2, 2]
+
+
+def build_trajectory(site_type, structures, reference_structure):
+    """Build a trajectory for one site type."""
+    Site._newid = 0
+
+    builder = (
+        TrajectoryBuilder()
+        .with_structure(structures[0])
+        .with_reference_structure(reference_structure)
+        .with_mobile_species("Li")
+        .with_structure_alignment(align_species="P")
+        .with_site_mapping(mapping_species=["S", "Cl"])
+    )
+
+    if site_type == "polyhedral":
+        for centre_species, label in [
+            ("Li", "type 1"), ("Mg", "type 2"), ("Na", "type 3"),
+            ("Be", "type 4"), ("K", "type 5"),
+        ]:
+            builder.with_polyhedral_sites(
+                centre_species=centre_species,
+                vertex_species="S",
+                cutoff=3.0,
+                n_vertices=4,
+                label=label,
+            )
+
+    elif site_type == "voronoi":
+        # Use all Li-related site centres (matching other site types)
+        centres = []
+        for sp in ["Li", "Mg", "Na", "Be", "K"]:
+            indices = [i for i, s in enumerate(reference_structure)
+                       if s.species_string == sp]
+            centres.extend([reference_structure[i].frac_coords.tolist()
+                           for i in indices])
+        builder.with_voronoi_sites(centres=centres)
+
[…]
