+++ tests/benchmark_containment.py
@@ -35,7 +35,7 @@ def build_argyrodite_trajectory(structures):
     coords = np.array([
         [0.5, 0.5, 0.5],
         [0.9, 0.9, 0.6],
-        [0.23, 0.92, 0.09],
+        [0.77, 0.585, 0.585],
         [0.25, 0.25, 0.25],
         [0.15, 0.15, 0.15],
         [0.0, 0.183, 0.183],
