### 2026-03-22 e34b784
Update changelog for v1.7.0 with pymatgen decoupling and builder validation
