### 2026-03-22 17bb14e
Update README to reflect pymatgen as input format, not internal dependency
