### 2026-03-22 6a28225
Add Goldmann et al. (2025) Li3ScCl6 citation to JOSS paper
