+++ paper/paper.bib
@@ -349,5 +349,19 @@ doi = {10.5281/zenodo.8401669}
   author = {Deng,  Zhi and Zhu,  Zhuoying and Chu,  Iek-Heng and Ong,  Shyue Ping},
   year = {2016},
   month = sep,
-  pages = {281–288}
+  pages = {281--288}
+}
+
+@article{GoldmannEtAl_ChemMater2025,
+  title = {Rotational Stacking Faults in the Ionic Conductor {L}i$_3${S}c{C}l$_6$},
+  volume = {37},
+  ISSN = {1520-5002},
+  DOI = {10.1021/acs.chemmater.5c02303},
+  number = {24},
+  journal = {Chem. Mater.},
+  publisher = {American Chemical Society (ACS)},
+  author = {Goldmann,  Benedek A. and Rosenbach,  Carolin and Evans,  Hayden A. and Helm,  Bianca and Wankmiller,  Bj\"{o}rn and Maus,  Oliver and Suard,  Emmanuelle and Nazar,  Linda F. and Hansen,  Michael Ryan and Morgan,  Benjamin J. and Islam,  M. Saiful and Zeier,  Wolfgang G.},
+  year = {2025},
+  month = dec,
+  pages = {9858--9868}
 }
