+++ paper/paper.md
@@ -51,7 +51,7 @@ The materials modelling community has developed various tools for site-based tra
 
 `site-analysis` enables researchers to apply site-projection analysis to characterise and quantify mechanisms of ion transport in solid electrolytes. The software provides both atom-centric views (tracking which sites each ion visits) and site-centric views (recording which ions occupy each site over time), supporting different analytical perspectives on the same transport processes. This dual representation, combined with multiple site-definition methods including dynamically-updating geometric sites, provides a flexible toolkit for analysing ionic transport mechanisms across diverse materials.
 
-The software has been used to analyse ion-transport mechanisms in lithium-ion and fluoride-ion solid electrolytes [@Morgan2021; @Mercadier2023; @Krenzer2023; @Hu2025]. The package reproduces functionality from an earlier Fortran code that used dynamically defined coordination polyhedra to study ion-transport and defect distributions [@Burbano2016; @Morgan2014; @Morgan2014b; @Morgan2012; @Morgan2011], which was itself motivated by earlier studies using projections onto tetrahedral sites [@Castiglione1999; @Castiglione2001; @Marrocchelli2009]. Documentation and tutorials are available at [https://site-analysis.readthedocs.io](https://site-analysis.readthedocs.io).
+The software has been used to analyse ion-transport mechanisms in lithium-ion and fluoride-ion solid electrolytes [@Morgan2021; @Mercadier2023; @Krenzer2023; @Hu2025; @GoldmannEtAl_ChemMater2025]. The package reproduces functionality from an earlier Fortran code that used dynamically defined coordination polyhedra to study ion-transport and defect distributions [@Burbano2016; @Morgan2014; @Morgan2014b; @Morgan2012; @Morgan2011], which was itself motivated by earlier studies using projections onto tetrahedral sites [@Castiglione1999; @Castiglione2001; @Marrocchelli2009]. Documentation and tutorials are available at [https://site-analysis.readthedocs.io](https://site-analysis.readthedocs.io).
 
 # Acknowledgements
 
