### 2026-03-25 #9606 @bjmorgan
> [@bjmorgan](https://github.com/bjmorgan) are you working on this points and/or can you formulate a response? Thanks

I am working on this (tracking at https://github.com/bjmorgan/site-analysis/issues/43)
