### 2026-03-27 d9ac067
Fix all Sphinx documentation build warnings
- conf.py: set language = 'en' (required by Sphinx 7+), remove invalid
  myst_html_meta dict, add myst_heading_anchors = 4 for cross-doc links
- pbc_handling.md: fix H1-to-H3 header jump, fix broken {doc} references
- atom.py, site.py: remove duplicate attribute docs (frac_coords,
  average_occupation) already documented as @property
- builders.py: convert markdown code blocks to RST :: blocks, fix section
  underlines and bullet list formatting for Napoleon compatibility
- dynamic_voronoi_site_collection.py, polyhedral_site_collection.py,
  spherica
[…]
