+++ site_analysis/builders.py
@@ -3,10 +3,8 @@
 This module provides a fluent builder interface for creating site analysis
 objects, making it easier to set up and run site analysis workflows.
 
-Examples:
-    Creating a simple trajectory with spherical sites:
-    
-    ```python
+Examples::
+
     # Using the builder directly
     trajectory = (TrajectoryBuilder()
                  .with_structure(structure)
@@ -17,7 +15,7 @@ Examples:
                      labels=["octahedral", "tetrahedral"]
                  )
                  .build())
-    
+
     # Using a factory function
     trajectory = create_trajectory_with_spherical_sites(
         structure=structure,
@@ -26,22 +24,18 @@ Examples:
         radii=2.0,
         labels=["octahedral", "tetrahedral"]
     )
-    ```
-    
-    Creating a trajectory with polyhedral sites:
-    
-    ```python
+
+    # Creating a trajectory with polyhedral sites
     trajectory = create_trajectory_with_polyhedral_sites(
         structure=target_structure,
         reference_structure=reference_structure,
         mobile_species="Li",
         centre_species="O",
-        vertex_species=["Li", "Na"], 
+        vertex_species=["Li", "Na"],
         cutoff=3.0,
         n_vertices=4,
         label="tetrahedral"
     )
-    ```
 """
 
 from __future__ import annotations
@@ -61,49 +55,45 @@ from typing import cast, Callable, Sequence, Any
 
 class TrajectoryBuilder:
     """Builder for creating Trajectory objects for site analysis.
-    
+
     This class provides a step-by-step approach to creating a Trajectory
     object for analysing site occupations in crystal structures.
-    
-    Structure Alignment and Site Mapping:
-    ------------------------------------
+
+    **Structure Alignment and Site Mapping**
+
     The builder supports separate control over structure alignment (finding optimal
     translations) and site mapping (identifying corresponding sites):
-    
-    - Structure alignment: Use with_structure_alignment() to control whether and how
-    structures are aligned before site creation.
-    
-    - Site mapping: Use with_site_mapping() to specify which species are used to
-    identify corresponding sites between structures.
-    
-    Default Behaviors:
-    - Alignment is enabled by default
-    - If mapping species are specified but alignment species are not, mapping species
-    will be used for alignment
-    - If alignment species are specified but mapping species are not, alignment species
-    will be used for mapping
-    
-    Example:
-        ```python
-        # Create a builder
+
+    - Structure alignment: Use ``with_structure_alignment()`` to control whether
+      and how structures are aligned before site creation.
+    - Site mapping: Use ``with_site_mapping()`` to specify which species are used
+      to identify corresponding sites between structures.
+
+    **Default Behaviors**
+
+    - Alignment is enabled by default.
+    - If mapping species are specified but alignment species are not, mapping
+      species will be used for alignment.
+    - If alignment species are specified but mapping species are not, alignment
+      species will be used for mapping.
+
+    Example::
+
         builder = TrajectoryBuilder()
-        
-        # Configure it
+
         builder.with_structure(structure)
             .with_reference_structure(reference_structure)
             .with_mobile_species("Li")
-            .with_structure_alignment(align=True, align_species=["O"])  # Align on framework
-            .with_site_mapping(mapping_species=["Na"])  # Map using Na atoms
+            .with_structure_alignment(align=True, align_species=["O"])
+            .with_site_mapping(mapping_species=["Na"])
             .with_polyhedral_sites(
                 centre_species="Li",
                 vertex_species="O",
                 cutoff=2.0,
                 n_vertices=4
             )
-        
-        # Build the trajectory
+
         trajectory = builder.build()
-        ```
     """
     
     def __init__(self) -> None:
[…]
