### 2026-03-27 ef4f3bf
Restore executable tutorials lost during rebase
The tutorials directory (notebooks, plotting helper, README, and
trajectory data) was silently dropped from the tree during the
pymatgen decoupling work. Restore from the last good commit (24f7611).
