### 2026-03-27 c13edea
Replace duplicate docs tutorials with overview page linking to GitHub notebooks
Remove tutorial notebooks and data from docs/source/tutorials/ (these
are already in tutorials/) and replace with a single overview page that
describes each tutorial and links to the GitHub repository. Update
README to direct users to clone the repository for tutorials.
