+++ docs/source/tutorials/residence_times_and_transitions.md
@@ -1,258 +1,0 @@
-# Residence Times and Transition Probabilities in Argyrodite
-
-This tutorial demonstrates how to compute residence times and transition probabilities from a site-analysis trajectory. We reuse the argyrodite Li6PS5Cl dataset from {doc}`the previous tutorial </tutorials/argyrodite_site_definitions>`, so the setup code here is largely the same — see that tutorial for a detailed explanation of the reference structure and trajectory builder configuration.
-
-## Prerequisites
-
-This tutorial requires the following packages:
-
-- `site-analysis`
-- `numpy`
-- `matplotlib`
-
-This tutorial uses the same trajectory data as the argyrodite tutorial.
-The data files are included in the `site-analysis` GitHub repository at
-`tutorials/data/`:
-
-- `Li6PS5Cl_0p_XDATCAR.gz` — 0% anion site exchange (fully ordered)
-- `Li6PS5Cl_50p_XDATCAR.gz` — 50% anion site exchange
-- `Li6PS5Cl_100p_XDATCAR.gz` — 100% anion site exchange
-
-These files are available in the [tutorials/data](https://github.com/bjmorgan/site-analysis/tree/main/tutorials/data) directory of the GitHub repository. The code examples below assume you are running from the repository root directory.
-
-## Setting up the trajectory
-
-We set up the reference structure and trajectory builder exactly as in the argyrodite tutorial. See {doc}`/tutorials/argyrodite_site_definitions` for a full explanation of these steps.
-
-```python
-from pymatgen.io.vasp import Xdatcar
-from pymatgen.core import Structure, Lattice
-import numpy as np
-from site_analysis import TrajectoryBuilder
-
-lattice = Lattice.cubic(a=10.155)
-
-coords = np.array(
-    [[0.5,     0.5,     0.5],     # P (type 0) - PS4 tetrahedra
-     [0.9,     0.9,     0.6],     # type 1 (Li in reference)
-     [0.77,    0.585,   0.585],   # type 2 (Mg in reference, 48h)
-     [0.25,    0.25,    0.25],    # type 3 (Na in reference)
-     [0.15,    0.15,    0.15],    # type 4 (Be in reference)
-     [0.0,     0.183,   0.183],   # type 5 (K in reference)
-     [0.0,     0.0,     0.0],     # S - anion position (4a site)
-     [0.75,    0.25,    0.25],    # S - anion position (4c site)
-     [0.11824, 0.11824, 0.38176]] # S - anion position (16e site)
-)
-
-reference_structure = Structure.from_spacegroup(
-    sg="F-43m",
-    lattice=lattice,
-    species=['P', 'Li', 'Mg', 'Na', 'Be', 'K', 'S', 'S', 'S'],
-    coords=coords) * [2, 2, 2]
-```
-
-```python
-def build_trajectory(structure):
-    builder = TrajectoryBuilder()
-    builder.with_reference_structure(reference_structure)
-    builder.with_structure(structure)
-    builder.with_mobile_species('Li')
-
-    builder.with_polyhedral_sites(
-        centre_species='Li', vertex_species='S',
-        cutoff=3.0, n_vertices=4, label='type 1')
-    builder.with_polyhedral_sites(
-        centre_species='Mg', vertex_species='S',
-        cutoff=3.0, n_vertices=4, label='type 2')
-    builder.with_polyhedral_sites(
-        centre_species='Na', vertex_species='S',
-        cutoff=3.0, n_vertices=4, label='type 3')
-    builder.with_polyhedral_sites(
-        centre_species='Be', vertex_species='S',
-        cutoff=3.0, n_vertices=4, label='type 4')
-    builder.with_polyhedral_sites(
[…]
