### 2026-03-27 9f77bd8
Fix tutorials README to launch Jupyter from tutorials directory
