### 2026-03-27 e635dc9
Add __str__ and _repr_html_ to TransitionTable
TransitionTable previously had no useful string representation,
requiring custom helpers or pandas to display transition data.
Add formatted text output via __str__ and HTML table output via
_repr_html_ for Jupyter notebooks. Format auto-detects from
the matrix dtype: integer for counts, 3 decimal places for
probabilities.
