### 2026-03-27 858b86a
Polish tutorial notebooks and docs overview
- Add circumradii example to comparing_site_definitions tutorial,
  demonstrating sphere overlap and the persistence rule
- Define inradius and circumradius terms on first use
- Clean up argyrodite tutorial: use dict+loop for site types,
  fix prerequisites text, add sub/superscripts for chemical formulae,
  remove speculative mechanism discussion
- Simplify residence_times_and_transitions tutorial code
- Fix tutorials.md to use HTML sub/superscripts for Li6PS5Cl
