### 2026-03-27 f12cd65
Fix mypy no-any-return for Structure.copy() on Linux
Annotate the return type of structure.copy() explicitly, as
pymatgen-core 2026.3.9 does not expose a typed return on all
platforms.
