### 2026-03-27 0ae423a
Replace pymatgen with pymatgen-core in dependencies
The library only uses pymatgen.core directly. The full pymatgen
package (needed for io.vasp and symmetry in tests/benchmarks) moves
to a dev dependency.
