### 2026-03-27 bc45ac4
Bump minimum Python version from 3.10 to 3.11
Required by pymatgen-core, which only supports 3.11+.
