### 2026-03-27 400cf09
Update changelog for v1.8.0 and bump version
