### 2026-03-27 7bb0a1e
Escape HTML in TransitionTable and add pymatgen to tutorial prerequisites
- Escape user-provided keys in _repr_html_ to prevent HTML injection
- List pymatgen in prerequisites for tutorials that import pymatgen.io.vasp
