### 2026-03-27 fb69f3c
Validate TransitionTable matrix has numeric dtype
Reject non-numeric dtypes at construction time rather than relying
on formatting to catch them.
