### 2026-03-27 #9606 @bjmorgan
@LIVazquezS I've now addressed all three points, released as v1.8.0.

The tutorials have been rewritten as self-contained Jupyter notebooks that can be run directly from the tutorials/ directory in the repository. Each notebook lists its prerequisites and data file locations upfront. There are now three tutorials, with the third tutorial covering residence time calculations and analysis of site–site transition probabilities.

The trajectory data is now bundled in tutorials/data/, and the notebooks include clear instructions for locating and using these files.

Could you take a look at the updated tutorials and see if these address your remaining points?
