### 2026-03-31 #9606 @LIVazquezS
@bjmorgan Thank you very much for address my points. Once more, I apologize for the delays with the review.

I think the provided tutorials give a great overview of the capabilities of `site-analysis`.

I recommend acceptance of the paper.
