### 2026-04-02 #9606 @Kevin-Mattheus-Moerman
@jkrajniak thanks for your review efforts here! I see you have all but one box ticked (performance). This is not a critical box, and there are two more postive/completed reviews, so we could proceed in principle. However, I wanted to just check if you feel there is an issue there that the authors should work on, or if you just feel you could not verify performance fully and left it blank (but that you are otherwise happy to recommend acceptance)?
