### 2026-05-18 fcd8238
Update Landgraf et al. citation to JACS version of record
