+++ paper/paper.md
@@ -29,7 +29,7 @@ Understanding ionic transport in crystalline solids is important for developing 
 
 # Statement of need
 
-Molecular dynamics (MD) simulations are widely used to study ionic transport in battery materials, solid electrolytes, and ceramic ion conductors [@LandgrafEtAl_ChemRxiv2024; @Mercadier2023; @Morgan2021; @PoletayevEtAl_NatureMater2022]. 
+Molecular dynamics (MD) simulations are widely used to study ionic transport in battery materials, solid electrolytes, and ceramic ion conductors [@LandgrafEtAl_JAmChemSoc2025; @Mercadier2023; @Morgan2021; @PoletayevEtAl_NatureMater2022]. 
 While these simulations produce detailed atomic trajectories, extracting mechanistic understanding from raw trajectory data can be challenging. A productive approach is to seek a coarse-grained description that retains only the essential features of ion transport. In many crystalline solids, ionic transport can be understood as a series of discrete "jumps" between crystallographic sites—persistent local minima on the potential energy surface where mobile ions reside between jumps. This site-hopping perspective motivates a coarse-graining strategy: transform continuous trajectories into sequences of site occupations, capturing only motion between sites while discarding vibrational motion within sites.
 
 Transforming MD trajectories into site-occupation sequences requires both defining what constitutes a "site" and implementing algorithms to track site occupancy throughout the simulation. While the concept is straightforward, the practical implementation presents challenges: sites must be defined and occupancy criteria must handle edge cases. Without accessible software tools, each research group must independently implement these methods, leading to duplicated effort and potential inconsistencies in analysis approaches.
