+++ paper/paper.bib
@@ -155,14 +155,17 @@ doi = {10.5281/zenodo.8401669}
   doi={10.1088/0953-8984/11/46/303}
 }
 
-@article{LandgrafEtAl_ChemRxiv2024,
+@article{LandgrafEtAl_JAmChemSoc2025,
   title = {Disorder-Mediated Ionic Conductivity in Irreducible Solid Electrolytes},
-  DOI = {10.26434/chemrxiv-2023-33r87-v2},
-  journal = {ChemRxiv},
+  DOI = {10.1021/jacs.5c02784},
+  journal = {J. Am. Chem. Soc.},
   publisher = {American Chemical Society (ACS)},
-  author = {Landgraf,  Victor and Tu,  Mengfu and Zhao,  Wenxuan and Lavrinenko,  Anastasiia and Cheng,  Zhu and Canals,  Jef and de Leeuw,  Joris and Ganapathy,  Swapna and Vasileiadis,  Alexandros and Wagemaker,  Marnix and Famprikis,  Theodosios},
-  year = {2024},
-  month = oct
+  author = {Landgraf, Victor and Tu, Mengfu and Zhao, Wenxuan and Lavrinenko, Anastasia K. and Cheng, Zhu and Canals, Jef and de Leeuw, Joris and Ganapathy, Swapna and Vasileiadis, Alexandros and Wagemaker, Marnix and Famprikis, Theodosios},
+  volume = {147},
+  number = {22},
+  pages = {18840--18852},
+  year = {2025},
+  month = may
 }
 
 @article{Hu2025,
