### 2026-05-18 3269674
Add skip-existing to PyPI publish step
Allows the publish workflow to handle release.published events that
re-fire for an existing tag (e.g. recreating a GitHub release to
re-trigger the Zenodo archive webhook) without failing on the
immutable PyPI version.
