+++ .github/workflows/python-publish.yml
@@ -50,3 +50,5 @@ jobs:
     
     - name: Publish to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        skip-existing: true
