### 2026-05-18 833c4d6
Add Zenodo DOI badge to README
