+++ README.md
@@ -5,6 +5,7 @@
 ![Build Status](https://github.com/bjmorgan/site-analysis/actions/workflows/build.yml/badge.svg)
 [![Documentation Status](https://readthedocs.org/projects/site-analysis/badge/?version=latest)](https://site-analysis.readthedocs.io/en/latest/?badge=latest)
 [![PyPI version](https://badge.fury.io/py/site-analysis.svg)](https://badge.fury.io/py/site-analysis)
+[![DOI](https://zenodo.org/badge/196561551.svg)](https://doi.org/10.5281/zenodo.20275363)
 [![status](https://joss.theoj.org/papers/0a447aeb167964e77c8d381f7d1db89a/status.svg)](https://joss.theoj.org/papers/0a447aeb167964e77c8d381f7d1db89a)
 
 `site-analysis` is a Python module for analysing molecular dynamics simulations of solid-state ion transport, by assigning positions of mobile ions to specific &ldquo;sites&rdquo; within the host structure.
