### 2026-05-18 8315789
Bump GitHub Actions to Node.js 24 versions
Updates actions/checkout (v4 -> v6.0.2), actions/setup-python
(v5 -> v6.2.0) and actions/upload-artifact (v4 -> v7.0.1) across
build, draft-pdf and python-publish workflows. Node.js 20 was
deprecated by the runner, prompting warnings on the previous
versions. SHA pins in draft-pdf.yml updated alongside the version
comments.
