### 2026-06-02 da52135
Remove published JOSS paper and update citation metadata
The JOSS paper has been published (doi:10.21105/joss.09606), so remove the
paper sources and the draft-PDF workflow from the repository.

Update CITATION.cff, the README, and the docs to cite the published paper,
retaining the Zenodo archive as a secondary, version-specific citation.
Also correct the README Python requirement to 3.11 to match pyproject.toml
and CI.
