+++ paper/paper.bib
@@ -1,370 +1,0 @@
-@article{Ong2013,
-  title={Python Materials Genomics (pymatgen): A robust, open-source python library for materials analysis},
-  author={Ong, Shyue Ping and Richards, William Davidson and Jain, Anubhav and Hautier, Geoffroy and Kocher, Michael and Cholia, Shreyas and Gunter, Dan and Chevrier, Vincent L. and Persson, Kristin A. and Ceder, Gerbrand},
-  journal={Computational Materials Science},
-  volume={68},
-  pages={314--319},
-  year={2013},
-  doi={10.1016/j.commatsci.2012.10.028}
-}
-
-@software{GEMDAT,
-author = {Azizi, Victor and Smeets, Stef and Lavrinenko, Anastasia K. and Ciarella, Simone and Famprikis, Theodosios and Landgraf, Victor and Vasileiadis, Alexandros},
-license = {Apache-2.0},
-title = {{GEMDAT}},
-url = {https://github.com/GEMDAT-repos/GEMDAT},
-version = {1.6.2},
-year = {2025},
-doi = {10.5281/zenodo.8401669}
-}
-
-@misc{pymatgen-analysis-diffusion,
-	title = {pymatgen-analysis-diffusion},
-	url = {https://github.com/materialsvirtuallab/pymatgen-analysis-diffusion},
-	year = {2021}
-}
-
-
-@misc{sitator,
-	title = {sitator},
-	url = {https://github.com/Linux-cpp-lisp/sitator},
-	year = {2020}
-}
-
-@article{Krenzer2023,
-  title={Nature of the Superionic Phase Transition of Lithium Nitride from Machine Learning Force Fields},
-  author={Krenzer, Georg and Klarbring, Johan and Tolborg, Kasper and Rossignol, Hugo and McCluskey, Andrew R. and Morgan, Benjamin J. and Walsh, Aron},
-  journal={Chemistry of Materials},
-  volume={35},
-  pages={6133--6140},
-  year={2023},
-  doi={10.1021/acs.chemmater.3c01271}
-}
-
-@article{PoletayevEtAl_NatureMater2022,
-  doi = {10.1038/s41563-022-01316-z},
-  url = {https://doi.org/10.1038/s41563-022-01316-z},
-  year = {2022},
-  month = jul,
-  publisher = {Springer Science and Business Media {LLC}},
-  volume = {21},
-  number = {9},
-  pages = {1066--1073},
-  author = {Andrey D. Poletayev and James A. Dawson and M. Saiful Islam and Aaron M. Lindenberg},
-  title = {Defect-driven anomalous transport in fast-ion conducting solid electrolytes},
-  journal = {Nature Materials}
-}
-
-@article{Mercadier2023,
-  title={Dynamic Lone Pairs and Fluoride-Ion Disorder in Cubic-{B}a{S}n{F}$_4$},
-  author={Mercadier, Briséis and Coles, Samuel W. and Duttine, Mathieu and Legein, Christophe and Body, Monique and Borkiewicz, Olaf J. and Lebedev, Oleg and Morgan, Benjamin J. and Masquelier, Christian and Dambournet, Damien},
-  journal={Journal of the American Chemical Society},
-  volume={145},
-  pages={17700--17714},
-  year={2023},
-  doi={10.1021/jacs.3c08232}
-}
-
-@article{Morgan2021,
-  title={Mechanistic Origin of Superionic Lithium Diffusion in Anion-Disordered {L}i$_6${P}{S}$_5${X} Argyrodites},
-  author={Morgan, Benjamin J.},
[…]
