+++ paper/paper.md
@@ -1,60 +1,0 @@
----
-title: 'site-analysis: A Python package for site-projection analysis of molecular dynamics trajectories' 
-
-tags:
-  - Python
-  - molecular dynamics
-  - ionic conductors
-  - diffusion
-  - crystallography
-  - solid-state ionics
-
-authors:
-  - name: Benjamin J. Morgan 
-    affiliation: "1, 2" 
-    orcid: 0000-0002-3056-8233 
-
-affiliations:
-  - index: 1
-    name: Department of Chemistry, University of Bath, Claverton Down, Bath, BA2 7AY, United Kingdom
-  - index: 2
-    name: The Faraday Institution, Quad One, Harwell Science and Innovation Campus, Didcot, OX11 0RA, United Kingdom
-date: 23 May 2025 
-bibliography: paper.bib
----
-
-# Summary
-
-Understanding ionic transport in crystalline solids is important for developing materials for batteries, fuel cells, and other electrochemical devices. Molecular dynamics (MD) simulations provide detailed information about the microscopic dynamics of mobile ions, but extracting mechanistic insights from these complex data can be challenging. In crystalline solids, ionic diffusion can often be considered to proceed by discrete jumps between "sites"—persistent local minima on the potential energy surface. `site-analysis` is a Python package that processes MD trajectories to produce time-resolved site-occupation data that provide a coarse-grained description of ion transport. The package implements four site-definition methods—spherical, polyhedral, Voronoi, and dynamic Voronoi—allowing users to choose the approach best suited to their requirements. For polyhedral and dynamic Voronoi sites, `site-analysis` implements a reference-based workflow that automates site generation from coordination environment specifications, allowing users to analyse trajectories without manually defining site geometries. `site-analysis` integrates with existing computational materials science workflows through the pymatgen ecosystem and provides both atom-centric and site-centric representations of ion dynamics as simple Python data structures.
-
-# Statement of need
-
-Molecular dynamics (MD) simulations are widely used to study ionic transport in battery materials, solid electrolytes, and ceramic ion conductors [@LandgrafEtAl_JAmChemSoc2025; @Mercadier2023; @Morgan2021; @PoletayevEtAl_NatureMater2022]. 
-While these simulations produce detailed atomic trajectories, extracting mechanistic understanding from raw trajectory data can be challenging. A productive approach is to seek a coarse-grained description that retains only the essential features of ion transport. In many crystalline solids, ionic transport can be understood as a series of discrete "jumps" between crystallographic sites—persistent local minima on the potential energy surface where mobile ions reside between jumps. This site-hopping perspective motivates a coarse-graining strategy: transform continuous trajectories into sequences of site occupations, capturing only motion between sites while discarding vibrational motion within sites.
-
-Transforming MD trajectories into site-occupation sequences requires both defining what constitutes a "site" and implementing algorithms to track site occupancy throughout the simulation. While the concept is straightforward, the practical implementation presents challenges: sites must be defined and occupancy criteria must handle edge cases. Without accessible software tools, each research group must independently implement these methods, leading to duplicated effort and potential inconsistencies in analysis approaches.
-
[…]
